### 2023-04-25 #5409 @editorialbot
[REVIEW]: Swiftest: An N-body Integrator for Gravitational Systems
**Submitting author:** @carlislewishard (<a href="http://orcid.org/0009-0001-0733-3268">Carlisle Wishard</a>)
**Repository:** https://github.com/profminton/swiftest
**Branch with paper.md** (empty if default branch):
**Version:** v2023.10.2
