### 2023-04-25 012e05c
Added timers to help troubleshoot speed issues with the big production runs
