### 2023-04-26 9b2c7d7
Removed timers and rearranged some calls to improve speed
