### 2023-04-26 5136422
Moved balance to just after dump step so we can control how often a balance gets done
