### 2023-04-26 7cc974c
Reverted orbital element utility functions to element-wise scalar operations and got rid of .cross. , .mag. operators. These ran slow when coarrays were used (for some reason I don't entirly understand)
