### 2023-04-28 1d2ce56
Explicitly set mass, radius, etc. to 0 for test particles
