### 2023-04-30 22cf767
Added deallocate statements to get memory issues under control
