### 2023-05-01 9d5306c
Found a typo in simulation_class.py
