+++ setup_sim.py
@@ -1,13 +1,20 @@
 import numpy as np
 import swiftest
+from astropy import constants as const
+from astroquery.jplhorizons import Horizons
+import datetime
 
 seed = None
 rng = np.random.default_rng(seed=seed)
-tstop = 2e6 # years
-dt = 8 / 365.25 # 8 DAYS
+tstop = 10 * 365.25 # 1e3 * 365.25 # years
+dt = 90.0 / 60.0 / 24 / 10 # 1/10 * 90 min to days
 dt_unit = 'Years'
 dt_max = 0
-G = 6.6743e-11 # SI Units
+
+scratch_dr = '/scratch/bell/anand43/swiftest_runs/'
+body_dr = 'tests/'
+simdir = 'test' # '2e3_part_radial'
+simdir = scratch_dr + body_dr + simdir
 
 # unit conversion factors
 S2DAY = 1 / (60.0 * 60 * 24)
@@ -23,18 +30,49 @@ rot = [0, 0, 0.00058] # rad/s
 
 #initial ring particle orbital parameters
 
-n_bodies = 500 # number of bodies
+n_bodies = 1 # number of bodies
 id_start = 100 # starting ID for bodies
 random_2pi = 360 * np.random.rand(n_bodies) # 360 degrees
 random_pi = 180 * np.random.rand(n_bodies) # 180 degrees
-random_a = (rng.random(n_bodies) * (3.0 - 0.5) + 0.5) # randomize the semi-major axis between 0.5 and 3.0 (AU))
-random_e = rng.random(n_bodies) # randomize eccentricity between 0 and 1
-random_i = random_pi / 2 + (np.pi/2 - np.pi/3) # between (-pi/3, pi/3) + pi/2 (equatorial)
-random_capom = random_2pi
-random_omega = random_2pi
-random_capm = random_2pi
-random_radius = (rng.random(n_bodies) * (10.0 - 1.0) + 1.0) * M2AU # randomize the radii of the ring particles between 1 and 10 m
-random_mass = (np.power(random_radius, 3) * 4.0 * np.pi / 3.0 * density) * KG2MSUN # randomize the mass 
+# ISS parameters on 4/28/23
+# random_a = 6.7999524e3                     # (rng.random(n_bodies) * (3.0 - 0.5) + 0.5) # randomize the semi-major axis between 0.5 and 3.0 (AU))
+# random_e = 4.3357182e-4                    # rng.random(n_bodies) # randomize eccentricity between 0 and 1
+# random_i = 6.7813805e1                     # random_pi / 2 + (np.pi/2 - np.pi/3) # between (-pi/3, pi/3) + pi/2 (equatorial)
+# random_capom = 2.2163223e2                 # random_2pi
+# random_omega = 4.4320487e1                 # random_2pi
+# random_capm = 3.2698275e2                  # random_2pi
+# random_radius = 100e-3 # km                # (rng.random(n_bodies) * (10.0 - 1.0) + 1.0) * M2AU # randomize the radii of the ring particles between 1 and 10 m
+# random_mass = 419725 # kg                  # (np.power(random_radius, 3) * 4.0 * np.pi / 3.0 * density) * KG2MSUN # randomize the mass 
+
+# use JPL Horizons for ISS details
+# date = '2023-04-27'
+ephemerides_start_date = '2023-04-27'
+tstart = datetime.date.fromisoformat(ephemerides_start_date)
+tstep = datetime.timedelta(days=1)
+tend = tstart + tstep
+ephemerides_end_date = tend.isoformat()
+ephemerides_step = '1d'
[…]
