### 2023-05-04 120d23d
fixed some typos, edited 'np.str' to 'str' in simulation_class.py
