### 2023-05-10 678e812
Fixed issue that was causing a segfault when an append operation was done using an unallocated source array
