### 2023-05-11 912b052
Fixed bug that caused out of array bounds error
