### 2023-05-11 704ce2d
Added intermediate local allocatable array to the coarray_collect methods
