### 2023-05-15 172f437
Fixed J2 and J4 bug
