### 2023-05-15 1620fdf
Cleaned up bugs that were preventing Swifter input file from being read properly
