+++ python/swiftest/swiftest/io.py
@@ -28,6 +28,7 @@ newfeaturelist = ("RESTART",
                   "YARKOVSKY",
                   "YORP",
                   "IN_FORM",
+                  "NC_IN",
                   "SEED",
                   "INTERACTION_LOOPS",
                   "ENCOUNTER_CHECK",
@@ -483,6 +484,7 @@ def swifter_stream(f, param):
     tlab : string list
         Labels for the tvec data
     """
+    
     while True:  # Loop until you read the end of file
         try:
             # Read single-line header
@@ -492,32 +494,52 @@ def swifter_stream(f, param):
         t = record[0]
         npl = record[1][0] - 1
         ntp = record[2][0]
-        
-        pvec = np.empty((6, npl))
+       
+        if param['OUT_FORM'] == 'XV':
+            rpvec = np.empty((npl,3))
+            vpvec = np.empty((npl,3))
+            rtvec = np.empty((ntp,3))
+            vtvec = np.empty((ntp,3))
+        elif param['OUT_FORM'] == 'EL': 
+            elempvec = np.empty((npl,6))
+            elemtvec = np.empty((ntp,6))
         plid = np.empty(npl, dtype='int')
-        tvec = np.empty((6, ntp))
         tpid = np.empty(ntp, dtype='int')
         if npl > 0:
             GMpl = np.empty(npl)
             Rpl = np.empty(npl)
             for i in range(npl):
                 # Read single-line pl frame for
-                record = f.read_record('<i4', '<f8', '<f8', '(6,)<f8')
+                if param['OUT_FORM'] == 'XV':
+                    record = f.read_record('<i4', '<f8', '<f8', '(3,)<f8', '(3,)<f8')
+                elif param['OUT_FORM'] == 'EL':
+                    record = f.read_record('<i4', '<f8', '<f8', '(6,)<f8')
                 plid[i] = record[0]
                 GMpl[i] = record[1]
                 Rpl[i] = record[2]
-                pvec[:, i] = record[3]
+                if param['OUT_FORM'] == 'XV':
+                    rpvec[i,:] = record[3]
+                    vpvec[i,:] = record[4]
+                elif param['OUT_FORM'] == 'EL':
+                    elempvec[i,:] = record[3]
         if ntp > 0:
             for i in range(ntp):
-                record = f.read_record('<i4', '(6,)<f8')
+                if param['OUT_FORM'] == 'XV':
+                    record = f.read_record('<i4', '(3,)<f8', '(3,)<f8')
+                elif param['OUT_FORM'] == 'EL':
+                    record = f.read_record('<i4', '(6,)<f8')
                 tpid[i] = record[0]
-                tvec[:, i] = record[1]
+                if param['OUT_FORM'] == 'XV':
+                    rtvec[i,:] = record[1]
+                    vtvec[i,:] = record[2]
+                elif param['OUT_FORM'] == 'EL':
+                    elemtvec[i,:] = record[1]
         
-        tlab = []
-        if param['OUT_FORM'] == 'XV' or param['OUT_FORM'] == 'XVEL':
+        tlab = ['id']
+        if param['OUT_FORM'] == 'XV': 
             tlab.append('rh')
             tlab.append('vh')
-        if param['OUT_FORM'] == 'EL' or param['OUT_FORM'] == 'XVEL':
+        elif param['OUT_FORM'] == 'EL':
             tlab.append('a')
             tlab.append('e')
             tlab.append('inc')
@@ -527,245 +549,13 @@ def swifter_stream(f, param):
         plab = tlab.copy()
         plab.append('Gmass')
         plab.append('radius')
[…]
