+++ python/swiftest/swiftest/simulation_class.py
@@ -345,7 +345,6 @@ class Simulation(object):
 
         # Parameters are set in reverse priority order. First the defaults, then values from a pre-existing input file,
         # then using the arguments passed via **kwargs.
-
         #--------------------------
         # Lowest Priority: Defaults:
         #--------------------------
@@ -371,7 +370,7 @@ class Simulation(object):
             self.read_encounters = read_encounters
             
         if read_param or read_data:
-            if self.read_param(read_init_cond = True, dask=dask):
+            if self.read_param(read_init_cond = True, dask=dask, **kwargs):
                 # We will add the parameter file to the kwarg list. This will keep the set_parameter method from
                 # overriding everything with defaults when there are no arguments passed to Simulation()
                 kwargs['param_file'] = self.param_file
@@ -851,8 +850,8 @@ class Simulation(object):
 
         # Setters returning parameter dictionary values
         param_dict = {}
-        param_dict.update(self.set_unit_system(**kwargs))
         param_dict.update(self.set_integrator(**kwargs))
+        param_dict.update(self.set_unit_system(**kwargs))
         param_dict.update(self.set_simulation_time(**kwargs))
         param_dict.update(self.set_init_cond_files(**kwargs))
         param_dict.update(self.set_output_files(**kwargs))
@@ -955,6 +954,11 @@ class Simulation(object):
                 self.binary_path = "NOT IMPLEMENTED FOR THIS CODE"
                 self.driver_executable = None
             update_list.append("driver_executable")
+        
+        if self.codename == "Swifter":
+            J2 = self.param.pop("J2",0.0)
+            J4 = self.param.pop("J4",0.0)
+            self.param = io.swiftest2swifter_param(self.param, J2, J4) 
 
         if integrator is not None:
             valid_integrator = ["symba","rmvs","whm","helio"]
@@ -1172,113 +1176,21 @@ class Simulation(object):
         if close_encounter_check is not None:
             self.param["CHK_CLOSE"] = close_encounter_check
             update_list.append("close_encounter_check")
-
-        if general_relativity is not None:
-            self.param["GR"] = general_relativity
-            update_list.append("general_relativity")
-
-        fragmentation_models = ["FRAGGLE"]
-        if collision_model is not None:
-            collision_model = collision_model.upper()
-            fragmentation = collision_model in fragmentation_models
-            if self.codename != "Swiftest" and self.integrator != "symba" and fragmentation:
-                warnings.warn("Fragmentation is only available on Swiftest SyMBA.",stacklevel=2)
-                self.param['COLLISION_MODEL'] = "MERGE"
-            else:
-                self.param['COLLISION_MODEL'] = collision_model
-                update_list.append("collision_model")
-                if fragmentation:
-                    if "MIN_GMFRAG" not in self.param and minimum_fragment_mass is None and minimum_fragment_gmass is None:
-                        warnings.warn("Minimum fragment mass is not set. Set it using minimum_fragment_gmass or minimum_fragment_mass",stacklevel=2)
-                    else:
-                        update_list.append("minimum_fragment_gmass")
-
-        if minimum_fragment_gmass is not None and minimum_fragment_mass is not None:
-            warnings.warn("Only set either minimum_fragment_mass or minimum_fragment_gmass, but not both!",stacklevel=2)
-
-        if minimum_fragment_gmass is not None:
-            self.param["MIN_GMFRAG"] = minimum_fragment_gmass
-            if "minmum_fragment_gmass" not in update_list:
-                update_list.append("minimum_fragment_gmass")
-        elif minimum_fragment_mass is not None:
-            self.param["MIN_GMFRAG"] = minimum_fragment_mass * self.GU
-            if "minimum_fragment_gmass" not in update_list:
-                update_list.append("minimum_fragment_gmass")
[…]
