### 2023-05-15 12bc34e
Improved the reading of Swifter files using the vec2xr method
