+++ python/swiftest/swiftest/init_cond.py
@@ -282,10 +282,12 @@ def vec2xr(param: Dict, **kwargs: Any):
 
     # Check for valid keyword arguments
     kwargs = {k:kwargs[k] for k,v in kwargs.items() if v is not None}
-    if "rot" not in kwargs and "Gmass" in kwargs:
-        kwargs['rot'] = np.zeros((len(kwargs['Gmass']),3))
-    if "Ip" not in kwargs and "Gmass" in kwargs:
-        kwargs['Ip'] = np.full((len(kwargs['Gmass']),3), 0.4)
+  
+    if "rotation" in param and param['rotation'] == True: 
+        if "rot" not in kwargs and "Gmass" in kwargs:
+            kwargs['rot'] = np.zeros((len(kwargs['Gmass']),3))
+        if "Ip" not in kwargs and "Gmass" in kwargs:
+            kwargs['Ip'] = np.full((len(kwargs['Gmass']),3), 0.4)
 
     if "time" not in kwargs:
         kwargs["time"] = np.array([0.0])
