### 2023-05-15 fed8c21
More change to clean up Swifter file saving
