### 2023-05-16 9adbcfe
Restored some floating point operations to make them closer to the Swifter version
