+++ src/swiftest/swiftest_util.f90
@@ -2609,7 +2609,7 @@ contains
       end select
       end select
 
-      call nbody_system%pl%set_rhill(nbody_system%cb)
+      if (.not.param%lrhill_present) call nbody_system%pl%set_rhill(nbody_system%cb)
 
       ! Take a minimal snapshot wihout all of the extra storage objects
       allocate(snapshot, mold=nbody_system)
