### 2023-05-16 9bb229b
Restructured oblateness acceleration to be in functional form so I can call it from the outer encounters like the original RMVS did
