### 2023-05-16 45c298e
Switched a bunch of subroutines to the strict math compiler flag category to tamp down on floating point nonsense.
