+++ src/CMakeLists.txt
@@ -13,17 +13,36 @@
 
 # Add the source files
 SET(STRICT_MATH_FILES
-            ${SRC}/swiftest/swiftest_kick.f90 
-            ${SRC}/helio/helio_kick.f90
-            ${SRC}/rmvs/rmvs_kick.f90
-            ${SRC}/symba/symba_kick.f90
-            ${SRC}/whm/whm_kick.f90
-            ${SRC}/swiftest/swiftest_user.f90
             ${SRC}/fraggle/fraggle_generate.f90
             ${SRC}/fraggle/fraggle_util.f90
             ${SRC}/fraggle/fraggle_module.f90
+            ${SRC}/helio/helio_drift.f90
+            ${SRC}/helio/helio_gr.f90
+            ${SRC}/helio/helio_kick.f90
+            ${SRC}/helio/helio_step.f90
             ${SRC}/misc/lambda_function_module.f90
             ${SRC}/misc/solver_module.f90
+            ${SRC}/operator/operator_module.f90
+            ${SRC}/operator/operator_cross.f90
+            ${SRC}/operator/operator_mag.f90
+            ${SRC}/operator/operator_unit.f90
+            ${SRC}/rmvs/rmvs_kick.f90
+            ${SRC}/rmvs/rmvs_step.f90
+            ${SRC}/swiftest/swiftest_drift.f90    
+            ${SRC}/swiftest/swiftest_gr.f90       
+            ${SRC}/swiftest/swiftest_kick.f90 
+            ${SRC}/swiftest/swiftest_user.f90
+            ${SRC}/swiftest/swiftest_obl.f90    
+            ${SRC}/swiftest/swiftest_orbel.f90
+            ${SRC}/symba/symba_drift.f90
+            ${SRC}/symba/symba_gr.f90
+            ${SRC}/symba/symba_kick.f90
+            ${SRC}/symba/symba_step.f90
+            ${SRC}/whm/whm_coord.f90
+            ${SRC}/whm/whm_drift.f90
+            ${SRC}/whm/whm_gr.f90
+            ${SRC}/whm/whm_kick.f90
+            ${SRC}/whm/whm_step.f90
 )
 
 SET(FAST_MATH_FILES
@@ -33,7 +52,6 @@ SET(FAST_MATH_FILES
             ${SRC}/misc/io_progress_bar_module.f90
             ${SRC}/encounter/encounter_module.f90
             ${SRC}/collision/collision_module.f90
-            ${SRC}/operator/operator_module.f90
             ${SRC}/walltime/walltime_module.f90
             ${SRC}/swiftest/swiftest_module.f90
             ${SRC}/whm/whm_module.f90
@@ -49,36 +67,18 @@ SET(FAST_MATH_FILES
             ${SRC}/encounter/encounter_check.f90
             ${SRC}/encounter/encounter_io.f90
             ${SRC}/encounter/encounter_util.f90
-            ${SRC}/helio/helio_drift.f90
-            ${SRC}/helio/helio_gr.f90
-            ${SRC}/helio/helio_step.f90
             ${SRC}/helio/helio_util.f90
             ${SRC}/netcdf_io/netcdf_io_implementations.f90  
-            ${SRC}/operator/operator_cross.f90
-            ${SRC}/operator/operator_mag.f90
-            ${SRC}/operator/operator_unit.f90
             ${SRC}/rmvs/rmvs_discard.f90
             ${SRC}/rmvs/rmvs_encounter_check.f90
-            ${SRC}/rmvs/rmvs_step.f90
             ${SRC}/rmvs/rmvs_util.f90
             ${SRC}/swiftest/swiftest_discard.f90  
[…]
