### 2023-05-17 4ed6bf4
Removed compiler flag that was causing test particle orbits to go haywire (-assume contiguous_assumed_shape) and moved more subroutines to the strict math category.
