+++ src/CMakeLists.txt
@@ -13,6 +13,7 @@
 
 # Add the source files
 SET(STRICT_MATH_FILES
+            ${SRC}/collision/collision_generate.f90
             ${SRC}/fraggle/fraggle_generate.f90
             ${SRC}/fraggle/fraggle_util.f90
             ${SRC}/fraggle/fraggle_module.f90
@@ -59,7 +60,6 @@ SET(FAST_MATH_FILES
             ${SRC}/helio/helio_module.f90
             ${SRC}/symba/symba_module.f90
             ${SRC}/collision/collision_check.f90  
-            ${SRC}/collision/collision_generate.f90
             ${SRC}/collision/collision_io.f90     
             ${SRC}/collision/collision_regime.f90   
             ${SRC}/collision/collision_resolve.f90  
