### 2023-05-18 297dedf
Bug fixes uncovered when compiling in gfortran with coarrays turned off
