### 2023-05-18 29afacf
Changed how coarray preprocessor variable is defined to use the older CMAKE TARGET_COMPILE_DEFINITIONS instead of the newer ADD_DEFINITIONS
