+++ src/CMakeLists.txt
@@ -111,6 +111,8 @@ SET_SOURCE_FILES_PROPERTIES(${SWIFTEST_src} PROPERTIES Fortran_PREPROCESS ON)
 # # Uncomment if you need to link to BLAS and LAPACK
 TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} ${NETCDF_LIBRARIES} ${NETCDF_FORTRAN_LIBRARIES})
 
+
+
 IF(USE_OPENMP)
     SET_TARGET_PROPERTIES(${SWIFTEST_DRIVER} PROPERTIES
                           COMPILE_FLAGS "${OpenMP_Fortran_FLAGS}"
@@ -118,6 +120,7 @@ IF(USE_OPENMP)
 ENDIF(USE_OPENMP)
 
 IF(USE_COARRAY)
+    TARGET_COMPILE_DEFINITIONS(${SWIFTEST_DRIVER} PRIVATE -DCOARRAY)
     SET_TARGET_PROPERTIES(${SWIFTEST_DRIVER} PROPERTIES
                           COMPILE_FLAGS "${Coarray_Fortran_FLAGS}"
                           LINK_FLAGS "${Coarray_Fortran_FLAGS}")
