### 2023-05-18 db6ad98
Fixed issue where the undefined nplm was being reported in non-SyMBA runs
