### 2023-05-18 43ae5ed
Switched variable from associated to integer based on an error that came up when compiling with ifx
