### 2023-05-18 9d78839
Added a regex expression to catch compiler options that are not supported by ifx
