### 2023-05-18 862888f
Another variable that needed to be set to integer due to ifx rules on do concurrent
