### 2023-05-18 4d52767
Fixed do concurrents to remove the associated variables from loop index
