+++ src/whm/whm_kick.f90
@@ -82,10 +82,12 @@ contains
       real(DP),                     intent(in)    :: t      !! Current time
       logical,                      intent(in)    :: lbeg   !! Logical flag that determines whether or not this is the beginning or end of the step
       ! Internals
-      integer(I4B)                                :: i
+      integer(I4B)                                :: i, npl, ntp
       real(DP), dimension(NDIM)                   :: ah0
    
-      associate(tp => self, ntp => self%nbody, pl => nbody_system%pl, cb => nbody_system%cb, npl => nbody_system%pl%nbody)
+      associate(tp => self, pl => nbody_system%pl, cb => nbody_system%cb)
+         npl = nbody_system%pl%nbody
+         ntp = self%nbody
          if (ntp == 0 .or. npl == 0) return
          nbody_system%lbeg = lbeg
 
@@ -151,16 +153,15 @@ contains
       class(swiftest_cb), intent(in)    :: cb !! WHM central body object
       class(whm_pl),      intent(inout) :: pl !! WHM massive body object
       ! Internals
-      integer(I4B)                 :: i
+      integer(I4B)                 :: i, npl
       real(DP), dimension(NDIM)    :: ah1h, ah1j
 
-      associate(npl => pl%nbody)
-         do concurrent (i = 2:npl, pl%lmask(i))
-            ah1j(:) = pl%xj(:, i) * pl%ir3j(i)
-            ah1h(:) = pl%rh(:, i) * pl%ir3h(i)
-            pl%ah(:, i) = pl%ah(:, i) + cb%Gmass * (ah1j(:) - ah1h(:))
-         end do
-      end associate
+      npl = pl%nbody
+      do concurrent (i = 2:npl, pl%lmask(i))
+         ah1j(:) = pl%xj(:, i) * pl%ir3j(i)
+         ah1h(:) = pl%rh(:, i) * pl%ir3h(i)
+         pl%ah(:, i) = pl%ah(:, i) + cb%Gmass * (ah1j(:) - ah1h(:))
+      end do
    
       return
    end subroutine whm_kick_getacch_ah1
@@ -178,22 +179,21 @@ contains
       class(swiftest_cb), intent(in)    :: cb !! Swiftest central body object
       class(whm_pl),      intent(inout) :: pl !! WHM massive body object
       ! Internals
-      integer(I4B)                 :: i
+      integer(I4B)                 :: i, npl
       real(DP)                     :: etaj, fac
       real(DP), dimension(NDIM)    :: ah2, ah2o
    
-      associate(npl => pl%nbody)
-         ah2(:) = 0.0_DP
-         ah2o(:) = 0.0_DP
-         etaj = cb%Gmass
-         do concurrent(i = 2:npl, pl%lmask(i))
-            etaj = etaj + pl%Gmass(i - 1)
-            fac = pl%Gmass(i) * cb%Gmass * pl%ir3j(i) / etaj
-            ah2(:) = ah2o + fac * pl%xj(:, i)
-            pl%ah(:,i) = pl%ah(:, i) + ah2(:)
-            ah2o(:) = ah2(:)
-         end do
-      end associate
+      npl = pl%nbody
+      ah2(:) = 0.0_DP
+      ah2o(:) = 0.0_DP
+      etaj = cb%Gmass
+      do concurrent(i = 2:npl, pl%lmask(i))
+         etaj = etaj + pl%Gmass(i - 1)
+         fac = pl%Gmass(i) * cb%Gmass * pl%ir3j(i) / etaj
+         ah2(:) = ah2o + fac * pl%xj(:, i)
+         pl%ah(:,i) = pl%ah(:, i) + ah2(:)
+         ah2o(:) = ah2(:)
[…]

[…]
