### 2023-05-19 22bad3d
Cleaned up some unused variables
