### 2023-05-19 922c3f0
Reviewed compiler flags and improved the consistency, and added ability to turn off OpenMP SIMD directives when building for CPU-agnostic container
