+++ cmake/Modules/SetFortranFlags.cmake
@@ -64,9 +64,51 @@ ENDIF(CMAKE_Fortran_FLAGS_RELEASE AND CMAKE_Fortran_FLAGS_TESTING AND CMAKE_Fort
 ### GENERAL FLAGS ###
 #####################
 
+# Free form
+SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                Fortran "-ffree-form" # GNU
+                ) 
+
 # Don't add underscores in symbols for C-compatability
 SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                 Fortran "-fno-underscoring")
+                Fortran "-fno-underscoring" # GNU
+                ) 
+
+# Determines whether the current Fortran Standard behavior of the compiler is fully implemented. 
+SET_COMPILE_FLAG(CMAKE_Fortran_Flags "${CMAKE_Fortran_FLAGS}"
+                Fortran "-standard-semantics" # Intel
+                        "/standard-semantics" # Intel Windows
+                )
+
+# Tells the compiler to issue compile-time messages for nonstandard language elements (Fortran 2018).                
+SET_COMPILE_FLAG(CMAKE_Fortran_Flags "${CMAKE_Fortran_FLAGS}"
+                Fortran "-stand f18"  # Intel
+                        "/stand:f18"  # Intel Windows
+                        "-fstd=f2018" # GNU
+                )  
+
+# Allows for lines longer than 80 characters without truncation
+SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                Fortran "-ffree-line-length-none" # GNU (gfortran)
+                )
+
+# Disables right margin wrapping in list-directed output
+SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                Fortran  "-no-wrap-margin" # Intel
+                         "/wrap-margin-"   # Intel Windows        
+                )
+
+# Aligns a variable to a specified boundary and offset
+SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                Fortran "-align all -align array64byte" # Intel
+                        "/align:all /align:array64byte" # Intel Windows
+                )
+
+# Enables changing the variable and array memory layout
+SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                Fortran "-pad"  # Intel
+                        "/Qpad" # Intel Windows
+                )
 
 # There is some bug where -march=native doesn't work on Mac
 IF(APPLE)
@@ -74,18 +116,10 @@ IF(APPLE)
 ELSE()
     SET(GNUNATIVE "-march=native")
 ENDIF()
-# Optimize for the host's architecture
-SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                 Fortran "-xhost"        # Intel
-                         "/QxHost"       # Intel Windows
-                         ${GNUNATIVE}    # GNU
-                )
-
 
 ###################
 ### DEBUG FLAGS ###
 ###################
-# NOTE: debugging symbols (-g or /debug:full) are already on by default
 
 # Disable optimizations
 SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS_DEBUG "${CMAKE_Fortran_FLAGS_DEBUG}"
@@ -106,33 +140,34 @@ SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS_DEBUG "${CMAKE_Fortran_FLAGS_DEBUG}"
[…]
