### 2023-05-19 8a8e506
Simplified initial conditions of basic simulation to run from the Python script
