+++ examples/Basic_Simulation/initial_conditions.py
@@ -74,10 +74,10 @@ omega_tp    = default_rng().uniform(0.0, 360.0, ntp)
 capm_tp     = default_rng().uniform(0.0, 360.0, ntp)
 
 sim.add_body(name=name_tp, a=a_tp, e=e_tp, inc=inc_tp, capom=capom_tp, omega=omega_tp, capm=capm_tp)
-
-# Run the simulation. Arguments may be defined here or thorugh the swiftest.Simulation() method.
-#sim.run(tstart=0.0, tstop=1.0e3, dt=0.01, istep_out=100, dump_cadence=10)
 sim.set_parameter(tstart=0.0, tstop=1.0e3, dt=0.01, istep_out=100, dump_cadence=0)
 # Display the run configuration parameters.
 sim.get_parameter()
 sim.save()
+
+# Run the simulation. Arguments may be defined here or thorugh the swiftest.Simulation() method.
+sim.run()
