### 2023-05-19 a85d5ad
Fixed bug that prevented progress bar from updating
