### 2023-05-19 20c7e4e
Fixed bug that prevented the tqdm progress bar to start
