### 2023-05-19 cda911f
Added cmake option to test whether do concurrent accepts local-spec. Newer versions of ifort segfault without it. Old ones can't compile with it.
