### 2023-05-20 eff54b5
Added local-spec to every remaining do concurrent
