+++ src/helio/helio_kick.f90
@@ -117,7 +117,11 @@ contains
          else
             call pl%set_beg_end(rend = pl%rh)
          end if
+#ifdef DOCONLOC
+         do concurrent(i = 1:npl, pl%lmask(i)) shared(pl,dt)
+#else
          do concurrent(i = 1:npl, pl%lmask(i)) 
+#endif
             pl%vb(1, i) = pl%vb(1, i) + pl%ah(1, i) * dt
             pl%vb(2, i) = pl%vb(2, i) + pl%ah(2, i) * dt
             pl%vb(3, i) = pl%vb(3, i) + pl%ah(3, i) * dt
@@ -152,7 +156,11 @@ contains
          ntp = self%nbody
          tp%ah(:, 1:ntp) = 0.0_DP
          call tp%accel(nbody_system, param, t, lbeg)
+#ifdef DOCONLOC
+         do concurrent(i = 1:ntp, tp%lmask(i)) shared(tp,dt)
+#else
          do concurrent(i = 1:ntp, tp%lmask(i)) 
+#endif
             tp%vb(:, i) = tp%vb(:, i) + tp%ah(:, i) * dt
          end do
       end associate
