### 2023-05-20 4fa3506
Cleaned up a few more unused variables caught during testing on fresh install
