### 2023-05-22 9556d18
Draft Dockerfile for creating a container for Swftest. I've commented out everything after netcdf-c compilation to troubleshoot the location of mpiifort for compiling the rest.
