### 2023-05-22 5b5b5c1
Added missing components to Dockerfile
