### 2023-05-23 8abdaec
Able to build the Docker container up to swiftest by using the ifx compiler, though I know this doesn't work for swiftest
