### 2023-05-23 3dfb89a
More improvements to Dockerfile. Now testing.
