+++ Dockerfile
@@ -20,7 +20,6 @@ RUN wget -O- https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRO
 RUN echo "deb [signed-by=/usr/share/keyrings/oneapi-archive-keyring.gpg] https://apt.repos.intel.com/oneapi all main" | tee /etc/apt/sources.list.d/oneAPI.list
 RUN apt-get -y update && apt-get upgrade -y
 RUN apt-get install -y intel-hpckit
-RUN . /opt/intel/oneapi/setvars.sh
 
 # Build the NetCDF libraries
 RUN mkdir -p /opt/build && mkdir -p /opt/dist
@@ -47,11 +46,65 @@ RUN cd netcdf-fortran && mkdir build && cd build && \
    cmake .. -DCMAKE_INSTALL_PREFIX="${INDIR}"  && \
    make && make install
 
-#Swiftest
+# #Swiftest
 RUN git clone -b debug https://github.com/carlislewishard/swiftest.git
-RUN cd swiftest && cmake -P distclean.cmake && mkdir build && cd build && \
-   FC="${INTEL_DIR}/mpi/latest/bin/mpiifort" cmake .. -DCMAKE_BUILD_TYPE=release -DCMAKE_INSTALL_PREFIX="${INDIR}" && \ 
-   make && make install
+ENV FC="${INTEL_DIR}/mpi/latest/bin/mpiifort"
+ENV NETCDF_HOME="${INDIR}"
+ENV NETCDF_FORTRAN_HOME="${INDIR}"
+ENV LANG=C.UTF-8
+ENV ACL_BOARD_VENDOR_PATH='/opt/Intel/OpenCLFPGA/oneAPI/Boards'
+ENV ADVISOR_2023_DIR='/opt/intel/oneapi/advisor/2023.1.0'
+ENV APM='/opt/intel/oneapi/advisor/2023.1.0/perfmodels'
+ENV CCL_CONFIGURATION='cpu_gpu_dpcpp'
+ENV CCL_ROOT='/opt/intel/oneapi/ccl/2021.9.0'
+ENV CLASSPATH='/opt/intel/oneapi/mpi/2021.9.0//lib/mpi.jar:/opt/intel/oneapi/dal/2023.1.0/lib/onedal.jar'
+ENV CLCK_ROOT='/opt/intel/oneapi/clck/2021.7.3'
+ENV CMAKE_PREFIX_PATH='/opt/intel/oneapi/tbb/2021.9.0/env/..:/opt/intel/oneapi/dnnl/2023.1.0/cpu_dpcpp_gpu_dpcpp/../lib/cmake:/opt/intel/oneapi/dal/2023.1.0:/opt/intel/oneapi/compiler/2023.1.0/linux/IntelDPCPP:/opt/intel/oneapi/ccl/2021.9.0/lib/cmake/oneCCL'
+ENV CMPLR_ROOT='/opt/intel/oneapi/compiler/2023.1.0'
[…]
