### 2023-05-23 3d8696b
Added missing dependencies to Dockerfile
