### 2023-05-23 9ada3e7
Adjusted flags to attempt to get cross compiling working with enough vector instructions
