### 2023-05-23 ffb8f30
Updated structure of project to better handle containerization, including a new cmake option to build with a lower level of vector support
