### 2023-05-23 05aa725
Added ability to control whether to use the containerized executable or not via the Python code
