+++ singularity/setenv.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+# source this file to set the path to the swiftest.sif file
+export SWIFTEST_SIF=${PWD}
