### 2023-05-23 f8265e2
Adjusted flags to allow the containerized version to work on more CPUs
