### 2023-05-23 6ce8e38
Created docker pull script
