### 2023-05-24 571b4a2
Fixed issue in the environment variable setter, as it was only pointing to the directory and not the swiftest.sif file itself
