+++ singularity/setenv.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 # source this file to set the path to the swiftest.sif file
-export SWIFTEST_SIF=${PWD}
+export SWIFTEST_SIF="${PWD}/swiftest.sif"
