### 2023-05-24 0605144
Moved swiftest_particle_info further up in the code before it is used in the swiftest_body type definition.
