### 2023-05-24 5245e2b
Use more robust way to check for valid bodies in a NetCDF file with the ieee_arithetic intrinsic module
