### 2023-05-25 660eb48
Improved the way that semi-interacting bodies are identified from input
