+++ src/swiftest/swiftest_io.f90
@@ -123,8 +123,6 @@ contains
       real(DP), dimension(NDIM)       :: L_total_now,  L_orbit_now,  L_spin_now
       real(DP)                        :: ke_orbit_now,  ke_spin_now,  pe_now,  E_orbit_now, be_now, be_cb_now, be_cb_orig, te_now
       real(DP)                        :: GMtot_now
-      character(len=STRMAX)           :: errmsg
-      integer(I4B), parameter         :: EGYIU = 72
       character(len=*), parameter     :: EGYTERMFMT = '(" DL/L0 = ", ES12.5, "; DE_orbit/|E0| = ", ES12.5, "; DE_total/|E0| = ", ES12.5, "; DM/M0 = ", ES12.5)'
 
       associate(nbody_system => self, pl => self%pl, cb => self%cb, npl => self%pl%nbody, display_unit => param%display_unit)
@@ -204,10 +202,6 @@ contains
       end associate
 
       return
-
-      667 continue
-      write(*,*) "Error writing energy and momentum tracking file: " // trim(adjustl(errmsg))
-      call base_util_exit(FAILURE)
    end subroutine swiftest_io_conservation_report
 
 
@@ -226,7 +220,6 @@ contains
       real(DP)                                       :: tfrac             !! Fraction of total simulation time completed
       type(progress_bar), save                       :: pbar              !! Object used to print out a progress bar
       character(len=64)                              :: pbarmessage
-      character(*), parameter                        :: symbacompactfmt = '(";NPLM",ES22.15,$)'
 #ifdef COARRAY
       character(*), parameter :: co_statusfmt = '("Image: ",I4, "; Time = ", ES12.5, "; fraction done = ", F6.3, ' // & 
                                              '"; Number of active pl, tp = ", I6, ", ", I6)'
@@ -465,7 +458,6 @@ contains
       character(len=STRMAX), dimension(:), allocatable :: arg
       integer(I4B), dimension(:), allocatable :: ierr
       integer :: i,narg
-      character(len=*),parameter    :: linefmt = '(A)'
 
       narg = command_argument_count() 
       if (narg > 0) then
@@ -653,6 +645,7 @@ contains
       ! Internals
       integer(I4B)                              :: itmax, idmax, tslot
       real(DP), dimension(:), allocatable       :: vals
+      logical, dimension(:), allocatable        :: plmask, tpmask
       real(DP), dimension(1)                    :: rtemp
       real(DP), dimension(NDIM)                 :: rot0, Ip0, L
       real(DP) :: mass0
@@ -689,7 +682,8 @@ contains
             self%L_total_orig(:) = self%L_orbit_orig(:) + self%L_spin_orig(:) 
 
             call netcdf_io_check( nf90_get_var(nc%id, nc%Gmass_varid, vals, start=[1,tslot], count=[idmax,1]), "netcdf_io_get_t0_values_system Gmass_varid" )
-            self%GMtot_orig = vals(1) + sum(vals(2:idmax), vals(2:idmax) == vals(2:idmax)) 
+            call nc%get_valid_masks(plmask,tpmask)
+            self%GMtot_orig = vals(1) + sum(vals(2:idmax), plmask(:))
 
             cb%GM0 = vals(1)
             cb%dGM = cb%Gmass - cb%GM0
@@ -1030,7 +1024,7 @@ contains
    end subroutine swiftest_io_netcdf_open
 
 
-   module subroutine swiftest_io_netcdf_get_valid_masks(self, plmask, tpmask)
+   module subroutine swiftest_io_netcdf_get_valid_masks(self, plmask, tpmask, plmmask, Gmtiny)
       !! author: David A. Minton
       !!
       !! Given an open NetCDF, returns logical masks indicating which bodies in the body arrays are active pl and tp type at the current time.
[…]
