### 2023-05-25 8c9e052
Removed the line in the executable that calls the shell default shell configuration file temporarily until I determine if it is necessary
