### 2023-05-25 54869be
Improved GNU warning flags in debug mode
