### 2023-05-25 511e425
Fixed inconsistent types in index flattener routines
