+++ src/swiftest/swiftest_util.f90
@@ -1101,12 +1101,12 @@ contains
       class(swiftest_pl),         intent(inout) :: self  !! Swiftest massive body object
       class(swiftest_parameters), intent(inout) :: param !! Current run configuration parameters
       ! Internals
-      integer(I4B) :: i, j, err
-      integer(I8B) :: k, npl
+      integer(I4B) :: err, i, j
+      integer(I8B) :: k, npl8
 
-      npl = int(self%nbody, kind=I8B)
-      associate(nplpl => self%nplpl)
-         nplpl = npl * (npl - 1_I8B) / 2_I8B ! number of entries in a strict lower triangle, npl x npl
+      associate(npl => self%nbody, nplpl => self%nplpl)
+         npl8 = int(npl, kind=I8B)
+         nplpl = npl8 * (npl8 - 1_I8B) / 2_I8B ! number of entries in a strict lower triangle, npl x npl
          if (param%lflatten_interactions) then
             if (allocated(self%k_plpl)) deallocate(self%k_plpl) ! Reset the index array if it's been set previously
             allocate(self%k_plpl(2, nplpl), stat=err)
@@ -1118,7 +1118,7 @@ contains
 #else
                do concurrent (i=1:npl, j=1:npl, j>i)
 #endif
-                  call swiftest_util_flatten_eucl_ij_to_k(self%nbody, i, j, k)
+                  call swiftest_util_flatten_eucl_ij_to_k(npl, i, j, k)
                   self%k_plpl(1, k) = i
                   self%k_plpl(2, k) = j
                end do
@@ -1145,17 +1145,18 @@ contains
       class(swiftest_pl),         intent(in)    :: pl    !! Swiftest massive body object
       class(swiftest_parameters), intent(inout) :: param !! Current run configuration parameters
       ! Internals
-      integer(I8B) :: i, j, counter, npl, ntp
+      integer(I4B) :: i, j
+      integer(I8B) :: counter, npl8, ntp8
 
-      ntp = int(self%nbody, kind=I8B)
-      npl = int(pl%nbody, kind=I8B)
-      associate(npltp => self%npltp)
-         npltp = npl * ntp
+      associate(ntp => self%nbody, npl => pl%nbody, npltp => self%npltp)
+         npl8 = int(npl, kind=I8B)
+         ntp8 = int(ntp, kind=I8B)
+         npltp = npl8 * ntp8 
          if (allocated(self%k_pltp)) deallocate(self%k_pltp) ! Reset the index array if it's been set previously
          allocate(self%k_pltp(2, npltp))
-         do i = 1_I8B, npl
-            counter = (i - 1_I8B) * npl + 1_I8B
-            do j = 1_I8B,  ntp
+         counter = 1_I8B
+         do i = 1, npl
+            do j = 1,  ntp
                self%k_pltp(1, counter) = i
                self%k_pltp(2, counter) = j
                counter = counter + 1_I8B
@@ -1643,7 +1644,8 @@ contains
       class(swiftest_parameters),   intent(inout) :: param  !! Current run configuration parameters
       ! Internals
       class(swiftest_pl), allocatable :: tmp !! The discarded body list.
-      integer(I4B) :: i, k, npl, nadd, nencmin, nenc_old, idnew1, idnew2, idold1, idold2
+      integer(I4B) :: i, npl, nadd, idnew1, idnew2, idold1, idold2
+      integer(I8B) :: k, nenc_old, nencmin
       logical, dimension(:), allocatable :: lmask
       class(encounter_list), allocatable :: plplenc_old
       logical :: lencounter
@@ -1708,7 +1710,7 @@ contains
          if (allocated(nbody_system%plpl_encounter)) then
[…]
