### 2023-05-25 d854c21
Improved the tslot finder using proper ieee arithmetic routine
