+++ src/CMakeLists.txt
@@ -25,6 +25,7 @@ SET(STRICT_MATH_FILES
             ${SRC}/helio/helio_step.f90
             ${SRC}/misc/lambda_function_module.f90
             ${SRC}/misc/solver_module.f90
+            ${SRC}/netcdf_io/netcdf_io_implementations.f90  
             ${SRC}/operator/operator_module.f90
             ${SRC}/operator/operator_cross.f90
             ${SRC}/operator/operator_mag.f90
@@ -69,7 +70,6 @@ SET(FAST_MATH_FILES
             ${SRC}/encounter/encounter_io.f90
             ${SRC}/encounter/encounter_util.f90
             ${SRC}/helio/helio_util.f90
-            ${SRC}/netcdf_io/netcdf_io_implementations.f90  
             ${SRC}/rmvs/rmvs_discard.f90
             ${SRC}/rmvs/rmvs_encounter_check.f90
             ${SRC}/rmvs/rmvs_util.f90
