### 2023-05-25 7b923bd
Cleaned up unused variables, type conversion inconsistencies (except for unit conversions using quad precision) and other things flagged by gfortran warnings
