+++ src/collision/collision_check.f90
@@ -77,7 +77,8 @@ contains
       ! Internals
       logical, dimension(:), allocatable        :: lcollision, lmask
       real(DP), dimension(NDIM)                 :: xr, vr
-      integer(I4B)                              :: i, j, k, nenc
+      integer(I4B)                              :: i, j
+      integer(I8B)                              :: k, nenc
       real(DP)                                  :: rlim, Gmtot
       logical                                   :: lany_closest
 
@@ -102,9 +103,9 @@ contains
             self%lclosest(:) = .false.
 
 #ifdef DOCONLOC
-            do concurrent(k = 1:nenc, lmask(k)) shared(self,pl,lmask, dt, lcollision) local(i,j,xr,vr,rlim,Gmtot)
+            do concurrent(k = 1_I8B:nenc, lmask(k)) shared(self,pl,lmask, dt, lcollision) local(i,j,xr,vr,rlim,Gmtot)
 #else
-            do concurrent(k = 1:nenc, lmask(k))
+            do concurrent(k = 1_I8B:nenc, lmask(k))
 #endif
                i = self%index1(k)
                j = self%index2(k)
@@ -120,7 +121,7 @@ contains
 
             if (lany_collision .or. lany_closest) then
                call pl%rh2rb(nbody_system%cb) ! Update the central body barycenteric position vector to get us out of DH and into bary
-               do k = 1, nenc
+               do k = 1_I8B, nenc
                   if (.not.lcollision(k) .and. .not. self%lclosest(k)) cycle
                   i = self%index1(k)
                   j = self%index2(k)
@@ -178,7 +179,8 @@ contains
       ! Internals
       logical, dimension(:), allocatable        :: lcollision, lmask
       real(DP), dimension(NDIM)                 :: xr, vr
-      integer(I4B)                              :: i, j, k, nenc
+      integer(I4B)                              :: i, j
+      integer(I8B)                              :: k, nenc
       logical                                   :: lany_closest
       character(len=STRMAX)                     :: timestr, idstri, idstrj, message
       class(collision_list_pltp), allocatable       :: tmp       
@@ -194,13 +196,13 @@ contains
 
             nenc = self%nenc
             allocate(lmask(nenc))
-             lmask(:) = (self%status(1:nenc) == ACTIVE) 
+            lmask(:) = (self%status(1:nenc) == ACTIVE) 
             select type(pl)
             class is (symba_pl)
-            select type(tp)
-            class is (symba_tp)
-               lmask(:) = lmask(:) .and. (tp%levelg(self%index2(1:nenc)) >= irec)
-            end select
+               select type(tp)
+               class is (symba_tp)
+                  lmask(:) = lmask(:) .and. (tp%levelg(self%index2(1:nenc)) >= irec)
+               end select
             end select
             if (.not.any(lmask(:))) return
 
@@ -209,9 +211,9 @@ contains
             self%lclosest(:) = .false.
 
 #ifdef DOCONLOC
-            do concurrent(k = 1:nenc, lmask(k)) shared(self,pl,tp,lmask, dt, lcollision) local(i,j,xr,vr)
+            do concurrent(k = 1_I8B:nenc, lmask(k)) shared(self,pl,tp,lmask, dt, lcollision) local(i,j,xr,vr)
 #else
-            do concurrent(k = 1:nenc, lmask(k))
+            do concurrent(k = 1_I8B:nenc, lmask(k))
 #endif
                i = self%index1(k)
                j = self%index2(k)
