### 2023-05-25 7472c5c
Fixed problem in SyMBA where nplm was not being updated properly after massive body discards
