### 2023-05-26 d46bec6
Restructured Docker and Singularity scripts. Added new GNU version of the Docker container that is smaller than the Intel one (but also slower).
