### 2023-05-26 64d6a30
Fixed typos and added some output
