+++ singularity/install.sh
@@ -4,6 +4,7 @@
 # $ . ./install.sh
 # 
 tag=${1:-intel}
+echo "Installing swiftest_driver.sif Singularity container and executable script from swiftest_driver:${tag} Docker container"
 singularity pull --force swiftest_driver.sif docker://daminton/swiftest_driver:${tag}
 cp -rf bin/swiftest_driver ../bin/
 source ./setenv.sh
