### 2023-05-26 6cce4f0
Removed unneeeded "container" argument because now the install.sh scripts will replace the executable with a script that calls the container
