+++ Dockerfile
@@ -1,0 +1,191 @@
+FROM ubuntu:20.04 as build
+
+# kick everything off
+RUN apt-get update && apt-get upgrade -y && \
+  DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+  ca-certificates curl git wget gpg-agent software-properties-common build-essential gnupg pkg-config && \
+  rm -rf /var/lib/apt/lists/*
+
+# Get CMAKE and install it
+RUN mkdir -p cmake/build && \
+   cd cmake/build && \
+   curl -LO https://github.com/Kitware/CMake/releases/download/v3.26.2/cmake-3.26.2-linux-x86_64.sh && \
+   /bin/bash cmake-3.26.2-linux-x86_64.sh --prefix=/usr/local --skip-license
+
+# Get the Intel compilers
+# download the key to system keyring
+RUN wget -O- https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB \
+| gpg --dearmor | tee /usr/share/keyrings/oneapi-archive-keyring.gpg > /dev/null
+# add signed entry to apt sources and configure the APT client to use Intel repository:
+RUN echo "deb [signed-by=/usr/share/keyrings/oneapi-archive-keyring.gpg] https://apt.repos.intel.com/oneapi all main" | tee /etc/apt/sources.list.d/oneAPI.list
+RUN apt-get -y update && apt-get upgrade -y
+RUN apt-get install -y intel-hpckit
+
+# Set Intel compiler environment variables
+ENV INTEL_DIR="/opt/intel/oneapi"
+ENV LANG=C.UTF-8
+ENV ACL_BOARD_VENDOR_PATH='/opt/Intel/OpenCLFPGA/oneAPI/Boards'
+ENV ADVISOR_2023_DIR='/opt/intel/oneapi/advisor/2023.1.0'
+ENV APM='/opt/intel/oneapi/advisor/2023.1.0/perfmodels'
+ENV CCL_CONFIGURATION='cpu_gpu_dpcpp'
+ENV CCL_ROOT='/opt/intel/oneapi/ccl/2021.9.0'
+ENV CLASSPATH='/opt/intel/oneapi/mpi/2021.9.0//lib/mpi.jar:/opt/intel/oneapi/dal/2023.1.0/lib/onedal.jar'
+ENV CLCK_ROOT='/opt/intel/oneapi/clck/2021.7.3'
+ENV CMAKE_PREFIX_PATH='/opt/intel/oneapi/tbb/2021.9.0/env/..:/opt/intel/oneapi/dnnl/2023.1.0/cpu_dpcpp_gpu_dpcpp/../lib/cmake:/opt/intel/oneapi/dal/2023.1.0:/opt/intel/oneapi/compiler/2023.1.0/linux/IntelDPCPP:/opt/intel/oneapi/ccl/2021.9.0/lib/cmake/oneCCL'
+ENV CMPLR_ROOT='/opt/intel/oneapi/compiler/2023.1.0'
[…]
