### 2023-06-04 ae98dd8
Added a universal writable .astropy directory to be able to call the docker image as a regular user to that simdir will have the correct ownership
