### 2023-06-04 da21136
Improved coarray outpt handling and preventing runs getting stuck on sync
