### 2023-06-05 0d4935b
Minor changes to troubleshoot energy calc NaN in new container.
