+++ singularity/install.sh
@@ -3,7 +3,7 @@
 # In order to use the executable script, the SWIFTEST_SIF environment variable must be set to point to the location of swiftest_driver.sif, which requires this script to be called via source:
 # $ . ./install.sh
 # 
-tag=${1:-intel}
+tag=${1:-latest}
 echo "Installing swiftest_driver.sif Singularity container and executable script from swiftest_driver:${tag} Docker container"
 singularity pull --force swiftest_driver.sif docker://daminton/swiftest_driver:${tag}
 cp -rf bin/swiftest_driver ../bin/
