### 2023-06-06 9bf02d0
Fixed compiler flag typos that prevented the -standard-semantics flag from being set. Also moved the standards compliance warnings to the DEBUG build
