+++ Dockerfile
@@ -2,24 +2,18 @@ FROM ubuntu:20.04 as build
 
 # kick everything off
 RUN apt-get update && apt-get upgrade -y && \
-  DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
-  ca-certificates curl git wget gpg-agent software-properties-common build-essential gnupg pkg-config && \
-  rm -rf /var/lib/apt/lists/*
-
-# Get CMAKE and install it
-RUN mkdir -p cmake/build && \
-   cd cmake/build && \
-   curl -LO https://github.com/Kitware/CMake/releases/download/v3.26.2/cmake-3.26.2-linux-x86_64.sh && \
-   /bin/bash cmake-3.26.2-linux-x86_64.sh --prefix=/usr/local --skip-license
-
-# Get the Intel compilers
-# download the key to system keyring
-RUN wget -O- https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB \
-| gpg --dearmor | tee /usr/share/keyrings/oneapi-archive-keyring.gpg > /dev/null
-# add signed entry to apt sources and configure the APT client to use Intel repository:
-RUN echo "deb [signed-by=/usr/share/keyrings/oneapi-archive-keyring.gpg] https://apt.repos.intel.com/oneapi all main" | tee /etc/apt/sources.list.d/oneAPI.list
-RUN apt-get -y update && apt-get upgrade -y
-RUN apt-get install -y intel-hpckit
+    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
+    ca-certificates curl git wget gpg-agent software-properties-common build-essential gnupg pkg-config && \
+    rm -rf /var/lib/apt/lists/* && \
+    mkdir -p cmake/build && \
+    cd cmake/build && \
+    curl -LO https://github.com/Kitware/CMake/releases/download/v3.26.2/cmake-3.26.2-linux-x86_64.sh && \
+    /bin/bash cmake-3.26.2-linux-x86_64.sh --prefix=/usr/local --skip-license && \
+    wget -O- https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB \
+            | gpg --dearmor | tee /usr/share/keyrings/oneapi-archive-keyring.gpg > /dev/null && \
+    echo "deb [signed-by=/usr/share/keyrings/oneapi-archive-keyring.gpg] https://apt.repos.intel.com/oneapi all main" | tee /etc/apt/sources.list.d/oneAPI.list && \
+    apt-get -y update && apt-get upgrade -y && \
+    apt-get install -y intel-hpckit && \
 
 # Set Intel compiler environment variables
 ENV INTEL_DIR="/opt/intel/oneapi"
@@ -94,24 +88,19 @@ ENV HDF5_INCLUDE_DIR="${HDF5_ROOT}/include"
 ENV HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
 
 # Get the HDF5, NetCDF-C and NetCDF-Fortran libraries
[…]
