### 2023-06-06 7f9cacb
Set debug flags correctly for the standards check
