### 2023-06-06 2aacbe3
Fixed line lengths based on standards compliance warning flags in debug mode
