+++ src/base/base_module.f90
@@ -22,90 +22,100 @@ module base
    !> User defined parameters that are read in from the parameters input file. 
    !>    Each paramter is initialized to a default values. 
    type, abstract :: base_parameters
-      character(STRMAX)                       :: integrator                             !! Name of the nbody integrator used
-      character(STRMAX)                       :: param_file_name                        !! The name of the parameter file
-      real(DP)                                :: t0                   =  0.0_DP         !! Integration reference time
-      real(DP)                                :: tstart               = -1.0_DP         !! Integration start time
-      real(DP)                                :: tstop                = -1.0_DP         !! Integration stop time
-      real(DP)                                :: dt                   = -1.0_DP         !! Time step
-      integer(I8B)                            :: iloop                = 0_I8B           !! Main loop counter
-      integer(I8B)                            :: nloops               = 0_I8B           !! Total number of loops to execute
-      integer(I8B)                            :: istart               = 0_I8B           !! Starting index for loop counter
-      integer(I4B)                            :: iout                 = 0               !! Output cadence counter
-      integer(I4B)                            :: idump                = 0               !! Dump cadence counter
-      integer(I4B)                            :: nout                 = 0               !! Current output step
-      integer(I4B)                            :: istep                = 0               !! Current value of istep (used for time stretching)
-      character(STRMAX)                       :: incbfile             = CB_INFILE       !! Name of input file for the central body
-      character(STRMAX)                       :: inplfile             = PL_INFILE       !! Name of input file for massive bodies
-      character(STRMAX)                       :: intpfile             = TP_INFILE       !! Name of input file for test particles
-      character(STRMAX)                       :: nc_in                = NC_INFILE       !! Name of system input file for NetCDF input
-      character(STRMAX)                       :: in_type              = "NETCDF_DOUBLE" !! Data representation type of input data files
-      character(STRMAX)                       :: in_form              = "XV"            !! Format of input data files ("EL" or ["XV"])
-      integer(I4B)                            :: istep_out            = -1              !! Number of time steps between saved outputs
-      integer(I4B)                            :: nstep_out            = -1              !! Total number of saved outputs
-      real(DP)                                :: fstep_out            = 1.0_DP          !! The output step time stretching factor
-      logical                                 :: ltstretch            = .false.         !! Whether to employ time stretching or not
-      character(STRMAX)                       :: outfile              = BIN_OUTFILE     !! Name of output binary file
-      character(STRMAX)                       :: out_type             = "NETCDF_DOUBLE" !! Binary format of output file
-      character(STRMAX)                       :: out_form             = "XVEL"          !! Data to write to output file
-      character(STRMAX)                       :: out_stat             = 'NEW'           !! Open status for output binary file
-      integer(I4B)                            :: dump_cadence         =  10             !! Number of output steps between dumping simulation data to file
-      real(DP)                                :: rmin                 = -1.0_DP         !! Minimum heliocentric radius for test particle
-      real(DP)                                :: rmax                 = -1.0_DP         !! Maximum heliocentric radius for test particle
-      real(DP)                                :: rmaxu                = -1.0_DP         !! Maximum unbound heliocentric radius for test particle
[…]
