### 2023-06-06 6b47437
Fixed typo in Dockerfile and set the Swiftest build type to RELEASE
