### 2023-06-06 20865a6
Removed do concurrents from operators and switched to manual cross product when in a do concurrent
