+++ src/fraggle/fraggle_generate.f90
@@ -59,7 +59,8 @@ contains
             call self%set_mass_dist(param) 
             call self%disrupt(nbody_system, param, t, lfailure)
             if (lfailure) then
-               call swiftest_io_log_one_message(COLLISION_LOG_OUT, "Fraggle failed to find a solution to match energy contraint. Treating this as a merge.") 
+               call swiftest_io_log_one_message(COLLISION_LOG_OUT, & 
+                                           "Fraggle failed to find a solution to match energy contraint. Treating this as a merge.") 
                call self%merge(nbody_system, param, t) ! Use the default collision model, which is merge
                return
             end if
@@ -131,8 +132,9 @@ contains
       real(DP), parameter                  :: fail_scale_initial = 1.0003_DP
       integer(I4B)                         :: nfrag_start
 
-      ! The minimization and linear solvers can sometimes lead to floating point exceptions. Rather than halting the code entirely if this occurs, we
-      ! can simply fail the attempt and try again. So we need to turn off any floating point exception halting modes temporarily 
+      ! The minimization and linear solvers can sometimes lead to floating point exceptions. Rather than halting the code entirely 
+      ! if this occurs, we can simply fail the attempt and try again. So we need to turn off any floating point exception halting 
+      ! modes temporarily 
       call ieee_get_halting_mode(IEEE_ALL,fpe_halting_modes)  ! Save the current halting modes so we can turn them off temporarily
       fpe_quiet_modes(:) = .false.
       call ieee_set_halting_mode(IEEE_ALL,fpe_quiet_modes)
@@ -168,7 +170,8 @@ contains
          if (.not.lfailure) then
             if (self%fragments%nbody /= nfrag_start) then
                write(message,*) self%fragments%nbody
-               call swiftest_io_log_one_message(COLLISION_LOG_OUT, "Fraggle found a solution with " // trim(adjustl(message)) // " fragments" )
+               call swiftest_io_log_one_message(COLLISION_LOG_OUT, "Fraggle found a solution with " // trim(adjustl(message)) &
+                                                                // " fragments" )
             end if
             call self%get_energy_and_momentum(nbody_system, param, phase="after")
 
@@ -176,7 +179,8 @@ contains
             dE = self%te(2) - self%te(1) 
 
             call swiftest_io_log_one_message(COLLISION_LOG_OUT, "All quantities in collision system natural units")
-            call swiftest_io_log_one_message(COLLISION_LOG_OUT, "*   Conversion factors (collision system units / nbody system units):")
+            call swiftest_io_log_one_message(COLLISION_LOG_OUT, & 
+                                             "*   Conversion factors (collision system units / nbody system units):")
             write(message,*) "*       Mass: ", self%mscale
             call swiftest_io_log_one_message(COLLISION_LOG_OUT, message)
             write(message,*) "*   Distance: ", self%dscale
@@ -200,7 +204,8 @@ contains
 
          end if
          call self%set_original_scale()
-         self%max_rot = MAX_ROT_SI * param%TU2S ! Re-compute the spin limit from scratch so it doesn't drift due to floating point errors every time we convert
+         self%max_rot = MAX_ROT_SI * param%TU2S ! Re-compute the spin limit from scratch so it doesn't drift due to floating point 
+                                                ! errors every time we convert
 
          ! Restore the big array
          if (lk_plpl) call pl%flatten(param)
@@ -248,7 +253,8 @@ contains
             end if
 
[…]
