+++ src/collision/collision_util.f90
@@ -146,13 +146,15 @@ contains
    module subroutine collision_util_get_energy_and_momentum(self, nbody_system, param, phase)
       !! Author: David A. Minton
       !!
-      !! Calculates total system energy in either the pre-collision outcome state (phase = "before") or the post-collision outcome state (lbefore = .false.)
+      !! Calculates total system energy in either the pre-collision outcome state (phase = "before") or the post-collision outcome 
+      !! state (lbefore = .false.)
       implicit none
       ! Arguments
       class(collision_basic),   intent(inout) :: self         !! Encounter collision system object
       class(base_nbody_system), intent(inout) :: nbody_system !! Swiftest nbody system object
       class(base_parameters),   intent(inout) :: param        !! Current Swiftest run configuration parameters
-      character(len=*),         intent(in)    :: phase        !! One of "before" or "after", indicating which phase of the calculation this needs to be done
+      character(len=*),         intent(in)    :: phase        !! One of "before" or "after", indicating which phase of the 
+                                                              !!    calculation this needs to be done
       ! Internals
       integer(I4B) :: i, phase_val, nfrag
 
@@ -179,9 +181,15 @@ contains
                do concurrent(i = 1:2)
 #endif
                   impactors%ke_orbit(i) = 0.5_DP * impactors%mass(i) * dot_product(impactors%vc(:,i), impactors%vc(:,i))
-                  impactors%ke_spin(i) = 0.5_DP * impactors%mass(i) * impactors%radius(i)**2 * impactors%Ip(3,i) * dot_product(impactors%rot(:,i), impactors%rot(:,i))
+                  impactors%ke_spin(i) = 0.5_DP * impactors%mass(i) * impactors%radius(i)**2 * impactors%Ip(3,i)  & 
+                                                * dot_product(impactors%rot(:,i), impactors%rot(:,i))
                   impactors%be(i) = -3 * impactors%Gmass(i) * impactors%mass(i) / (5 * impactors%radius(i))
-                  impactors%L_orbit(:,i) = impactors%mass(i) * impactors%rc(:,i) .cross. impactors%vc(:,i)
+                  impactors%L_orbit(1,i) = impactors%mass(i) * (impactors%rc(2,i) * impactors%vc(3,i) &
+                                                              - impactors%rc(3,i) * impactors%vc(2,i))
+                  impactors%L_orbit(2,i) = impactors%mass(i) * (impactors%rc(3,i) * impactors%vc(1,i) &
+                                                              - impactors%rc(1,i) * impactors%vc(3,i))
+                  impactors%L_orbit(3,i) = impactors%mass(i) * (impactors%rc(1,i) * impactors%vc(2,i) &
+                                                              - impactors%rc(2,i) * impactors%vc(1,i))
                   impactors%L_spin(:,i) = impactors%mass(i) * impactors%radius(i)**2 * impactors%Ip(3,i) * impactors%rot(:,i)
                end do
                self%L_orbit(:,phase_val) = sum(impactors%L_orbit(:,1:2),dim=2)
@@ -190,7 +198,8 @@ contains
                self%ke_orbit(phase_val) = sum(impactors%ke_orbit(1:2))
                self%ke_spin(phase_val) = sum(impactors%ke_spin(1:2))
                self%be(phase_val) = sum(impactors%be(1:2))
-               call swiftest_util_get_potential_energy(2, [(.true., i = 1, 2)], 0.0_DP, impactors%Gmass, impactors%mass, impactors%rb, self%pe(phase_val))
+               call swiftest_util_get_potential_energy(2, [(.true., i = 1, 2)], 0.0_DP, impactors%Gmass, impactors%mass, & 
+                                                            impactors%rb, self%pe(phase_val))
[…]

[…]
