### 2023-06-06 b8be368
Removed -standard-semantics from the default flag list, because it is not compatible with the pre-built NetCDF Fortran library. Kept it in the containerized version.
