### 2023-06-06 f7b8f46
Updated coarray code so it can compile in the container.
