+++ src/swiftest/swiftest_coarray.f90
@@ -54,6 +54,7 @@ contains
         return
     end subroutine swiftest_coarray_balance_system
 
+
     module subroutine swiftest_coarray_coclone_body(self)
         !! author: David A. Minton
          !!
@@ -93,6 +94,21 @@ contains
 
         return
     end subroutine swiftest_coarray_coclone_body
+
+    module subroutine swiftest_coarray_coclone_kin(self)
+        !! author: David A. Minton
+        !!
+        !! Broadcasts the image 1 object to all other images in a coarray 
+        implicit none
+        ! Arguments
+        class(swiftest_kinship),intent(inout),codimension[*]  :: self  !! Swiftest kinship object
+
+        call coclone(self%parent)
+        call coclone(self%nchild)
+        call coclone(self%child)
+
+        return
+     end subroutine swiftest_coarray_coclone_kin
 
     module subroutine swiftest_coarray_coclone_nc(self)
         !! author: David A. Minton
@@ -327,10 +343,6 @@ contains
         ! Internals
         integer(I4B) :: i
 
-        call self%cb%coclone()
-        call self%pl%coclone()
-        call self%tp%coclone()
-
         call coclone(self%maxid)
         call coclone(self%t)
         call coclone(self%GMtot)
@@ -474,7 +486,7 @@ contains
         integer(I4B), intent(in),optional :: src_img
         ! Internals
         type(swiftest_kinship), dimension(:), codimension[:], allocatable :: tmp
-        integer(I4B) :: img, si
+        integer(I4B) :: i, img, si
         integer(I4B), allocatable :: n[:]
         logical, allocatable :: isalloc[:]
 
@@ -484,21 +496,17 @@ contains
             si = 1
         end if
 
-        allocate(isalloc[*])
-        allocate(n[*])
+        sync all
         isalloc = allocated(var)
         if (isalloc) n = size(var)
         sync all 
         if (.not. isalloc[si]) return
 
         allocate(tmp(n[si])[*])
-        if (this_image() == si) then
-            do img = 1, num_images()
-                tmp(:)[img] = var 
-            end do
-            sync images(*)
-        else 
-            sync images(si)
+        do i = 1, n[si]
+            call tmp(i)%coclone()
+        end do
+        if (this_image() /= si) then
             if (allocated(var)) deallocate(var)
             allocate(var, source=tmp)
         end if
