### 2023-06-07 35e0099
Updates to singularity scripts and also added apptainer scripts
