+++ apptainer/setenv.sh
@@ -1,0 +1,4 @@
+#!/bin/sh --
+# This will set the SWIFTEST_SIF environment variable as long as it is executed by source.
+# $ . ./setenv.sh
+export SWIFTEST="${PWD}/swiftest.sif"
