### 2023-06-07 5f2e9bf
Fixed typo in Singularity/Apptainer install scripts
