### 2023-06-07 da9b300
More typo fixes and cleanup
