### 2023-06-07 c8ba090
Typos in install scripts
