+++ singularity/setenv.sh
@@ -1,4 +1,0 @@
-#!/bin/sh --
-# This will set the SWIFTEST_SIF environment variable as long as it is executed by source.
-# $ . ./setenv.sh
-export SWIFTEST_SIF="${PWD}/swiftest.sif"
