+++ docker/install.sh
@@ -1,5 +1,6 @@
 #!/bin/sh --
 tag=${1:-latest}
 echo "Installing swiftest:${tag} Docker container and executable script"
-docker pull daminton/swiftest:${tag}
+docker pull mintongroup/swiftest:${tag}
 cp -rf bin/swiftest ../bin/
+cp -rf bin/swiftest_driver ../bin/
