### 2023-06-09 bfe55af
Made a clean, streamlined Dockerfile that contains almost no dependencies in the end. It can also spit out the binary executables so it could be used as a build tool rather than running from inside the container
