+++ Dockerfile
@@ -1,85 +1,73 @@
-FROM ubuntu:20.04 as build
+# Copyright 2023 - David Minton
+# This file is part of Swiftest.
+# Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
+# as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+# Swiftest is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
+# of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+# You should have received a copy of the GNU General Public License along with Swiftest. 
+# If not, see: https://www.gnu.org/licenses. 
+#
+# This Dockerfile will build the Swiftest driver program with minimal external dependencies using the Intel Oneapi toolkit. 
+# This is done by building static versions of a minimal set of libraries that NetCDF-Fortran needs (Netcdf-C, HDF5, and Zlib). 
+# These, along with the Intel runtime libraries, are linked statically to the executable. Only the OS-specific libraries are linked
+# dynamically. 
 
-# kick everything off
-RUN apt-get update && apt-get upgrade -y && \
-    DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
-    ca-certificates curl git wget gpg-agent software-properties-common build-essential gnupg pkg-config && \
-    rm -rf /var/lib/apt/lists/* && \
-    mkdir -p cmake/build && \
-    cd cmake/build && \
-    curl -LO https://github.com/Kitware/CMake/releases/download/v3.26.2/cmake-3.26.2-linux-x86_64.sh && \
-    /bin/bash cmake-3.26.2-linux-x86_64.sh --prefix=/usr/local --skip-license && \
-    wget -O- https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS.PUB \
-            | gpg --dearmor | tee /usr/share/keyrings/oneapi-archive-keyring.gpg > /dev/null && \
-    echo "deb [signed-by=/usr/share/keyrings/oneapi-archive-keyring.gpg] https://apt.repos.intel.com/oneapi all main" | tee /etc/apt/sources.list.d/oneAPI.list && \
-    apt-get -y update && apt-get upgrade -y && \
-    apt-get install -y intel-hpckit 
+# This build target compiles all dependencies and the swiftest driver itself
+FROM intel/oneapi-hpckit:2023.1.0-devel-ubuntu22.04 as build
 
-# Set Intel compiler environment variables
-ENV INTEL_DIR="/opt/intel/oneapi"
[…]
