### 2023-06-09 1453773
Addedn environment.yml file to repo so that the conda environment can be created
