### 2023-06-09 a0e5edc
Fixed issue that was leading to an incompatibility between glibc versions due to the miniconda3 container being based on an older Ubuntu.
