### 2023-06-10 9a60cee
Minor tweaks to the Dockerfile to isolate the dependency build from the driver build
