### 2023-06-13 a454958
changed rot matrix calculation; passing extra arg to obl_acc
