+++ src/swiftest/swiftest_obl.f90
@@ -40,7 +40,7 @@ contains
       B(3,3) = +detinv * (A(1,1)*A(2,2) - A(1,2)*A(2,1))
    end function
 
-   module subroutine swiftest_obl_rot_matrix(self, nbody_system, rot_matrix, rot_matrix_inv)
+   module subroutine swiftest_obl_rot_matrix(n, rot, rot_matrix, rot_matrix_inv)
       !! author: Kaustub P. Anand
       !! 
       !! Generate a rotation matrix and its inverse to rotate the coordinate frame to align the rotation axis along the z axis for correct spin calculation
@@ -48,16 +48,18 @@ contains
 
       implicit none
       ! Arguments
-      class(swiftest_body),         intent(inout) :: self   !! Swiftest body object
-      class(swiftest_nbody_system), intent(inout) :: nbody_system !! Swiftest nbody system object
-      real(DP), dimension(NDIM, NDIM), intent(inout) :: rot_matrix ! rotation matrix and its inverse
-      real(DP), dimension(NDIM, NDIM), intent(inout) :: rot_matrix_inv ! inverse of the rotation matrix
+            !! class(swiftest_body),         intent(inout) :: self   !! Swiftest body object
+            !! class(swiftest_nbody_system), intent(inout) :: nbody_system !! Swiftest nbody system object
+      integer(I4B),             intent(in)           :: n               !! Number of bodies
+      real(DP), dimension(NDIM), intent(in)       :: rot             !! Central body rotation matrix
+      real(DP), dimension(NDIM, NDIM), intent(inout) :: rot_matrix      !! rotation matrix and its inverse
+      real(DP), dimension(NDIM, NDIM), intent(inout) :: rot_matrix_inv  !! inverse of the rotation matrix
 
       ! Internals
-      real(DP)     :: theta ! angle to rotate it through
-      real(DP), dimension(3) :: u, z_hat ! unit vector about which we rotate and z_hat
-      real(DP), dimension(3, 3) :: S_matrix ! rotation matrices
-      integer        :: i, j ! dummy variable
+      real(DP)     :: theta !! angle to rotate it through
+      real(DP), dimension(3) :: u, z_hat !! unit vector about which we rotate and z_hat
+      real(DP), dimension(3, 3) :: S_matrix !! rotation matrices
+      integer        :: i, j !! dummy variable
 
       ! Assumed that NDIM = 3
       
@@ -65,50 +67,47 @@ contains
       rot_matrix_inv(:, :) = 0.0_DP
       z_hat(:) = [0, 0, 1]
 
-      if (self%nbody == 0) return
+      if (n == 0) return
 
-      associate(n => self%nbody, cb => nbody_system%cb)
-         if (cb%rot(1) == 0 .and. cb%rot(2) == 0) then
-            do i = 1, NDIM
-                  rot_matrix(i, i) = 1.0
-                  rot_matrix_inv(i, i) = 1.0
-            end do
-
-            return ! rotation axis is about the z-axis, no need to change
-         end if
-         
-         u(:) = cb%rot(:) .cross. z_hat(:)
-         u(:) = .unit. u(:)
-         theta = acos(dot_product((.unit. cb%rot(:)), z_hat(:)))
-         
[…]
