### 2023-06-13 850d691
changed 'rotation' to 'ROTATION' in vec2xr in init_cond.py to ensure default Ip and rot is set
