### 2023-06-14 #5409 @mikegrudic
Hi folks,

I've read the paper and taken the code for a test drive, and think it's all looking pretty good overall. I was able to clone, build, and run the code with the instructions provided. The one thing I am missing from the checklist is either an automated test, or otherwise some documented procedure for verifying that the code is running some test case correctly. This could be as simple as a script that runs one of the standard test problems and checks that the output is within certain bounds.

Mike
