### 2023-06-15 3706292
fixed CMAKE flag typo in readme
