### 2023-06-20 4980834
Fixed a typo in comments; tested and checked a non-z-axis central body rotation vector. Precession works
