+++ setup_sim_ring_capture.py
@@ -1,246 +1,0 @@
-import numpy as np
-import swiftest
-import matplotlib.pyplot as plt
-from mpl_toolkits.mplot3d import axes3d
-import pandas as pd
-from astropy import constants as const
-
-seed = None
-rng = np.random.default_rng(seed=seed)
-n_bodies = int(1e4) # number of planet bodies
-array_shift = np.random.randint(0, n_bodies)
-tstop = 100 # rotation periods
-dt = 1.0e-4
-dt_unit = 'Haumea_rot_period'
-# dt_unit = 'd'
-dt_max = 0
-G = 6.6743e-11 # SI Units
-scratch_dr = '/scratch/bell/anand43/swiftest_runs/'
-body_dr = 'Haumea/'
-simdir = 'moons' # '2e3_part_radial'
-simdir = scratch_dr + body_dr + simdir
-
-# initial central body parameters
-# Haumea parameters taken from P. S. Jean Carvalho (2019)
-
-cb_mass = 4.006e21 # kg
-cb_radius = 711.9097e3 # m
-T_rotation = 3.915341 * 60 * 60 # hours to sec
-
-mass = cb_mass / cb_mass
-radius = cb_radius / cb_radius
-longest_axis = 1161e3 # m
-rmin = longest_axis / cb_radius
-volume = 4.0 / 3.0 * np.pi * radius**3
-density = mass / volume 
-rot = 2 * np.pi / T_rotation # rad/s
-rot = [[0, 0, rot]] 
-J2 = 0.305
-J4 = -0.216
-print(f'R_min = {rmin}')
-
-# moons parameters
-# values taken from JPL Horizons
-
-moon_name = ['Hi\'iaka', 'Namaka']
-moon_id = np.arange(10, len(moon_name) + 10, step = 1)
-moon_mass = np.array([1.151e9, 0.036e9]) / const.G / cb_mass
-moon_radius = np.array([160e3, 85e3]) / cb_radius   
-moon_rh = np.array([[-2.3446616e7, -3.6088243e7, -2.9864834e7], [-8.6355596e6, 1.5824178e7, 2.4140061e7]]) / cb_radius
-moon_vh = np.array([[-5.6613699e1, 3.5390515e0, 3.8830064e1], [6.6949687e1, 5.157315e1, -1.0614025e1]]) / cb_radius * T_rotation
-moon_T_rotation = np.array([9.8, np.inf]) * 60 * 60 # h to seconds
-moon_rot = []
-for i in range(len(moon_T_rotation)):
-    moon_rot.append([0, 0, 2 * np.pi / moon_T_rotation[i]]) # rad/s
-
-# set up swiftest simulation
-
-sim = swiftest.Simulation(simdir = simdir, integrator = "symba", tstop = tstop, dt = dt, istep_out = 1e4, dump_cadence = 10, rotation = True, collision_model = "FRAGGLE", rmin = rmin, rmax = 50 * rmin, MU2KG = cb_mass, DU2M = cb_radius, TU2S = T_rotation, init_cond_format = 'XV') # TU = dt_unit, MU_name = 'M_Haumea', DU_name = 'R_Haumea'
-
-#initial ring particle orbital parameters
-
-id_start = 100 # starting ID for bodies
-
-random_radius = rng.random(n_bodies) * (500 - 100) + 100 # randomize the radii of the ring particles between 100 and 500 (m)
-random_radius = random_radius / sim.DU2M
-random_mass = np.power(random_radius, 3) * 4.0 * np.pi / 3.0 * density # randomize the mass 
-
-z_sign = rng.random(n_bodies) - 0.5 # randomize the sign of the z-component
-
[…]
