### 2023-06-26 156556d
started python functions for spherical harmonics
