+++ python/swiftest/swiftest/sph_harmonics_gravity.py
@@ -1,0 +1,45 @@
+"""
+ Copyright 2023 - David Minton, Carlisle Wishard, Jennifer Pouplin, Jake Elliott, Dana Singh, & Kaustub Anand
+ This file is part of Swiftest.
+ Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
+ as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ Swiftest is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
+ of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ You should have received a copy of the GNU General Public License along with Swiftest. 
+ If not, see: https://www.gnu.org/licenses. 
+"""
+
+# python functions to read in and set up spherical harmonics coefficients for non-standard gravity terms
+# using pySHTOOLS (in python) by Mark W.
+# 
+
+import swiftest
+import numpy as np
+from astroquery.jplhorizons import Horizons
+import astropy.units as u
+from astropy.coordinates import SkyCoord
+import datetime
+import xarray as xr
+import pyshtools as pysh
+from typing import (
+    Literal,
+    Dict,
+    List,
+    Any
+)
+
+def clm_from_ellipsoid():
+    """
+    Initializes a Swiftest dataset containing the major planets of the Solar System at a particular data from JPL/Horizons
+
+    Parameters
+    ----------
+    param : dict
+        Swiftest paramuration parameters. This method uses the unit conversion factors to convert from JPL's AU-day system into the system specified in the param file
+    ephemerides_start_date : string
+        Date to use when obtaining the ephemerides in the format YYYY-MM-DD.
+
+    Returns
+    -------
+    ds : xarray dataset
+    """
