### 2023-06-26 1c12eac
Wrote a function to calculate the grav clm values for an ellipsoid
