+++ python/swiftest/swiftest/sph_harmonics_gravity.py
@@ -30,8 +30,8 @@ from typing import (
 
 def clm_from_ellipsoid(lmax, mass, density, a, b = None, c = None):
     """
-    Creates and returns the gravity coefficients for an ellipsoid with principal axes a, b, c upto a certain maximum degree lmax. Uses pyshtools.
-    No units necessary for a, b, & c. However, they need to be in the same units (DU).
+    Creates and returns the gravity coefficients for an ellipsoid with principal axes a, b, c upto a certain maximum degree lmax. 
+    Uses pyshtools. No units necessary for a, b, & c. However, they need to be in the same units (DU).
 
     Parameters
     ----------
@@ -68,6 +68,51 @@ def clm_from_ellipsoid(lmax, mass, density, a, b = None, c = None):
 
     # create shape grid and convert to Spherical Harmonics (.expand())
     shape_SH = pysh.SHGrid.from_ellipsoid(lmax = lmax, a = a, b = b, c = c).expand()
+
+    # get coefficients
+    clm_class = pysh.SHGravcoeffs.from_shape(shape_SH, rho = density, gm = Gmass)
+    clm = clm_class.to_array()
+
+    return clm
+
+def clm_from_relief(lmax, mass, density, grid):
+    """
+    Creates and returns the gravity coefficients for a body with a given DH grid upto a certain maximum degree lmax. 
+    Uses pyshtools.
+
+    Parameters
+    ----------
+    lmax : int
+        The maximum spherical harmonic degree resolvable by the grid.
+    mass : float
+        mass of the central body
+    density : float
+        density of the central body
+    grid : array, shape []
+        DH grid of the surface relief of the body
+
+    Returns
+    -------
+    clm : ndarry, shape (2, lmax+1, lmax+1)
+        numpy ndarray of the gravitational potential spherical harmonic coefficients. 
+        This is a three-dimensional array formatted as coeffs[i, degree, order], 
+        where i=0 corresponds to positive orders and i=1 to negative orders.
+
+    """
+
+    G = swiftest.constants.GC
+    Gmass = G * mass # SHTOOLS uses an SI G value, and divides it before using the mass
+            # FIND A BETTER WAY TO CHANGE UNITS
+
+    # cap lmax to 20 to ensure fast performance
+    lmax_limit = 20
+    if(lmax > lmax_limit): # FIND A BETTER WAY to judge this cut off point, i.e., relative change between coefficients
+        lmax = lmax_limit
+        print(f'Setting maximum spherical harmonic degree to {lmax_limit}')
+
+    # convert to spherical harmonics
+    shape_SH = pysh.SHGrid.from_array(grid).expand()
+        # shape_SH = SHExpandDH(grid)
 
     # get coefficients
     clm_class = pysh.SHGravcoeffs.from_shape(shape_SH, rho = density, gm = Gmass)
