### 2023-06-26 b0f86fa
Update Git repo location in README.md
I followed the instructions and cloned from the Purdue repo and ran into an issue (using `.not.` on an integer) that appears to have been fixed in the GitHub version, so I presume the README needs to be updated.
