### 2023-06-27 #5409 @warrickball
Hello!

I see some activity on trying to install Swiftest, which everyone has succeeded at one way or another. Because no-one else was going down that road, I also tried building Swiftest on our cluster with `ifort` and the suggested modules, which worked once I told CMake to use ifort, not GCC. @carlislewishard, it looks like some bugfixes and docs updates are still required.

I also see that there aren't any tests. Though JOSS prefers these to be automated, it's easier in some languages than others, but you do need tests that can be run and documentation on how users can do so. This might just be a case of tweaking your examples so that they check some output and report whether the calculation has finished successfully. As a personal remark, I find enormous value in automating this, beyond just getting quick feedback on development errors. E.g. if, as a user, I can find a GitHub action to test a program, then I can also look at that action to see how the code is being built successfully.

@hannorein and @aatrani, have you had a chance to test the functionality yet? My local build did run one or two examples but, since I'm not actually reviewing, I haven't checked any output.
