### 2023-06-27 #5409 @carlislewishard
Hi all, I apologize for my delay on responding to some of your comments. I recently defended and I have been consumed by thesis edits. I will try to address issues with the README and various compilation flags/modules/requirements etc today. I have been working on turning one of our examples into a fleshed out automated test but I will admit that I have little experience setting up automated tests with CMAKE so it may take me a bit more time. Thank you for your patience.
