### 2023-06-27 aebd770
Update README.md
fixed link to clone repo, also added in cmake flags for gfortran
