### 2023-06-27 f1bfc2f
Update README.md
added flag to specify fortran compiler
