### 2023-06-27 47b0ad5
added error analysis python script
