### 2023-06-27 e5181a6
Updated output_reader python script
