+++ examples/Basic_Simulation/output_reader.py
@@ -43,4 +43,4 @@ ax.set_ylim(0, 0.4)
 ax.text(1.5, 0.35, f"t = {sim.data['time'].isel(time=0).values} years", size=10, ha="left")
 plt.tight_layout()
 plt.show()
-fig.savefig("output.eps", dpi=300, facecolor='white', transparent=False, bbox_inches="tight")
+fig.savefig("../examples/Basic_Simulation/output.eps", dpi=300, facecolor='white', transparent=False, bbox_inches="tight")
