### 2023-06-27 99aff8e
prevented output_reader from showing plot
