### 2023-06-27 #5409 @warrickball
> I recently defended...

Congratulations!

Regarding the "automated" test, note that the requirement is to have instructions like "Once the software is built, you can test it by running <some sequence of commands>." Though the checklist item says **Automated tests**, it doesn't have to be automated in the sense of continuous integration (e.g. GitHub actions), thought that's ideal.

Also, as an editor I'm mainly concerned with whether the review is progressing, not necessarily how long it takes, especially compared to things like thesis corrections!
