### 2023-06-27 #5409 @carlislewishard
Thank you!

That makes sense, thank you for the clarification. I believe that I have implemented a truly automated test using one of our example scripts (see Issue #5) but if that automated test poses problems for any of the reviewers, its good to know that I can adjust things to be more user friendly.
