### 2023-06-28 #5409 @hannorein
@warrickball No. I still struggle getting it to run. @profminton mentioned that he is working on putting together a container with the complete Swiftest project (Python + Fortran). If that still upcoming, then it would make thing easier (hopefully), if not, then I'll try once more installing things from scratch.
