### 2023-06-29 8814dff
Update initial_conditions.py
set a seed for the random number generator, clarified names of bodies, and adjusted eccentricity and inclination of bodies to be less energetic
