### 2023-06-29 6181c88
Update output_reader.py
adjusted bounds of plot to account for less excited system
