+++ examples/Basic_Simulation/output_reader.py
@@ -39,7 +39,7 @@ fig, ax = plt.subplots(nrows=1, ncols=1, figsize=(5,3))
 ax.set(xlabel="Semimajor Axis (AU)", ylabel="Eccentricity", title="Simulation Initial Conditions (t=0)")
 ax.scatter(sim.data['a'].isel(time=0), sim.data['e'].isel(time=0), c=colors, s=sizes, edgecolor='black')
 ax.set_xlim(0, 2.0)
-ax.set_ylim(0, 0.4)
-ax.text(1.5, 0.35, f"t = {sim.data['time'].isel(time=0).values} years", size=10, ha="left")
+ax.set_ylim(0, 0.25)
+ax.text(1.5, 0.2, f"t = {sim.data['time'].isel(time=0).values} years", size=10, ha="left")
 plt.tight_layout()
 fig.savefig("../examples/Basic_Simulation/output.eps", dpi=300, facecolor='white', transparent=False, bbox_inches="tight")
