### 2023-06-29 6a82e69
Update errors.py
adjusted energy error limit slightly to account for less energetic system
