+++ python/swiftest/requirements.txt
@@ -1,86 +1,16 @@
-argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613036642480/work
-astropy==4.2.1
-astroquery==0.4.2
-async-generator==1.10
-attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
-backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
-beautifulsoup4==4.9.3
-bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
-certifi==2021.5.30
-cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
-chardet==4.0.0
-cryptography==3.4.7
-cycler==0.10.0
-decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
-defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
-entrypoints==0.3
-html5lib==1.1
-idna==2.10
-importlib-metadata==4.5.0
-ipykernel==5.5.5
-ipython @ file:///tmp/build/80754af9/ipython_1617118429768/work
-ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
-ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
-jedi==0.17.0
-jeepney==0.6.0
-Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
-jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
-jupyter==1.0.0
-jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
-jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
-jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213308260/work
-jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
-jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
-keyring==23.0.1
-kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
-MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
-matplotlib==3.4.2
-mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
-mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853974840/work
-mkl-service==2.3.0
-nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
-nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914821128/work
-nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
-nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
-notebook @ file:///tmp/build/80754af9/notebook_1621523661196/work
-numpy==1.20.3
-olefile==0.46
-packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
-pandas==1.2.4
-pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120451932/work
-parso @ file:///tmp/build/80754af9/parso_1617223946239/work
-pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
-pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
[…]
