### 2023-07-25 6c7769c
Fixed add_solar_system_body method to be able to handle small bodies properly
