### 2023-07-25 9ee517a
Moved automated test to the correct CmakeLists.txt file
