### 2023-07-26 b564e69
started work on coding spherical harmonics for gravitational acceleration
