### 2023-07-27 2e1e93d
Overhauled initial conditions generator for add_solar_system_body. Can more robustly retrieve initial conditions for many more bodies in the solar system, including physical properties like mass, radius, and rotation rates if they exist. Updated the Basic_Simulation example to reflect the new capabilities.
