+++ python/swiftest/swiftest/simulation_class.py
@@ -2157,36 +2157,31 @@ class Simulation(object):
         return range_dict
 
     def add_solar_system_body(self,
-                              name: str | List[str],
+                              name: str | List[str] | None = None,
                               ephemeris_id: int | List[int] | None = None,
                               date: str | None = None,
                               source: str = "HORIZONS"):
         """
-        Adds a solar system body to an existing simulation Dataset from the JPL Horizons ephemeris service.
+        Adds a solar system body to an existing simulation Dataset from the JPL Horizons ephemeris service. The JPL Horizons service
+        will be searched for a body matching the string passed by `name`, or alternatively `ephemeris_id` if passed. Bodies will be
+        named in the Swiftest initial conditions Dataset using `name`. Use `ephemeris_id` to have finer control over which body is
+        searched in Horizons while using a custom name.
+        
+        If `name` is not passed, then the target name property is used as the name. You must pass either `name` and/or `ephemeris_id`
+        
+        When passing `name` == "Earth" or `name` == "Pluto", it a body is generated that has initial conditions matching the system
+        barycenter and mass equal to the sum of Earth+Moon or Pluto+Charon. 
+        
+        To obtain initial conditions for either Earth or Pluto alone, pass `ephemeris_id` == "399" for Earth or 
+        `ephemeris_id` == "999" for Pluto.  
 
-        The following are name/ephemeris_id pairs that are currently known to Swiftest, and therefore have
-        physical properties that can be used to make massive bodies.
-
-        Sun     : 0
-        Mercury : 1
-        Venus   : 2
-        Earth   : 3
-        Mars    : 4
-        Jupiter : 5
-        Saturn  : 6
-        Uranus  : 7
-        Neptune : 8
-        Pluto   : 9
 
         Parameters
         ----------
-        name : str | List[str]
-            Add solar system body by name.
-            Bodies not on this list will be added as test particles, but additional properties can be added later if
-            desired.
+        name : str, optional | List[str]
+            Add solar system body by name. This will be the name used in the Swiftest initial conditions Dataset unless not supplied 
         ephemeris_id : int | List[int], optional but must be the same length as `name` if passed.
-            Use id if the body you wish to add is recognized by Swiftest. In that case, the id is passed to the
-            ephemeris service and the name is used. The body specified by `id` supercedes that given by `name`.
+            In that case, the id is passed to the ephemeris service and the name is used. 
         date : str, optional
             ISO-formatted date sto use when obtaining the ephemerides in the format YYYY-MM-DD. Defaults to value
             set by `set_ephemeris_date`.
@@ -2198,16 +2193,22 @@ class Simulation(object):
         None
             initial conditions data stored as an Xarray Dataset in the init_cond instance variable
         """
+        
+        if name == None and ephemeris_id == None:
+            warnings.warn("Either `name` and/or `ephemeris_id` must be supplied to add_solar_system_body")
+            return None
 
-        if type(name) is str:
-            name = [name]
         if ephemeris_id is not None:
-            if type(ephemeris_id) is int:
+            if type(ephemeris_id) is int or type(ephemeris_id) is str:
                 ephemeris_id = [ephemeris_id]
-            if len(ephemeris_id) != len(name):
+            if name is None:
+                name = [None] * len(ephemeris_id)
+            elif len(ephemeris_id) != len(name):
                 warnings.warn(f"The length of ephemeris_id ({len(ephemeris_id)}) does not match the length of name ({len(name)})",stacklevel=2)
                 return None
         else:
+            if type(name) is str:
+                name = [name]
             ephemeris_id = [None] * len(name)
 
[…]
