+++ examples/Basic_Simulation/initial_conditions.py
@@ -42,7 +42,13 @@ sim.clean()
 rng = default_rng(seed=123)
 
 # Add the modern planets and the Sun using the JPL Horizons Database.
-sim.add_solar_system_body(["Sun","Mercury","Venus","Earth","Mars","Jupiter","Saturn","Uranus","Neptune","Pluto"])
+sim.add_solar_system_body(["Sun","Mercury","Venus","Earth","Mars","Jupiter","Saturn","Uranus","Neptune"])
+
+# Add in some main belt asteroids
+sim.add_solar_system_body(name=["Ceres","Vesta","Pallas","Hygiea"])
+
+# Add in some big KBOs
+sim.add_solar_system_body(name=["Pluto","Eris","Sedna","Haumea","Makemake","Quaoar","Orcus","Gonggong","Salacia"])
 
 # Add 5 user-defined massive bodies.
 npl         = 5
