### 2023-07-28 2b6885f
Updated Dockerfile to enable passing extra cmake build options via the argument EXTRA_CMAKE_OPTIONS
