### 2023-07-28 ea63fdd
Updated README.md with new directions on compiling
