### 2023-07-28 2975351
Finished changes to the README to improve the compiling instructions
