### 2023-07-28 fa59443
Added additional row in the CMake options table to show how to add aditional flags, such as include paths
