### 2023-07-28 5970e5e
Added sanity check to ensure tstep_out <= tstop
