### 2023-07-28 486b3a3
Fixed typo in Dockerfile
