### 2023-07-28 4eac910
Improved automated test with unittest
