+++ test/test_suite.py
@@ -1,0 +1,161 @@
+#!/usr/bin/env python3
+
+"""
+ Copyright 2023 - David Minton, Carlisle Wishard, Jennifer Pouplin, Jake Elliott, & Dana Singh
+ This file is part of Swiftest.
+ Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
+ as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ Swiftest is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
+ of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ You should have received a copy of the GNU General Public License along with Swiftest. 
+ If not, see: https://www.gnu.org/licenses. 
+"""
+
+"""
+Tests that energy and momentum errors are within tolerances in a Swiftest simulation
+
+Input
+------
+
+Output
+------
+None
+"""
+
+import swiftest
+import unittest
+import os
+import numpy as np
+from numpy.random import default_rng
+
+rng = default_rng(seed=123)
+
+major_bodies = ["Sun","Mercury","Venus","Earth","Mars","Jupiter","Saturn","Uranus","Neptune"]
+param = {}
+
+
+class TestSwiftest(unittest.TestCase):
+   
+   def test01_gen_ic(self):
+      """
+      Tests that Swiftest is able to successfully generate a set of initial conditions in a file without any exceptions being raised
+      """
+      print("\ntest_gen_ic: Test whether we can generate simulation initial conditions test")
+      # Files that are expected to be generated:
+      simdir = "simdata"
+      file_list = [simdir, os.path.join(simdir,"param.in"), os.path.join(simdir,"init_cond.nc")]
+      
+      sim = swiftest.Simulation()
+      sim.clean()
+
+      # Add the modern planets and the Sun using the JPL Horizons Database.
+      # Add the modern planets and the Sun using the JPL Horizons Database.
+      sim.add_solar_system_body(major_bodies)
+   
+      # Display the run configuration parameters.
+      param = sim.get_parameter(verbose=False)
+      sim.save()
+      
+      for f in file_list:
+         self.assertTrue(os.path.exists(f))
+         
+   def test02_read_ic(self):
+      """
+      Tests that Swiftest is able to read a set of pre-existing initial conditions files and that they contain the correct data
+      """
+      print("\ntest_read_ic: Test whether we can read back initial conditions files created by test_gen_ic")
+      sim = swiftest.Simulation(read_param=True)
+      # Check if all names in Dataset read in from file match the expected list of names
+      self.assertTrue((major_bodies == sim.init_cond['name']).all(), msg="Name mismatch in Dataset")
+      
+      # Check to see if all parameter values read in from file match the expected parameters saved when generating the file
+      self.assertTrue(all([v == param[k] for k,v in sim.param.items() if k in param]))
+     
+   def test03_integrators(self):
+      """
+      Tests that Swiftest is able to integrate a collection of massive bodies and test particles with all available integrators
+      """ 
[…]
