+++ test/test_suite.py
@@ -158,4 +158,3 @@ class TestSwiftest(unittest.TestCase):
       
 if __name__ == '__main__':
    unittest.main()
-
