### 2023-07-29 b2439e7
Changed test to use ADD_TEST
