### 2023-07-31 5e57497
Removed conda-build from default environment because it was wreaking havoc on vs code terminals. Also improved the Dockerfile so that conda-build doesn't have to be in the swiftest-env environment
