+++ Dockerfile
@@ -136,20 +136,18 @@ ENV SHELL="/bin/bash"
 ENV PATH="/opt/conda/bin:${PATH}"
 ENV LD_LIBRARY_PATH="/usr/local/lib"
 
+COPY --from=build_driver /usr/local/bin/swiftest_driver /opt/conda/bin/swiftest_driver
+COPY ./python/. /opt/conda/pkgs/
+COPY environment.yml .
+
 RUN conda update --all -y && \
   conda install conda-libmamba-solver -y && \
-  conda config --set solver libmamba 
-
-COPY environment.yml .
-RUN conda env create -f environment.yml && \
+  conda config --set solver libmamba && \
+  conda install conda-build -y && \
+  conda env create -f environment.yml && \
+  cd /opt/conda/pkgs/swiftest && conda develop . --name swiftest-env && \
   conda init bash && \
-  echo "conda activate swiftest-env" >> ~/.bashrc 
-
-COPY ./python/. /opt/conda/pkgs/
-COPY --from=build_driver /usr/local/bin/swiftest_driver /opt/conda/bin/swiftest_driver
-
-# Start new shell to activate the environment and install Swiftest
-RUN cd /opt/conda/pkgs/swiftest && conda develop . && \
+  echo "conda activate swiftest-env" >> ~/.bashrc  && \
   conda clean --all -y && \
   mkdir -p /.astropy && \
   chmod -R 777 /.astropy && \
