### 2023-07-31 b3bf286
Improved energy/momentum error test to compute the slope of the error and angular momentum growth instead of just the point at the end, which is more instructive for error checking
