### 2023-07-31 4d3211b
Rearranged because -u is a valid option for gfortran.
