### 2023-08-01 05a00cb
Added static linking of gfortran and gcc libraries
