### 2023-08-01 bd1fb58
Reorganized testing and made the compiler-option finding much more robust and efficient by getting the compiler id first instead of trying every possible flag.
