+++ cmake/Modules/SetFortranFlags.cmake
@@ -46,15 +46,6 @@ ELSE()
     MESSAGE(FATAL_ERROR "CMAKE_BUILD_TYPE not valid! ${BUILD_TYPE_MSG}")
 ENDIF(BT STREQUAL "RELEASE")
 
-IF (CMAKE_Fortran_COMPILER_ID STREQUAL "GNU")
-    IF (APPLE)
-        SET(MACHINE_CODE_VALUE "tune=native" CACHE STRING "Tells the compiler which processor features it may target, including which instruction sets and optimizations it may generate.")
-    ELSE ()
-        SET(MACHINE_CODE_VALUE "arch=native" CACHE STRING "Tells the compiler which processor features it may target, including which instruction sets and optimizations it may generate.")
-    ENDIF ()
-ELSE ()
-    SET(MACHINE_CODE_VALUE "host" CACHE STRING "Tells the compiler which processor features it may target, including which instruction sets and optimizations it may generate.")
-ENDIF ()
 
 #########################################################
 # If the compiler flags have already been set, return now
@@ -77,340 +68,456 @@ ENDIF(CMAKE_Fortran_FLAGS_RELEASE AND CMAKE_Fortran_FLAGS_TESTING AND CMAKE_Fort
 #####################
 
 # Free form
-SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                Fortran "-ffree-form" # GNU
-                ) 
+IF (COMPILER_OPTIONS STREQUAL "GNU")
+        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                        Fortran "-ffree-form" # GNU
+                        ) 
 
-# Don't add underscores in symbols for C-compatability
-SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                Fortran "-fno-underscoring" # GNU
-                ) 
-# Compile code assuming that IEEE signaling NaNs may generate user-visible traps during floating-point operations. 
-# Setting this option disables optimizations that may change the number of exceptions visible with signaling NaNs. 
-SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                Fortran "-fsignaling-nans " # GNU
-                ) 
-               
-# Allows for lines longer than 80 characters without truncation
-SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                Fortran "-ffree-line-length-512" # GNU (gfortran)
-                )
+        # Don't add underscores in symbols for C-compatability
+        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                        Fortran "-fno-underscoring" # GNU
+                        ) 
+        # Compile code assuming that IEEE signaling NaNs may generate user-visible traps during floating-point operations. 
+        # Setting this option disables optimizations that may change the number of exceptions visible with signaling NaNs. 
+        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                        Fortran "-fsignaling-nans " # GNU
+                        ) 
+                
+        # Allows for lines longer than 80 characters without truncation
+        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+                        Fortran "-ffree-line-length-512" # GNU (gfortran)
+                        )
+ELSEIF (COMPILER_OPTIONS STREQUAL "Intel")
+        # Disables right margin wrapping in list-directed output
[…]
