### 2023-08-02 517a2f4
Fixed issues that were causing the horizons initial conditions generator to fail if a valid body could not be identified.
