### 2023-08-02 f9baf87
Improved the horizons query method to accept additional keyword arguments to pass to the astroquery.Horizons method. This allows one to disambiguate names by passing 'id_type' as an optional keyword
