### 2023-08-02 c3ea9f2
Improved the version control of version number. Added function to CMakeLists.txt that will update the version number for both the Fortran and Python source files based on the contents of version.txt. Also improved how base_util_exit output is handled by allowing output to be directed to the proper display unit set by param%display_unit
