+++ src/symba/symba_step.f90
@@ -197,7 +197,7 @@ contains
                write(*, *) "SWIFTEST Warning:"
                write(*, *) "   In symba_step_recur_system, local time step is too small"
                write(*, *) "   Roundoff error will be important!"
-               call base_util_exit(FAILURE)
+               call base_util_exit(FAILURE,param%display_unit)
             END IF
             irecp = ireci + 1
             if (ireci == 0) then
