### 2023-08-02 081439d
Fixed another edge case when parsing Horizons output for mass
