### 2023-08-04 6bee96a
Edited the normalization for C_lm. Rewrite the g formula
