+++ src/swiftest/swiftest_sph.f90
@@ -13,10 +13,11 @@
 submodule (swiftest) s_swiftest_sph
 use swiftest
 use operators
+use SHTOOLS
 
 contains
 
-    module subroutine swiftest_sph_g_acc_one(gm, r_0, phi, theta, rh, c_lm, g_sph)
+    module subroutine swiftest_sph_g_acc_one(gm, r_0, phi, theta, rh, c_lm, g_sph, l_max)
         !! author: Kaustub P. Anand
         !!
         !! Calculate the acceleration terms for one pair of bodies given c_lm, theta, phi, r
@@ -24,35 +25,43 @@ contains
 
         implicit none
         ! Arguments
-        real(DP), intent(in)        :: gm !! GMass of the central body
-        real(DP), intent(in)        :: r_0 !! radius of the central body
-        real(DP), intent(in)        :: phi !! Azimuthal/Phase angle (radians)
-        real(DP), intent(in)        :: theta !! Inclination/Zenith angle (radians)
-        real(DP), intent(in), dimension(NDIM)       :: rh !! distance vector of body
-        real(DP), intent(in), dimension(2, :, :)    :: c_lm !! Spherical Harmonic coefficients
-        real(DP), intent(out), dimension(NDIM)      :: g_sph !! acceleration vector
+        real(DP), intent(in)        :: gm                       !! GMass of the central body
+        real(DP), intent(in)        :: r_0                      !! radius of the central body
+        real(DP), intent(in)        :: phi                      !! Azimuthal/Phase angle (radians)
+        real(DP), intent(in)        :: theta                    !! Inclination/Zenith angle (radians)
+        real(DP), intent(in), dimension(NDIM)       :: rh       !! distance vector of body
+        real(DP), intent(in), dimension(2, :, :)    :: c_lm     !! Spherical Harmonic coefficients
+        real(DP), intent(out), dimension(NDIM)      :: g_sph    !! acceleration vector
 
         ! Internals
-        integer        :: l, m !! SPH coefficients
-        real(DP)       :: r_mag !! magnitude of rh
+        integer        :: l, m              !! SPH coefficients
+        real(DP)       :: r_mag             !! magnitude of rh
+        real(DP)       :: p                 !! Associated Lengendre Polynomials at a given cos(theta)
+        integer        :: l_max             !! max Spherical Harmonic l order value
 
         r_mag = sqrt(dot_product(rh(:), rh(:)))
-
-        ! WHAT DO I DO WITH THE COMPLEX TERM????
+        l_max = size(c_lm, 2)
+        PlBar(p, lmax, cos(theta))      ! Associated Legendre Polynomials
 
         do l = 1, l_max
             do m = 1, l
 
-                swiftest_sph_dylm_dtheta(ylm, ylm1, theta, phi, l, m, dylm_dtheta, dyl_m_dtheta)
-                ! m > 0
-                g_sph(1) += gm * r_0**l / r_mag**(l + 1) * c_lm[0, l, m] * (dylm_dtheta / (rh(3) * cos(phi)) - (l + 1) * rh(1) * ylm / r_mag**2 - i * m * ylm / rh(2)) ! i = sqrt(-1)
[…]
