### 2023-08-04 2b3d7c0
Rearranged the CMake build process to first build source code into a library and link the driver executable to the library.
