### 2023-08-04 f98f9ae
Turned the swiftest_driver into a subroutine that is now called by a simple main program called main.f90. This is so that I can eventually convert the driver executable into a library function called directly inside Python using f2py
