+++ src/swiftest/swiftest_driver.f90
@@ -6,186 +6,194 @@
 !! of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 !! You should have received a copy of the GNU General Public License along with Swiftest. 
 !! If not, see: https://www.gnu.org/licenses. 
+submodule(swiftest) s_swiftest_driver
+contains
 
-program swiftest_driver
-   !! author: David A. Minton
-   !!
-   !! Driver program for the Swiftest integrators. Unlike the earlier Swift and Swifter drivers, in Swiftest all integrators 
-   !!    are run from this single program. 
-   !!
-   !! Adapted from Swifter by David E. Kaufmann's Swifter driver programs swifter_[bs,helio,ra15,rmvs,symba,tu4,whm].f90
-   !! Adapted from Hal Levison and Martin Duncan's Swift driver programs
-   use netcdf
-   use swiftest
-   implicit none
+   subroutine swiftest_driver(integrator, param_file_name, display_style)
 
-   class(swiftest_nbody_system), allocatable :: nbody_system      !! Polymorphic object containing the nbody system to be integrated
-   type(swiftest_parameters)                 :: param             !! Run configuration parameters
-   class(swiftest_storage),      allocatable :: system_history    !! Stores the system history between output dumps
-   character(len=:), allocatable             :: integrator        !! Integrator type code (see globals for symbolic names)
-   character(len=:), allocatable             :: param_file_name   !! Name of the file containing user-defined parameters
-   character(len=:), allocatable             :: display_style     !! Style of the output display {"STANDARD", "COMPACT", "PROGRESS"}). Default is "STANDARD"
-   type(walltimer)                           :: integration_timer !! Object used for computing elapsed wall time
+      !! author: David A. Minton
+      !!
+      !! Driver program for the Swiftest integrators. Unlike the earlier Swift and Swifter drivers, in Swiftest all integrators 
+      !!    are run from this single program. 
+      !!
+      !! Adapted from Swifter by David E. Kaufmann's Swifter driver programs swifter_[bs,helio,ra15,rmvs,symba,tu4,whm].f90
+      !! Adapted from Hal Levison and Martin Duncan's Swift driver programs
+      use netcdf
+      use swiftest
+      implicit none
 
-   call swiftest_io_get_args(integrator, param_file_name, display_style)
+      ! Arguments
+      character(len=:), intent(in), allocatable :: integrator      !! Symbolic code of the requested integrator  
+      character(len=:), intent(in), allocatable :: param_file_name !! Name of the input parameters file
+      character(len=:), intent(in), allocatable :: display_style   !! Style of the output display {"STANDARD", "COMPACT", "PROGRESS"}). Default is "STANDARD"   
 
-   !> Read in the user-defined parameters file and the initial conditions of the nbody_system
-   param%integrator = trim(adjustl(integrator))
-   param%display_style = trim(adjustl(display_style))
-   call param%read_in(param_file_name)
+      ! Internals
+      class(swiftest_nbody_system), allocatable :: nbody_system      !! Polymorphic object containing the nbody system to be integrated
+      type(swiftest_parameters)                 :: param             !! Run configuration parameters
+      class(swiftest_storage),      allocatable :: system_history    !! Stores the system history between output dumps
+      type(walltimer)                           :: integration_timer !! Object used for computing elapsed wall time
+
+      !> Read in the user-defined parameters file and the initial conditions of the nbody_system
+      param%integrator = trim(adjustl(integrator))
+      param%display_style = trim(adjustl(display_style))
+      call param%read_in(param_file_name)
 #ifdef COARRAY
[…]
