+++ src/swiftest/swiftest_driver.f90
@@ -9,7 +9,7 @@
 submodule(swiftest) s_swiftest_driver
 contains
 
-   subroutine swiftest_driver(integrator, param_file_name, display_style)
+   module subroutine swiftest_driver(integrator, param_file_name, display_style)
 
       !! author: David A. Minton
       !!
@@ -18,8 +18,6 @@ contains
       !!
       !! Adapted from Swifter by David E. Kaufmann's Swifter driver programs swifter_[bs,helio,ra15,rmvs,symba,tu4,whm].f90
       !! Adapted from Hal Levison and Martin Duncan's Swift driver programs
-      use netcdf
-      use swiftest
       implicit none
 
       ! Arguments
