### 2023-08-04 023c420
Fixed issues with Dockerfile when compiling the new code structure with the library.
