+++ CMakeLists.txt
@@ -68,6 +68,11 @@ SET(MOD ${CMAKE_SOURCE_DIR}/include)
 #SET(TEST ${CMAKE_SOURCE_DIR}/test)
 SET(PY ${CMAKE_SOURCE_DIR}/python/swiftest)
 
+set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${LIB})
+set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${LIB})
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${BIN})
+
+
 FUNCTION(REPLACE_VERSION IN_FILE LANGUAGE)
     # Make list of strings from file
     FILE(STRINGS ${IN_FILE} LINES)
