+++ src/CMakeLists.txt
@@ -105,7 +105,8 @@ ENDIF ()
 SET_SOURCE_FILES_PROPERTIES(${SWIFTEST_src} ${DRIVER_src} PROPERTIES Fortran_PREPROCESS ON)
 
 # Create a library from the source files, except the driver
-ADD_LIBRARY(swiftest ${SWIFTEST_src})
+SET(SWIFTEST_LIBRARY swiftest)
+ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
 
 # Define the executable name
 SET(SWIFTEST_DRIVER swiftest_driver)
@@ -174,11 +175,11 @@ IF(WIN32)
 ELSE()
     SET(CMAKE_INSTALL_PREFIX /usr/local)
 ENDIF(WIN32)
-INSTALL(TARGETS swiftest ${SWIFTEST_DRIVER}
-        LIBRARY DESTINATION ${LIB}
-        ARCHIVE DESTINATION ${LIB}
-        RUNTIME DESTINATION ${BIN}
-        )
+INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY}
+        RUNTIME DESTINATION bin
+        LIBRARY DESTINATION lib
+        ARCHIVE DESTINATION lib
+)
 
 
 
