### 2023-08-04 f42c2da
Added new target that allows one to export the static library file to the host after compiling it
