### 2023-08-04 cb3f621
Added new build target for exporting swiftest.mod
