+++ Dockerfile
@@ -134,6 +134,11 @@ COPY --from=build_driver /usr/local/bin/swiftest_driver /
 FROM scratch as export_library
 COPY --from=build_driver /usr/local/lib/libswiftest.a /
 
+# This build target exports the module file to the host
+FROM scratch as export_module
+COPY --from=build_driver /usr/local/include/swiftest.mod /
+
+
 # This build target creates a container with a conda environment with all dependencies needed to run the Python front end and 
 # analysis tools
 FROM continuumio/miniconda3 as python
