### 2023-08-04 b08403f
Added swiftest.mod to the install command in CMake
