+++ src/CMakeLists.txt
@@ -180,6 +180,5 @@ INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY}
         LIBRARY DESTINATION lib
         ARCHIVE DESTINATION lib
 )
-
-
+INSTALL(FILES ${MOD}/swiftest.mod DESTINATION include)
 
