### 2023-08-04 436eeb4
Updated Dockerfile with swiftest.mod
