### 2023-08-04 b7fddf4
simplified driver program to include the swiftest_driver subroutine
