### 2023-08-07 10b098c
Restructured Python and created pydriver extension module
The pydriver extension module will (eventually) allow us to run the swiftest fortran driver from a module instead of via a subprocess call to the CLI program.
