### 2023-08-07 3957b4b
Created a new bindings module for fortran and python that allows a driver program to be run directly from the Python side without using subprocess. Next step will be to fix the Simulation class so that it uses it
