+++ python/setup.py
@@ -14,16 +14,20 @@ from setuptools import setup, find_packages, Extension
 from Cython.Build import cythonize
 import os
 
-# Build the pydriver extension that allows us to run the Fortran driver as a Python module. 
-root_dir = 'pydriver'
-include_dir = os.path.join(root_dir,'include')
-lib_dir = os.path.join(root_dir,'lib')
-pydriver_extension = [Extension('swiftest.pydriver',
-                         [os.path.join(root_dir,'pydriver.pyx')],
-                         extra_compile_args=['-fPIC', '-O3'],
-                         library_dirs=[lib_dir],
-                         libraries=['swiftest','netcdff','netcdf','hdf5_hl','hdf5','m','z'],
-                         include_dirs=[include_dir],
+# Build the pybindings extension that allows us to run the Fortran driver as a Python module. 
+root_dir = 'pybindings'
+include_dirs = "/Users/daminton/git/swiftest/apple_install/usr/local/include;/Users/daminton/git/swiftest/apple_install/usr/local/include"
+include_dirs = include_dirs.split()
+include_dirs.append(root_dir)
+link_flags = "-lswiftest  /Users/daminton/git/swiftest/apple_install/usr/local/lib/libnetcdff.a /Users/daminton/git/swiftest/apple_install/usr/local/lib/libnetcdf.a -L/Users/daminton/git/swiftest/apple_install/usr/local/lib -lhdf5_hl -lhdf5 -lm -lz -lbz2 -lxml2 -lcurl"
+link_flags = link_flags.split()
+
+pybindings_extension = [Extension('swiftest.bindings',
+                         [os.path.join(root_dir,'pybindings.pyx')],
+                         extra_compile_args=['-fPIC', '-O3','-fopenmp'],
+                         extra_link_args=link_flags,
+                         libraries=['gfortran','omp'],
+                         include_dirs=include_dirs,
                          )]
 
 setup(name='swiftest',
@@ -31,5 +35,5 @@ setup(name='swiftest',
       author='David A. Minton',
       author_email='daminton@purdue.edu',
       url='https://github.itap.purdue.edu/MintonGroup/swiftest',
-      ext_modules = cythonize(pydriver_extension),
+      ext_modules = cythonize(pybindings_extension),
       packages=find_packages())
