### 2023-08-08 1da0ce1
Fixed some typos that were preventing libraries from being linked properly
