+++ cmake/Modules/SetFortranFlags.cmake
@@ -123,9 +123,9 @@ IF (NOT BUILD_SHARED_LIBS)
                 IF (USE_OPENMP)
                         SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
                                 Fortran "-qopenmp-link=static"  # Intel
-                )
+                        )
                 ENDIF (USE_OPENMP)
-       ELSEIF (COMPIELR_OPTIONS STREQUAL "GNU") 
+       ELSEIF (COMPILER_OPTIONS STREQUAL "GNU") 
                 # Set GNU static libraries
                 SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
                         Fortran  "-static-libgfortran" 
