### 2023-08-08 d955405
Continued restructuring and started adding skbuild to the python side to help automate the building of the extension module
