+++ python/CMakeLists.txt
@@ -8,3 +8,7 @@ SET(SETUP_PY_IN "${PY}/setup.py.in")
 SET(SETUP_PY_OUT "${PY}/setup.py")
 
 CONFIGURE_FILE(${SETUP_PY_IN} ${SETUP_PY_OUT})
+
+# ADD_LIBRARY(pybindings MODULE hello/_hello.cxx)
+# python_extension_module(_hello)
+# install(TARGETS _hello LIBRARY DESTINATION hello)
