### 2023-08-08 447eaaf
Another major restructuring to place the python code more central to the project. More inclusion of scikit-build stuff in the cmake files, though not complete
