### 2023-08-08 7742abb
Added a maximum order for the SpH coefficients
