### 2023-08-08 e47cd6c
Added framework for spherical harmonic gravity
