### 2023-08-08 6918164
More restructuring and progress getting the scikit-build setup working.
