+++ setup.py
@@ -11,31 +11,47 @@
 """
 
 from skbuild import setup
-from setuptools import find_packages, Extension
-from Cython.Build import cythonize
+from setuptools import find_packages #, Extension
+#from Cython.Build import cythonize
 import os
 
 # Build the pybindings extension that allows us to run the Fortran driver as a Python module. 
-root_dir = 'pybindings'
-include_dirs = ""
-include_dirs = include_dirs.split()
-include_dirs.append(root_dir)
-link_flags = ""
-link_flags = link_flags.split()
+# root_dir = os.path.join(,'bindings')
+# include_dirs = ""
+# include_dirs = include_dirs.split()
+# include_dirs.append(root_dir)
+# link_flags = ""
+# link_flags = link_flags.split()
 
-pybindings_extension = [Extension('swiftest.bindings',
-                         [os.path.join(root_dir,'pybindings.pyx')],
-                         extra_compile_args=['-fPIC', '-O3'],
-                         extra_link_args=link_flags,
-                         include_dirs=include_dirs,
-                         )]
-
+# pybindings_extension = [Extension('swiftest.bindings',
+#                          [os.path.join(root_dir,'bindings.pyx')],
+#                          extra_compile_args=['-fPIC', '-O3'],
+#                          extra_link_args=link_flags,
+#                          include_dirs=include_dirs,
+#                          package_data={"": [os.path.join(root_dir,'bindings.h')]} 
+#                          )]
+with open('version.txt') as version_file:
+    version = version_file.read().strip()
 setup(name='swiftest',
-      version='2023.8.0',
+      version=version,
       author='David A. Minton',
       author_email='daminton@purdue.edu',
       url='https://github.itap.purdue.edu/MintonGroup/swiftest',
       python_requires=">3.8",
       license="GPLv3",
-      ext_modules = cythonize(pybindings_extension),
+      #ext_modules = cythonize(pybindings_extension),
+      install_requires= [
+            'numpy>=1.24.3',
+            'pandas>=1.5.3',
+            'scipy>=1.10.1',
+            'xarray>=2022.11.0',
+            'dask>=2022.1',
+            'bottleneck>=1.3.5',
+            'h5netcdf>=1.0.2',
+            'netcdf4>=1.6.2',
+            'matplotlib>=3.7.1',
+            'astropy>=5.1',
+            'astroquery>=0.4.6',
+            'tqdm>=4.65.0',
+      ],
       packages=find_packages())
