+++ pyproject.toml
@@ -1,0 +1,8 @@
+[build-system]
+requires = [
+    "setuptools>=42",
+    "scikit-build>=0.13",
+    "cmake>=3.20.1",
+    "ninja",
+]
+build-backend = "setuptools.build_meta"
