+++ distclean.cmake
@@ -19,6 +19,8 @@ ELSE()
     SET(TOPDIR "${CMAKE_SOURCE_DIR}")
 ENDIF()
 
+SET(SKBUILD_DIR "_skbuild")
+
 MACRO(GET_PARENT_DIRECTORIES search_string return_list grandparents)
     FILE(GLOB_RECURSE new_list ${search_string})
     SET(dir_list "")
@@ -61,6 +63,7 @@ SET(DEL ${TOPDIRECTORIES}
         ${CMAKEFILES}
         ${CMAKETESTING}
         ${CMAKETESTFILES}
+        ${SKBUILD_DIR}
 )
 
 # If we are not in the build dir, delete that as well
