+++ src/bindings/CMakeLists.txt
@@ -1,5 +1,9 @@
 # Use the old method to get Python packages, as that's what scikit-build uses
 CMAKE_POLICY(SET CMP0148 OLD)
+
+SET(CYTHON_FLAGS
+    "--directive binding=True,boundscheck=False,wraparound=False,embedsignature=True,always_allow_keywords=True"
+    CACHE STRING "The directives for Cython compilation.")
 FIND_PACKAGE(PythonInterp REQUIRED)
 FIND_PACKAGE(Cython REQUIRED)
 FIND_PACKAGE(PythonExtensions REQUIRED)
@@ -14,19 +18,6 @@ IF(${PYTHON_VERSION_MAJOR} EQUAL 3
       ADD_COMPILE_OPTIONS("-Wno-deprecated-declarations")
 ENDIF()
 
-
-# Detect when building against a conda environment set the _using_conda variable
-# for use both in this file and in the parent
-GET_FILENAME_COMPONENT(_python_bin_dir ${PYTHON_EXECUTABLE} DIRECTORY)
-IF(EXISTS "${_python_bin_dir}/../conda-meta")
-  MESSAGE("-- Detected conda environment, setting INSTALL_RPATH_USE_LINK_PATH")
-  SET(_using_conda On)
-  SET(_using_conda On PARENT_SCOPE)
-ELSE()
-  SET(_using_conda Off)
-  SET(_using_conda Off PARENT_SCOPE)
-ENDIF()
-
 SET(SWIFTEST_BINDINGS swiftest_bindings)
 
 INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
@@ -34,4 +25,15 @@ ADD_CYTHON_TARGET(${SWIFTEST_BINDINGS} bindings.pyx)
 ADD_LIBRARY(${SWIFTEST_BINDINGS} MODULE ${SWIFTEST_BINDINGS})
 TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} ${SWIFTEST_LIBRARY} ${SWIFTEST_LIBRARY} ${NETCDF_FORTRAN_LIBRARIES} ${NETCDF_LIBRARIES} $ENV{LIBS})
 PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
-INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION ${PY})
+INSTALL(TARGETS ${SWIFTEST_BINDINGS} DESTINATION swiftest)
+
+IF(APPLE)
+   SET_TARGET_PROPERTIES(${SWIFTEST_BINDINGS} PROPERTIES INSTALL_RPATH "@loader_path")
+ELSE()
+   SET_TARGET_PROPERTIES(${SWIFTEST_BINDINGS} PROPERTIES INSTALL_RPATH "\$ORIGIN")
+ENDIF()
+
+IF(_using_conda OR DEFINED ENV{CIBUILDWHEEL})
+   MESSAGE(STATUS "Conda detected. Setting INSTALL_RPATH_USE_LINK_PATH to True")
+   SET_TARGET_PROPERTIES(${SWIFTEST_BINDINGS} PROPERTIES INSTALL_RPATH_USE_LINK_PATH True)
+ENDIF()
