+++ pyproject.toml
@@ -3,6 +3,7 @@ requires = [
     "setuptools>=42",
     "scikit-build>=0.13",
     "cmake>=3.20.1",
+    "cython>=3.0.0",
     "ninja",
 ]
 build-backend = "setuptools.build_meta"
