### 2023-08-09 24c9333
Rearranged and tweaked until I finally have a working scikit-build setup
