+++ src/CMakeLists.txt
@@ -136,6 +136,24 @@ ENDIF ()
 # Turn preprocessor on for all files
 SET_SOURCE_FILES_PROPERTIES(${SWIFTEST_src} ${DRIVER_src} PROPERTIES Fortran_PREPROCESS ON)
 
+
+#Set strict vs fast math flags
+STRING(TOUPPER "${CMAKE_BUILD_TYPE}" BT)
+IF(BT STREQUAL "RELEASE" OR BT STREQUAL "PROFILE")
+   SET_PROPERTY(SOURCE ${STRICT_MATH_FILES} APPEND_STRING PROPERTY COMPILE_FLAGS "${STRICTMATH_FLAGS}")
+   SET_PROPERTY(SOURCE ${FAST_MATH_FILES} APPEND_STRING PROPERTY COMPILE_FLAGS "${FASTMATH_FLAGS}")
+ENDIF()
+
+IF(BT STREQUAL "DEBUG")
+   ADD_DEFINITIONS(-DDEBUG)
+ELSEIF(BT STREQUAL "PROFILE")
+   ADD_DEFINITIONS(-DPROFILE)
+ENDIF()
+
+#####################################################
+# Add the needed libraries and special compiler flags
+#####################################################
+
 # Create a library from the source files, except the driver
 SET(SWIFTEST_LIBRARY swiftest)
 ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
@@ -143,10 +161,6 @@ ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
 # Define the executable name
 SET(SWIFTEST_DRIVER swiftest_driver)
 ADD_EXECUTABLE(${SWIFTEST_DRIVER} ${DRIVER_src})
-
-#####################################################
-# Add the needed libraries and special compiler flags
-#####################################################
 
 TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_FORTRAN_LIBRARIES} ${NETCDF_LIBRARIES} $ENV{LIBS})
 
@@ -162,24 +176,6 @@ IF(USE_COARRAY)
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY LINK_FLAGS  "${Coarray_Fortran_FLAGS} ")
 ENDIF(USE_COARRAY)
 
-STRING(REPLACE ";" " " SWIFTEST_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS} -L${LIB} ${CMAKE_LINK_LIBRARY_FLAG}${SWIFTEST_LIBRARY} ${NETCDF_FORTRAN_LIBRARIES} ${NETCDF_LIBRARIES} $ENV{LIBS}")
-SET(SWIFTEST_LINK_FLAGS "${SWIFTEST_LINK_FLAGS}" CACHE STRING "Link flags to pass to setup.py")
-
-GET_TARGET_PROPERTY(SWIFTEST_INCLUDE_DIR ${SWIFTEST_DRIVER} INCLUDE_DIRECTORIES)
-SET(SWIFTEST_INCLUDE_DIR "${SWIFTEST_INCLUDE_DIR}" CACHE STRING "Include directory needed for linking the library to an executable")
-
-#Set strict vs fast math flags
-STRING(TOUPPER "${CMAKE_BUILD_TYPE}" BT)
-IF(BT STREQUAL "RELEASE" OR BT STREQUAL "PROFILE")
[…]
