+++ swiftest/CMakeLists.txt
@@ -18,22 +18,12 @@ IF(${PYTHON_VERSION_MAJOR} EQUAL 3
       ADD_COMPILE_OPTIONS("-Wno-deprecated-declarations")
 ENDIF()
 
-SET(SWIFTEST_BINDINGS swiftest_bindings)
+SET(SWIFTEST_BINDINGS _bindings)
 
-INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR})
-ADD_CYTHON_TARGET(${SWIFTEST_BINDINGS} bindings.pyx)
+INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR} ${MOD})
+ADD_CYTHON_TARGET(${SWIFTEST_BINDINGS})
 ADD_LIBRARY(${SWIFTEST_BINDINGS} MODULE ${SWIFTEST_BINDINGS})
-TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} ${SWIFTEST_LIBRARY} ${SWIFTEST_LIBRARY} ${NETCDF_FORTRAN_LIBRARIES} ${NETCDF_LIBRARIES} $ENV{LIBS})
+#MESSAGE(STATUS "Module linked libraries\nswiftest: ${SWIFTEST_LIBRARY}\nnetcdf: ${NETCDF_LIBRARIES}\nLIBS: $ENV{LIBS}")
+TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} swiftest ${NETCDF_LIBRARIES} $ENV{LIBS})
 PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
-INSTALL(TARGETS ${SWIFTEST_BINDINGS} DESTINATION swiftest)
-
-IF(APPLE)
-   SET_TARGET_PROPERTIES(${SWIFTEST_BINDINGS} PROPERTIES INSTALL_RPATH "@loader_path")
-ELSE()
-   SET_TARGET_PROPERTIES(${SWIFTEST_BINDINGS} PROPERTIES INSTALL_RPATH "\$ORIGIN")
-ENDIF()
-
-IF(_using_conda OR DEFINED ENV{CIBUILDWHEEL})
-   MESSAGE(STATUS "Conda detected. Setting INSTALL_RPATH_USE_LINK_PATH to True")
-   SET_TARGET_PROPERTIES(${SWIFTEST_BINDINGS} PROPERTIES INSTALL_RPATH_USE_LINK_PATH True)
-ENDIF()
+INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION swiftest)
