### 2023-08-09 92540d8
Fixed issue with compiler flags when "generic" is passed as MACHINE_CODE_VALUE using Intel comilers.
