+++ cmake/Modules/SetUpNetCDF.cmake
@@ -12,3 +12,5 @@ IF(NOT NETCDF_FOUND)
     ENABLE_LANGUAGE(C) # Some libraries need a C compiler to find 
     FIND_PACKAGE(NETCDF REQUIRED)
 ENDIF(NOT NETCDF_FOUND)
+
+INCLUDE_DIRECTORIES(${NETCDF_INCLUDE_DIR})
