### 2023-08-09 9da818d
Fixed typo in swiftest_driver usage statement
