### 2023-08-09 dd2730f
Added some new info to the setup.py and streamlined/reorganize the tests
