### 2023-08-09 66c6187
Fixed some issues related to getting the NetCDF libraries linked and included correctly without relying on environment variables
