+++ src/CMakeLists.txt
@@ -24,12 +24,6 @@ ELSE ()
     MESSAGE(FATAL_ERROR "Compiler not recognized!") 
 ENDIF ()
 
-# Set some options the user may choose
-OPTION(USE_COARRAY "Use Coarray Fortran for parallelization of test particles" OFF)
-OPTION(USE_OPENMP "Use OpenMP for parallelization" ON)
-OPTION(USE_SIMD "Use SIMD vectorization" ON)
-OPTION(BUILD_SHARED_LIBS "Build using shared libraries" ON)
-
 INCLUDE(SetParallelizationLibrary)
 INCLUDE(SetUpNetCDF)
 IF (COMPILER_OPTIONS STREQUAL "Intel")
@@ -159,7 +153,16 @@ ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
 SET(SWIFTEST_DRIVER swiftest_driver)
 ADD_EXECUTABLE(${SWIFTEST_DRIVER} ${DRIVER_src})
 
-TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_FORTRAN_LIBRARIES} ${NETCDF_LIBRARIES} $ENV{LIBS})
+TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR})
+TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_INCLUDE_DIR})
+
+IF (BUILD_SHARED_LIBS)
+    TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES})
+    TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY})
+ELSE ()
+    TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES} $ENV{LIBS})
+    TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES} $ENV{LIBS})
+ENDIF (BUILD_SHARED_LIBS)
 
 IF(USE_OPENMP)
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY COMPILE_FLAGS  "${OpenMP_Fortran_FLAGS} ")
