### 2023-08-09 0091b10
Updated copyrights headers in CMakeLists.txt files and adjusted versioning for CMake
