+++ pyproject.toml
@@ -1,8 +1,8 @@
 [build-system]
 requires = [
     "setuptools>=42",
-    "scikit-build>=0.13",
-    "cmake>=3.20.1",
+    "scikit-build>=0.17",
+    "cmake>=3.5.0",
     "cython>=3.0.0",
     "ninja",
 ]
