### 2023-08-09 1f62eaf
rearranged dockerfile and updated the setting of  the machine code for Intel processors
