+++ Dockerfile
@@ -98,22 +98,15 @@ ENV NETCDF_LIBRARY=${NETCDF_HOME}
 ENV FOR_COARRAY_NUM_IMAGES=1
 ENV OMP_NUM_THREADS=1
 ENV FC="${ONEAPI_ROOT}/mpi/latest/bin/mpiifort"
-ENV CC="${ONEAPI_ROOT}/mpi/latest/bin/mpicc -cc=icx"
-ENV CXX="${ONEAPI_ROOT}/mpi/latest/bin/mpicc -cc=icpx"
+ENV CC="${ONEAPI_ROOT}/mpi/latest/bin/mpicc"
+ENV CXX="${ONEAPI_ROOT}/mpi/latest/bin/mpicxx"
 ENV FFLAGS="-fPIC -standard-semantics"
 ENV LDFLAGS="-L${INSTALL_DIR}/lib"
 ENV LIBS="-lhdf5_hl -lhdf5 -lz"
 ENV PATH /root/miniconda3/bin:$PATH
 
 COPY --from=build_deps ${INSTALL_DIR}/. ${INSTALL_DIR}/
-COPY ./cmake/ /swiftest/cmake/
-COPY ./src/ /swiftest/src/
-COPY ./CMakeLists.txt /swiftest/
-COPY ./swiftest/ /swiftest/swiftest/
-COPY ./version.txt /swiftest/
-COPY ./setup.py /swiftest/
-COPY ./requirements.txt /swiftest/
-COPY ./pyproject.toml /swiftest/
+
 COPY ./environment.yml /swiftest/
 
 RUN wget https://repo.anaconda.com/miniconda/Miniconda3-py311_23.5.2-0-Linux-x86_64.sh  && \
@@ -122,6 +115,7 @@ RUN wget https://repo.anaconda.com/miniconda/Miniconda3-py311_23.5.2-0-Linux-x86
   source /root/.bashrc && conda update --all -y && \
   conda install conda-libmamba-solver -y && \
   conda config --set solver libmamba && \
+  cd swiftest && \
   conda env create -f environment.yml && \
   conda init bash && \
   echo "conda activate swiftest-env" >> ~/.bashrc  && \
@@ -129,18 +123,16 @@ RUN wget https://repo.anaconda.com/miniconda/Miniconda3-py311_23.5.2-0-Linux-x86
   conda install -c conda-forge scikit-build -y&& \ 
   conda install -c anaconda cython -y
 
-# RUN cd swiftest && \
-#   cmake -S . -B build -DCMAKE_INSTALL_PREFIX="${INSTALL_DIR}" \
-#   -DMACHINE_CODE_VALUE=${MACHINE_CODE_VALUE} \
-#   -DCMAKE_BUILD_TYPE=${BUILD_TYPE} \
-#   -DUSE_COARRAY=OFF \
-#   -DBUILD_SHARED_LIBS=OFF \
-#   ${EXTRA_CMAKE_OPTIONS} && \
-#   cmake --build build && \
-#   cmake --install build 
-
-# RUN cd swiftest && \
-#    pip install . -v
+COPY ./cmake/ /swiftest/cmake/
+COPY ./src/ /swiftest/src/
+COPY ./CMakeLists.txt /swiftest/
+COPY ./swiftest/ /swiftest/swiftest/
+COPY ./version.txt /swiftest/
+COPY ./setup.py /swiftest/
+COPY ./requirements.txt /swiftest/
+COPY ./pyproject.toml /swiftest/
+RUN cd swiftest && \
+   pip install . -v
 
 # This build target creates a container that executes just the driver program
 FROM ubuntu:20.04 as driver
