### 2023-08-09 adbb88d
Let the Dockerfile be able to pass CMake arguments to setup.py and removed an unneeded required package
