### 2023-08-11 c052e4c
Began process to develop a more robust set of build scripts for all of my test platforms
