### 2023-08-11 e438bdd
Continuted improvements to automated build scripts
