### 2023-08-11 ee38471
Cleaned up and completed python SpH functions
