### 2023-08-11 4602c19
Fixes to build scripts so that they can work with Docker or just conda/mamba
