### 2023-08-11 a5d044f
Cleaned cruft from Dockerfile
