+++ Dockerfile
@@ -28,159 +28,4 @@ RUN if [ "$BUILDIMAGE" = "intel/oneapi-hpckit:2023.1.0-devel-ubuntu20.04" ]; the
         echo "conda activate swiftest-build-env" >> ~/.bashrc  && \
         /bin/bash -lic "${SCRIPT_DIR}/build_dependencies.sh GNU"; \
     fi 
-
-# ENV INSTALL_DIR="/usr/local"
-# ENV FC="${ONEAPI_ROOT}/mpi/latest/bin/mpiifort"
-# ENV CC="${ONEAPI_ROOT}/mpi/latest/bin/mpicc -cc=icx"
-# ENV CXX="${ONEAPI_ROOT}/mpi/latest/bin/mpicc -cc=icpx"
-# ENV F77="${FC}"
-# ENV CFLAGS="-fPIC"
-
-# # Get the HDF5, NetCDF-C, and NetCDF-Fortran libraries
-# RUN wget -qO- https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.14/hdf5-1.14.1/src/hdf5-1.14.1-2.tar.gz | tar xvz && \
-#     wget -qO- https://github.com/Unidata/netcdf-c/archive/refs/tags/v4.9.2.tar.gz | tar xvz && \
-#     wget -qO- https://github.com/Unidata/netcdf-fortran/archive/refs/tags/v4.6.1.tar.gz | tar xvz && \
-#     wget -qO- https://www.zlib.net/zlib-1.2.13.tar.gz | tar xvz
-
-# RUN apt-get update && \
-#   DEBIAN_FRONTEND=noninteractive apt-get install -y --no-install-recommends \
-#   m4 && \
-#   rm -rf /var/lib/apt/lists/* 
-
-# RUN cd zlib-1.2.13 && \
-#   ./configure --prefix=${INSTALL_DIR} --static && \
-#   make && \
-#   make install
-
-# RUN cd hdf5-1.14.1-2 && \
-#   ./configure --disable-shared \
-#               --enable-build-mode=production \
-#               --disable-fortran \
-#               --disable-java \
-#               --disable-cxx \
-#               --prefix=${INSTALL_DIR} \
-#               --with-zlib=${INSTALL_DIR} && \
-#               make && \
-#               make install
-
-# RUN cd netcdf-c-4.9.2 && \
-#   ./configure --disable-shared \
-#               --disable-dap \
-#               --disable-libxml2 \
-#               --disable-byterange \
-#               --prefix=${INSTALL_DIR} && \
-#               make && \
-#               make install
-
-# ENV NCDIR="${INSTALL_DIR}"
-# ENV NFDIR="${INSTALL_DIR}"
-# ENV HDF5_ROOT="${INSTALL_DIR}"
-# ENV HDF5_LIBDIR="${HDF5_ROOT}/lib"
-# ENV HDF5_INCLUDE_DIR="${HDF5_ROOT}/include"
-# ENV HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
-
-# # NetCDF-Fortran library
-# ENV FCFLAGS="${CFLAGS} -standard-semantics"
-# ENV FFLAGS=${CFLAGS}
-# ENV CPPFLAGS="-I${INSTALL_DIR}/include"
-# ENV LIBS="-L/usr/local/lib -L/usr/lib/x86_64-linux-gnu -lnetcdf -lhdf5_hl -lhdf5 -lm -lz" 
-# RUN cd netcdf-fortran-4.6.1 && \
-#   ./configure --disable-shared --prefix=${NFDIR} && \
-#   make && \
-#   make install  
-
[…]
