### 2023-08-11 82c34d6
Added file copy lines to Dockerfile
