### 2023-08-11 b73911d
More updates to Dockerfile to get it closer to being able to do a complete build
