### 2023-08-11 aebe98a
Finished up the acceleration calculation from C_lm. Final lookover/test needed
