+++ src/swiftest/swiftest_sph.f90
@@ -17,7 +17,7 @@ use SHTOOLS
 
 contains
 
-    module subroutine swiftest_sph_g_acc_one(gm, r_0, phi, theta, rh, c_lm, g_sph)
+    module subroutine swiftest_sph_g_acc_one(GMcb, r_0, phi, theta, rh, c_lm, g_sph, GMpl, aoblcb)
         !! author: Kaustub P. Anand
         !!
         !! Calculate the acceleration terms for one pair of bodies given c_lm, theta, phi, r
@@ -25,33 +25,34 @@ contains
 
         implicit none
         ! Arguments
-        real(DP), intent(in)        :: gm                       !! GMass of the central body
-        real(DP), intent(in)        :: r_0                      !! radius of the central body
-        real(DP), intent(in)        :: phi                      !! Azimuthal/Phase angle (radians)
-        real(DP), intent(in)        :: theta                    !! Inclination/Zenith angle (radians)
-        real(DP), intent(in), dimension(NDIM)       :: rh       !! distance vector of body
-        real(DP), intent(in), dimension(2, :, :)    :: c_lm     !! Spherical Harmonic coefficients
-        real(DP), intent(out), dimension(NDIM)      :: g_sph    !! acceleration vector
-
+        real(DP), intent(in)        :: GMcb                        !! GMass of the central body
+        real(DP), intent(in)        :: r_0                         !! radius of the central body
+        real(DP), intent(in)        :: phi                         !! Azimuthal/Phase angle (radians)
+        real(DP), intent(in)        :: theta                       !! Inclination/Zenith angle (radians)
+        real(DP), intent(in), dimension(NDIM)       :: rh          !! distance vector of body
+        real(DP), intent(in), dimension(2, :, :)    :: c_lm        !! Spherical Harmonic coefficients
+        real(DP), intent(out), dimension(NDIM)      :: g_sph       !! acceleration vector
+        real(DP), dimension(:),   intent(in),  optional :: GMpl    !! Masses of input bodies if they are not test particles
+        real(DP), dimension(:),   intent(inout), optional :: aoblcb  !! Barycentric acceleration of central body (only needed if input bodies are massive)
+     
         ! Internals
         integer        :: l, m              !! SPH coefficients
         real(DP)       :: r_mag             !! magnitude of rh
-        real(DP)       :: p                 !! Associated Lengendre Polynomials at a given cos(theta)
+        real(DP), dimension(2, :, :)  :: p, dp             !! Associated Lengendre Polynomials at a given cos(theta)
         integer        :: l_max             !! max Spherical Harmonic l order value
+        real(DP)       :: plm, dplm         !! Associated Legendre polynomials at a given l, m
 
+        g_sph(:) = 0.0_DP
         r_mag = sqrt(dot_product(rh(:), rh(:)))
-        l_max = size(c_lm, 2)
-        PlBar(p, lmax, cos(theta))      ! Associated Legendre Polynomials
+        l_max = size(c_lm, 2) - 1
+        PlmON_d1(p, dp, l_max, cos(theta))      ! Orthonormalized Associated Legendre Polynomials and the 1st Derivative
 
-        do l = 1, l_max
-            do m = 1, l
+        do l = 0, l_max
+            do m = 0, l
 
                 ! Associated Legendre Polynomials   
                 plm = p(PlmIndex(l, m))         ! p_l,m
[…]
