### 2023-08-11 9af17e5
More tweaks to build scripts
