### 2023-08-11 2124551
Fixed issues with build scripts and the build directory
