+++ buildscripts/fetch_dependencies.sh
@@ -11,8 +11,8 @@
 # You should have received a copy of the GNU General Public License along with Swiftest. 
 # If not, see: https://www.gnu.org/licenses. 
 SCRIPT_DIR=$(realpath $(dirname $0))
-BUILD_DIR=$(realpath ${SCRIPT_DIR}/../build)
-
+ROOT_DIR=$(realpath ${SCRIPT_DIR}/..)
+BUILD_DIR="${ROOT_DIR}/build"
 mkdir -p ${BUILD_DIR}
 cd $BUILD_DIR
 
@@ -20,3 +20,5 @@ wget -qO- https://www.zlib.net/zlib-1.2.13.tar.gz | tar xvz
 wget -qO- https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.14/hdf5-1.14.1/src/hdf5-1.14.1-2.tar.gz | tar xvz 
 wget -qO- https://github.com/Unidata/netcdf-c/archive/refs/tags/v4.9.2.tar.gz | tar xvz 
 wget -qO- https://github.com/Unidata/netcdf-fortran/archive/refs/tags/v4.6.1.tar.gz | tar xvz 
+
+echo "Files downloaded to ${BUILD_DIR}"
