### 2023-08-11 47faaa0
Adjustments to build script for Mac
