### 2023-08-12 20abadb
more tweaks to build script for Mac
