### 2023-08-12 d5168fe
Separated out the Intel and GNU versions of the Dockerfile into separate files for clarity
