### 2023-08-12 fbcb4da
Adjusted Dockerfile names to keep them consistent with build script variables
