### 2023-08-14 8f61a97
Got a working Dockerfile for GNU
