### 2023-08-15 49f81d3
More fixes to build scripts. Pushing for testing
