+++ environment.yml
@@ -20,3 +20,4 @@ dependencies:
   - tqdm>=4.65.0
   - x264>=1!157.20191217
   - ffmpeg>=4.3.2
+  - py>=1.11

[…]
