### 2023-08-15 17b0339
Fixed typos in Intel Dockerfile and build_dependencies script
