### 2023-08-15 6d4167e
Updated the build_swiftest script to be able to successfully generate a wheel file on the M1 Mac
