### 2023-08-15 e72484a
More tweaks to build scripts for robustness
