+++ src/CMakeLists.txt
@@ -151,8 +151,9 @@ ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
 SET(SWIFTEST_DRIVER swiftest_driver)
 ADD_EXECUTABLE(${SWIFTEST_DRIVER} ${DRIVER_src})
 
-TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR})
+MESSAGE(STATUS "NetCDF Include directory: ${NETCDF_INCLUDE_DIR}")
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_INCLUDE_DIR})
+TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR})
 
 IF (BUILD_SHARED_LIBS)
     TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES})
@@ -173,7 +174,6 @@ IF(USE_COARRAY)
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY COMPILE_FLAGS  "${Coarray_Fortran_FLAGS} ")
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY LINK_FLAGS  "${Coarray_Fortran_FLAGS} ")
 ENDIF(USE_COARRAY)
-
 
 # Check to see if the compiler allows for local-spec in do concurrent statements. Set a preprocessor variable if it does
 IF (USE_OPENMP)
