### 2023-08-15 f46be52
Split the dependency build up so that each one can be build separately. More adjustments on the road to getting a stable build
