+++ setup.py
@@ -12,12 +12,6 @@
 
 from skbuild import setup
 
-cmake_args = [
-            '-DCMAKE_BUILD_TYPE=RELEASE',
-            '-DUSE_COARRAY:BOOL=OFF',
-            '-DUSE_OPENMP:BOOL=ON',
-            ]
-
 with open('version.txt') as version_file:
     version = version_file.read().strip()
     
@@ -47,7 +41,6 @@ setup(name='swiftest',
         'Programming Language :: Python :: 3',
       ],
       keywords='astronomy astrophysics planetary nbody integrator symplectic wisdom-holman',
-      cmake_args=cmake_args,
       install_requires= [
             'numpy>=1.24.3',
             'scipy>=1.10.1',
