+++ buildscripts/build_zlib.sh
@@ -21,6 +21,7 @@ USTMT="Usage: ${0} <-c Intel|GNU-Linux|GNU-Mac> [-p {/usr/local}|/prefix/path]"
 IFORT=false
 PREFIX=/usr/local
 COMPILER=""
+CARG=""
 while getopts ":c:p:" ARG; do
     case "${ARG}" in
     c)
@@ -46,20 +47,34 @@ export FC=${FC}
 export F77=${F77}
 export CPP=${CPP}
 
-printf "*********************************************************\n"
-printf "*          STARTING DEPENDENCY BUILD                    *\n"
-printf "*********************************************************\n"
-printf "Using ${COMPILER} compilers:\nFC: ${FC}\nCC: ${CC}\nCXX: ${CXX}\n"
-printf "Installing to ${PREFIX}\n"
-printf "\n"
+export LD_LIBRARY_PATH="${PREFIX}/lib:${LD_LIBRARY_PATH}"
+export CPPFLAGS="${CPPFLAGS} -isystem {$PREFIX}/include"
+export LDFLAGS="${LDFLAGS} -L${PREFIX}/lib"
+export CPATH="${PREFIX}/include:${CPATH}"
+export CFLAGS="${CFLAGS} -Wno-unused-but-set-variable"
 
-${SCRIPT_DIR}/build_zlib.sh -c $COMPILER -p $PREFIX
-${SCRIPT_DIR}/build_hdf5.sh -c $COMPILER -p $PREFIX
-${SCRIPT_DIR}/build_netcdf-c.sh -c $COMPILER -p $PREFIX
-${SCRIPT_DIR}/build_netcdf-fortran.sh -c $COMPILER -p $PREFIX
+if [ $COMPILER = "GNU-Mac" ]; then
+    export MACOSX_DEPLOYMENT_TARGET=13 
+    export LDFLAGS="${LDFLAGS} -Wl,-no_compact_unwind"
+    export CFLAGS="${CFLAGS} -Wno-deprecated-non-prototype"
+fi
 
-printf "\n"
 printf "*********************************************************\n"
-printf "*             DEPENDENCIES ARE BUILT                    *\n"
+printf "*            BUILDING ZLIB STATIC LIBRARY               *\n"
 printf "*********************************************************\n"
-printf "Dependencys are installed to: ${PREFIX}\n\n"
+printf "LIBS: ${LIBS}\n"
+printf "CFLAGS: ${CFLAGS}\n"
+printf "CPPFLAGS: ${CPPFLAGS}\n"
+printf "CPATH: ${CPATH}\n"
+printf "LD_LIBRARY_PATH: ${LD_LIBRARY_PATH}\n"
+printf "LDFLAGS: ${LDFLAGS}\n"
+printf "*********************************************************\n"
+
+cd ${BUILD_DIR}/zlib-*
+./configure --prefix=${PREFIX} --static 
+make 
+make install
+if [ $? -ne 0 ]; then
+   printf "zlib could not be compiled.\n"
+   exit 1
+fi
