+++ src/CMakeLists.txt
@@ -163,10 +163,10 @@ ELSE ()
     TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES} $ENV{LIBS})
 ENDIF (BUILD_SHARED_LIBS)
 
-IF(USE_OPENMP)
+IF(USE_OPENMP OR USE_SIMD)
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY COMPILE_FLAGS  "${OpenMP_Fortran_FLAGS} ")
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY LINK_FLAGS  "${OpenMP_Fortran_FLAGS} ")
-ENDIF(USE_OPENMP)
+ENDIF()
 
 IF(USE_COARRAY)
     TARGET_COMPILE_DEFINITIONS(${SWIFTEST_LIBRARY} PUBLIC -DCOARRAY)
