### 2023-08-15 c64eefd
Removed old cruft from when the executable was called
