### 2023-08-15 4102dd5
Fixed typo in build script for building the distribution package
