+++ buildscripts/build_swiftest.sh
@@ -107,6 +107,6 @@ printf "SKBUILD_CONFIGURE_OPTIONS: ${SKBUILD_CONFIGURE_OPTIONS}\n"
 printf "*********************************************************\n"
 
 python3 -m pip install build pip
-python3 -m pip build
+python3 -m build
 python3 -m pip install . -v 
 
