### 2023-08-15 bdb51ba
Trying new capture method and fixed a bug in the output descriptor
