### 2023-08-15 1d5161a
Restructured the driver method so that running the simulation acts sort of like it used to
