+++ swiftest/_bindings.pyx
@@ -1,7 +1,7 @@
 # cython: language_level=3, c_string_type=unicode, c_string_encoding=ascii
 
 cdef extern from "_bindings.h":
-    void bindings_c_driver(char* integrator, char* param_file_name, char* display_style)
+    void bindings_c_driver(char* integrator, char* param_file_name, char* display_style) noexcept nogil
 
 def driver(integrator, param_file_name, display_style):
     b_integrator = bytes(integrator,'ascii') + b'\x00'
@@ -13,5 +13,6 @@ def driver(integrator, param_file_name, display_style):
         char* c_param_file_name = b_param_file_name 
         char* c_display_style = b_display_style 
 
-    bindings_c_driver(c_integrator, c_param_file_name, c_display_style)
+    with nogil:
+        bindings_c_driver(c_integrator, c_param_file_name, c_display_style)
     return
