+++ src/swiftest/swiftest_driver.f90
@@ -182,15 +182,6 @@ contains
          call nbody_system%dump(param, system_history)
          call nbody_system%display_run_information(param, integration_timer, phase="last")
       end associate
-
-#ifdef COARRAY
-      if (this_image() == 1) then
-#endif
-         call base_util_exit(SUCCESS,unit=param%display_unit)
-#ifdef COARRAY
-      end if ! (this_image() == 1) 
-#endif
-
    return
    end subroutine swiftest_driver
 
