### 2023-08-15 c6bdcd8
Removed all of the output stream parsing and now just output the Fortran progress bar version of the runs when called inside Python.
