+++ swiftest/simulation_class.py
@@ -23,12 +23,10 @@ from pathlib import Path
 import datetime
 import xarray as xr
 import numpy as np
-from functools import partial
 import numpy.typing as npt
 import shutil
 import warnings
-import sys
-from tqdm.auto import tqdm
+import contextlib
 from typing import (
     Literal,
     Dict,
@@ -36,7 +34,7 @@ from typing import (
     Tuple,
     Any
 )
-import contextlib
+from cython import nogil
 
 @contextlib.contextmanager
 def _cwd(newdir):
@@ -46,94 +44,6 @@ def _cwd(newdir):
         yield
     finally:
         os.chdir(olddir)
-        
-        
-import os
-import sys
-import threading
-import time
-
-
-class OutputGrabber(object):
-    """
-    Class used to grab standard output or another stream.
-    """
-    escape_char = "\b"
-
-    def __init__(self, stream=None, threaded=True):
-        self.origstream = stream
-        self.threaded = threaded
-        if self.origstream is None:
-            self.origstream = sys.stdout
-        self.origstreamfd = self.origstream.fileno()
-        self.capturedtext = ""
-        # Create a pipe so the stream can be captured:
-        self.pipe_out, self.pipe_in = os.pipe()
-
-    def __enter__(self):
-        self.start()
-        return self
-
-    def __exit__(self, type, value, traceback):
-        self.stop()
-
-    def start(self):
-        """
-        Start capturing the stream data.
-        """
-        self.capturedtext = ""
-        # Save a copy of the stream:
-        self.streamfd = os.dup(self.origstreamfd)
-        # Replace the original stream with our write pipe:
-        os.dup2(self.pipe_in, self.origstreamfd)
-        if self.threaded:
-            # Start thread that will read the stream:
-            self.workerThread = threading.Thread(target=self.readOutput)
-            self.workerThread.start()
-            # Make sure that the thread is running and os.read() has executed:
-            time.sleep(0.01)
-
-    def stop(self):
-        """
-        Stop capturing the stream data and save the text in `capturedtext`.
-        """
-        # Print the escape character to make the readOutput method stop:
-        self.origstream.write(self.escape_char)
-        # Flush the stream to make sure all our data goes in before
-        # the escape character:
-        self.origstream.flush()
-        if self.threaded:
-            # wait until the thread finishes so we are sure that
-            # we have until the last character:
-            self.workerThread.join()
-        else:
-            self.readOutput()
-        # Close the pipe:
-        os.close(self.pipe_in)
-        os.close(self.pipe_out)
-        # Restore the original stream:
-        os.dup2(self.streamfd, self.origstreamfd)
-        # Close the duplicate stream:
-        os.close(self.streamfd)
-
-    def readOutput(self):
-        """
-        Read the stream data (one byte at a time)
-        and save the text in `capturedtext`.
-        """
-        
-        def _outstream(self):
-            capturedtext = ''
-            while True:
-                char = os.read(self.pipe_out,1).decode(self.origstream.encoding)
-                if not char or self.newline in char:
-                    break
-                capturedtext += char
-            yield capturedtext
-        
-        while True:
-            self.capturedtext = _outstream(self)
-        
 
 class Simulation(object):
     """
@@ -406,27 +316,7 @@ class Simulation(object):
             Parameter input file equivalent: None
         """
 
-        # Configuration parameters will be stored in a json file alongside the Python source scripts.
-        self._config_file = Path(_pyfile).parent / "swiftest_configuration.json"
-        config_exists = self._config_file.exists()
-        if config_exists:
-            try:
-                with open(self._config_file, 'r') as f:
[…]
