### 2023-08-15 15a1b56
Fixed typos in build scripts
