+++ buildscripts/build_netcdf-fortran.sh
@@ -55,7 +55,7 @@ export HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
 export NCDIR="${PREFIX}"
 export NFDIR="${PREFIX}"
 export LD_LIBRARY_PATH="${PREFIX}/lib:${LD_LIBRARY_PATH}"
-export CPPFLAGS="${CPPFLAGS} -isystem {$PREFIX}/include"
+export CPPFLAGS="${CPPFLAGS} -isystem ${PREFIX}/include"
 export LDFLAGS="${LDFLAGS} -L${PREFIX}/lib"
 export CPATH="${PREFIX}/include:${CPATH}"
 export CFLAGS="${CFLAGS} -Wno-unused-but-set-variable"
