### 2023-08-16 9243080
Updated scritps to work with cibuildwheel
