### 2023-08-16 ea4ae19
Removed large example from official repo to cut down on size.
