+++ buildscripts/build_swiftest.sh
@@ -106,4 +106,6 @@ printf "NETCDF_FORTRAN_HOME: ${NETCDF_FORTRAN_HOME}\n"
 printf "SKBUILD_CONFIGURE_OPTIONS: ${SKBUILD_CONFIGURE_OPTIONS}\n"
 printf "*********************************************************\n"
 
-pipx run cibuildwheel --platform macos
+#pipx run cibuildwheel --platform macos
+python3 -m pip install build pip
+python3 -m build
