+++ examples/swiftest_performance/symba-performance-comparison_JOSS/swift/pl32k/param.in
@@ -1,34 +1,0 @@
-!
-!Parameter file for the Entire Solar System for the Terrestrial Disk
-!
-NPLMAX          32002                            !turns this off
-NTPMAX          0                               !turns this off
-T0              0.0e0                           !starting time in years
-TSTOP           0.030                     !simulation length in years
-DT              0.001                           !step size in years
-PL_IN           Jul27_32k_fully.in !input file in XV not EL
-TP_IN           tp.in
-IN_TYPE         ASCII                           !format of input file
-ISTEP_OUT       1  
-ISTEP_DUMP      1  
-BIN_OUT         bin.dat                         !output file
-OUT_TYPE        REAL8                           !double precision real output
-OUT_FORM        XV                              !output type in XV OR EL
-OUT_STAT        UNKNOWN
-J2              0.0                             !higher order harmonics
-J4              0.0                             !higher order harmonics
-CHK_CLOSE       yes                             !check for planetary close encounters
-CHK_RMIN        0.005                           !check for close solar encounters in AU
-CHK_RMAX        10000.0                         !drop things from the simulation in AU
-CHK_EJECT       -1.0                            !turns this off
-CHK_QMIN        -1.0                            !turns this off
-!CHK_QMIN_COORD HELIO                           !turns this off
-!CHK_QMIN_RANGE 1.0 1000.0                      !turns this off
-ENC_OUT         enc.dat                         !output file for close encounters
-EXTRA_FORCE     no                              !this would be user defined forces
-BIG_DISCARD     no                              !output all planets if anything is discarded
-RHILL_PRESENT   yes 
-WALLTIME        yes
-ENERGY          no
-
-
