### 2023-08-16 0bae0cd
Made an adjustment to the FindPython3 call in the CMake based on a failure of the Linux build in cibuildwheels
