+++ CMakeLists.txt
@@ -29,7 +29,7 @@ IF(PROJECT_SOURCE_DIR STREQUAL PROJECT_BINARY_DIR)
 ENDIF()
 
 # Ensure scikit-build modules
-FIND_PACKAGE(Python3 COMPONENTS Interpreter Development REQUIRED)
+FIND_PACKAGE(Python3 COMPONENTS Interpreter Development.Module REQUIRED)
 IF (NOT SKBUILD)
   EXECUTE_PROCESS(
     COMMAND "${Python3_EXECUTABLE}"
