### 2023-08-16 428f804
Fixed typo in the set_compiler script that was accidentaly setting the c++ compiler to cpp
