+++ pyproject.toml
@@ -9,10 +9,8 @@ requires = [
 build-backend = "setuptools.build_meta"
 
 [tool.cibuildwheel.linux]
-environment = {SCRIPT_DIR="buildscripts", BUILD_DIR="build", PREFIX="/usr/local", SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF", CFLAGS="-fPIC", CPPFLAGS="-isystem /usr/local/include", NETCDF_FORTRAN_HOME="/usr/local", LIBS="-lm -lz -ldl"}
+environment = {SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF",CPATH="/usr:/usr/local",LD_LIBRARY_PATH="/usr/lib:/usr/lib64:/usr/local/lib:/usr/lib/x86_64-linux-gnu:/usr/lib/aarch64-linux-gnu"}
 before-all = [
-    "yum install wget m4 doxygen -y",
-    "bash ${SCRIPT_DIR}/fetch_dependencies.sh -d ${BUILD_DIR}",
-    "bash ${SCRIPT_DIR}/build_dependencies.sh -c GNU-Linux -p ${PREFIX}"
+    "yum install netcdf-fortran-static netcdf-static hdf5-static -y",
 ]
     
