### 2023-08-16 ccc9cd4
More improvements to FindNETCDF toward getting a working cibuildwheel setup
