+++ cmake/Modules/FindNETCDF.cmake
@@ -9,14 +9,54 @@
 
 # - Finds the NetCDF libraries 
 
+FIND_PATH(NFBIN
+NAMES nf-config
+HINTS 
+   ENV NETCDF_FORTRAN_HOME
+   ENV PATH
+PATH_SUFFIXES
+   bin
+)
+
+IF (NFBIN)
+   SET(CMD "${NFBIN}/nf-config")
+   LIST(APPEND CMD "--includedir")
+   MESSAGE(STATUS "Searching for NetCDF-Fortran include directory using ${CMD}")
+   EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFINCLUDE_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
+   IF (NFINCLUDE_DIR)
+      MESSAGE(STATUS "Found in ${NFINCLUDE_DIR}")
+   ELSE ()
+      MESSAGE(STATUS "Cannot execute ${CMD}")
+      MESSAGE(STATUS "OUTPUT: ${NFINCLUDE_DIR}")
+      MESSAGE(STATUS "RESULT: ${RES}")
+      MESSAGE(STATUS "ERROR : ${ERR}")
+   ENDIF ()
+
+   SET(CMD "${NFBIN}/nf-config")
+   LIST(APPEND CMD "--prefix")
+   MESSAGE(STATUS "Searching for NetCDF-Fortran library directory using ${CMD}")
+   EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFPREFIX_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
+   IF (NFPREFIX_DIR)
+      MESSAGE(STATUS "Found in ${NFPREFIX_DIR}")
+   ELSE ()
+      MESSAGE(STATUS "Cannot execute ${CMD}")
+      MESSAGE(STATUS "OUTPUT: ${NFPREFIX_DIR}")
+      MESSAGE(STATUS "RESULT: ${RES}")
+      MESSAGE(STATUS "ERROR : ${ERR}")
+   ENDIF ()
+ENDIF()
+
 FIND_PATH(NETCDF_INCLUDE_DIR 
    NAMES netcdf.mod 
    HINTS 
+      ${NFINCLUDE_DIR}
       ENV NETCDF_INCLUDE_DIR 
       ENV NETCDF_FORTRAN_HOME
       ENV CPATH
    PATH_SUFFIXES
       include
+      modules
+      mod
    REQUIRED
 )
 
@@ -32,6 +72,7 @@ ENDIF()
 FIND_LIBRARY(NETCDF_FORTRAN_LIBRARY 
    NAMES ${NETCDFF} 
    HINTS
+      ${NFPREFIX_DIR}
       ENV NETCDF_FORTRAN_HOME
       ENV NETCDF_HOME
       ENV LD_LIBRARY_PATH
