### 2023-08-16 2d6cf87
More tweaks. Turns out the cibuildwheel netcdf libraries are super old and I can't get them to link
