### 2023-08-16 7198e0d
Added missing wget package for fetching dependencies
