### 2023-08-16 abb66f5
Reducing the dependencies and ensuring PIC for ARM
