+++ src/CMakeLists.txt
@@ -150,6 +150,9 @@ ADD_EXECUTABLE(${SWIFTEST_DRIVER} ${DRIVER_src})
 # Create a library from the source files, except the driver
 SET(SWIFTEST_LIBRARY swiftest)
 ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
+IF (NOT BUILD_SHARED_LIBS) 
+    SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} PROPERTY POSITION_INDEPENDENT_CODE)
+ENDIF ()
 
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_INCLUDE_DIR})
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR})
