+++ swiftest/CMakeLists.txt
@@ -30,6 +30,9 @@ SET(SWIFTEST_BINDINGS _bindings)
 
 ADD_CYTHON_TARGET(${SWIFTEST_BINDINGS})
 ADD_LIBRARY(${SWIFTEST_BINDINGS} MODULE ${SWIFTEST_BINDINGS})
+IF (NOT BUILD_SHARED_LIBS) 
+    SET_PROPERTY(TARGET ${SWIFTEST_BINDINGS} PROPERTY POSITION_INDEPENDENT_CODE)
+ENDIF ()
 TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} swiftest ${NETCDF_LIBRARIES} $ENV{LIBS})
 PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_BINDINGS} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${MOD} ${NETCDF_INCLUDE_DIR})
