### 2023-08-16 9b24467
Put back hdf5 library.
