### 2023-08-17 6ddec8a
Added OpenMP flag to binding compile step
