+++ cmake/Modules/FindNETCDF.cmake
@@ -118,7 +118,7 @@ ELSE ()
       IF (EXTRA_FLAGS)
          SEPARATE_ARGUMENTS(EXTRA_FLAGS NATIVE_COMMAND "${EXTRA_FLAGS}")
          LIST(REMOVE_DUPLICATES EXTRA_FLAGS)
-         LIST(FILTER EXTRA_FLAGS EXCLUDE REGEX "netcdf+|-L+")
+         LIST(FILTER EXTRA_FLAGS EXCLUDE REGEX "netcdf+")
          MESSAGE(STATUS "Extra library flags: ${EXTRA_FLAGS}")
       ELSE ()
          MESSAGE(STATUS "Cannot execute ${CMD}")
