### 2023-08-17 09e107c
Streamlined the dependency building scripts
