### 2023-08-17 9eb9c5d
Streamlined the dependency building scripts
