+++ buildscripts/build_all.sh
@@ -66,7 +66,13 @@ case $OS in
         ;;
 esac
 
-${SCRIPT_DIR}/fetch_dependencies.sh -d ${BUILD_DIR} && \
+mkdir -p ${BUILD_DIR}
+cd $BUILD_DIR
+wget -qO- https://www.zlib.net/zlib-1.2.13.tar.gz | tar xvz
+wget -qO- https://support.hdfgroup.org/ftp/HDF5/releases/hdf5-1.14/hdf5-1.14.1/src/hdf5-1.14.1-2.tar.gz | tar xvz 
+wget -qO- https://github.com/Unidata/netcdf-c/archive/refs/tags/v4.9.2.tar.gz | tar xvz 
+wget -qO- https://github.com/Unidata/netcdf-fortran/archive/refs/tags/v4.6.1.tar.gz | tar xvz 
+
 ${SCRIPT_DIR}/build_dependencies.sh -c $COMPILER -p ${PREFIX} && \
 ${SCRIPT_DIR}/build_swiftest.sh -c $COMPILER -p ${PREFIX}
 
