+++ pyproject.toml
@@ -12,7 +12,7 @@ build-backend = "setuptools.build_meta"
 environment = {SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF", FFLAGS="${FFLAGS} -fPIC"}
 before-all = [
     "yum install wget m4 doxygen -y",
-    "buildscripts/fetch_dependencies.sh -d build",
-    "buildscripts/build_dependencies.sh -c GNU-Linux -p /usr/local"
+    "buildscripts/build_dependencies.sh",
+    "buildscripts/build_swiftest.sh -c GNU-Linux -p /usr/local"
 ]
     

[…]
