### 2023-08-17 170360d
Added extra flags needed to get a successful build and link up the openmp library
