+++ cmake/Modules/SetFortranFlags.cmake
@@ -141,6 +141,11 @@ IF (NOT BUILD_SHARED_LIBS)
                 SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
                         Fortran  "-static-libquadmath" 
                 )
+                IF (USE_OPENMP)
+                        SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
+                                Fortran "-lgomp"  
+                        )
+                ENDIF (USE_OPENMP)
         ENDIF ()
 
 ENDIF (NOT BUILD_SHARED_LIBS)
