### 2023-08-17 efda5c2
Restructured scripts to consolidate variable definitions and argument parsing into one place.
