+++ buildscripts/build_zlib.sh
@@ -11,52 +11,8 @@
 # You should have received a copy of the GNU General Public License along with Swiftest. 
 # If not, see: https://www.gnu.org/licenses. 
 SCRIPT_DIR=$(realpath $(dirname $0))
-
-# Parse arguments
-USTMT="Usage: ${0} <-c Intel|GNU-Linux|GNU-Mac> <-d /path/to/dependency/source> [-p {/usr/local}|/prefix/path] "
-PREFIX=/usr/local
-COMPILER=""
-DEPENCENCY_DIR=""
-CARG=""
-while getopts ":c:p:d:" ARG; do
-    case "${ARG}" in
-    c)
-        COMPILER="${OPTARG}"
-        ;;
-    p)
-        PREFIX="${OPTARG}"
-        ;;
-    d)
-        DEPENDENCY_DIR="${OPTARG}"
-        ;;
-    :)      
-        echo "Error: -${OPTARG} requires an argument."
-        echo $USTMT
-        exit 1
-        ;;
-    *)
-        ;;
-    esac
-done
-CMD="${SCRIPT_DIR}/set_compilers.sh -c $COMPILER"
-read -r CC CXX FC F77 CPP < <($CMD)
-export CC=${CC}
-export CXX=${CXX}
-export FC=${FC}
-export F77=${F77}
-export CPP=${CPP}
-
-export LD_LIBRARY_PATH="${PREFIX}/lib:${LD_LIBRARY_PATH}"
-export CPPFLAGS="${CPPFLAGS} -isystem ${PREFIX}/include"
-export LDFLAGS="${LDFLAGS} -L${PREFIX}/lib"
-export CPATH="${PREFIX}/include:${CPATH}"
-export CFLAGS="${CFLAGS} -Wno-unused-but-set-variable"
-
-if [ $COMPILER = "GNU-Mac" ]; then
-    export MACOSX_DEPLOYMENT_TARGET=13 
-    export LDFLAGS="${LDFLAGS} -Wl,-no_compact_unwind"
-    export CFLAGS="${CFLAGS} -Wno-deprecated-non-prototype"
-fi
+set -a
+. ${SCRIPT_DIR}/_build_getopts.sh
 
 printf "*********************************************************\n"
 printf "*            BUILDING ZLIB STATIC LIBRARY               *\n"
@@ -72,7 +28,6 @@ printf "*********************************************************\n"
 cd ${DEPENDENCY_DIR}/zlib-*
 ./configure --prefix=${PREFIX} --static 
 make 
-
 if [ -w ${PREFIX} ]; then
     make install
 else
