### 2023-08-17 b4c0dd8
Made the PREFIX a realpath
