### 2023-08-17 bf1c3b7
Create the prefix or dependency directories if they don't exist.
