### 2023-08-17 6824e60
Make all the install prefix subdirectories if they don't exist already
