### 2023-08-17 6f46b41
Fixed issue when relative path was used as prefix or dependency
