### 2023-08-17 4aa0e1a
Changed the dependency fetcher to only download the dependencies if they aren't already there, otherwise run make distclean
