### 2023-08-17 8c99d27
Fixed problem that occurs if the LIBS environment variable has leading or trailing spaces
