+++ swiftest/CMakeLists.txt
@@ -37,7 +37,12 @@ IF(USE_OPENMP OR USE_SIMD)
     SET_PROPERTY(TARGET ${SWIFTEST_BINDINGS} APPEND_STRING PROPERTY COMPILE_FLAGS  "${OpenMP_Fortran_FLAGS} ")
 ENDIF()
 
-TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} swiftest ${NETCDF_LIBRARIES} $ENV{LIBS})
+IF (DEFINED ENV{LIBS})
+    STRING(STRIP "$ENV{LIBS}" LIBS)
+    SEPARATE_ARGUMENTS(LIBS NATIVE_COMMAND "$LIBS")
+ENDIF()
+
+TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} swiftest ${NETCDF_LIBRARIES} $LIBS)
 PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_BINDINGS} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${MOD} ${NETCDF_INCLUDE_DIR})
 INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION swiftest)
