+++ cmake/Modules/FindNETCDF.cmake
@@ -131,7 +131,8 @@ ELSE ()
       MESSAGE(FATAL_ERROR "Cannot find nc-config")
    ENDIF ()
    IF (DEFINED ENV{LIBS})
-      SEPARATE_ARGUMENTS(LIBS NATIVE_COMMAND "$ENV{LIBS}")
+      STRING(STRIP "$ENV{LIBS}" LIBS)
+      SEPARATE_ARGUMENTS(LIBS NATIVE_COMMAND "$LIBS")
       LIST(APPEND EXTRA_FLAGS ${LIBS})
    ENDIF()
 
