### 2023-08-17 0cb5674
Fixed some issues related to getting the Mac OSX deployment target version correctly formatted
