+++ swiftest/tests/test_input_output.py
@@ -60,10 +60,6 @@ class TestSwiftest(unittest.TestCase):
       for f in file_list:
          self.assertTrue(os.path.exists(f))
          
-   def test02_read_ic(self):
-      """
-      Tests that Swiftest is able to read a set of pre-existing initial conditions files and that they contain the correct data
-      """
       print("\ntest_read_ic: Test whether we can read back initial conditions files created by test_gen_ic")
       sim = swiftest.Simulation(read_param=True)
       # Check if all names in Dataset read in from file match the expected list of names
@@ -72,12 +68,21 @@ class TestSwiftest(unittest.TestCase):
       # Check to see if all parameter values read in from file match the expected parameters saved when generating the file
       self.assertTrue(all([v == param[k] for k,v in sim.param.items() if k in param]))
      
-   def test03_integrators(self):
+   def test02_integrators(self):
       """
       Tests that Swiftest is able to integrate a collection of massive bodies and test particles with all available integrators
       """ 
       print("\ntest_integrators: Tests that Swiftest is able to integrate a collection of massive bodies and test particles with all available integrators")
-      sim = swiftest.Simulation(read_param=True)
+      sim = swiftest.Simulation()
+      sim.clean()
+
+      # Add the modern planets and the Sun using the JPL Horizons Database.
+      # Add the modern planets and the Sun using the JPL Horizons Database.
+      sim.add_solar_system_body(major_bodies)
+   
+      # Display the run configuration parameters.
+      param = sim.get_parameter(verbose=False)
+      sim.save()
       
       # Add 10 user-defined test particles.
       ntp = 10
@@ -90,18 +95,14 @@ class TestSwiftest(unittest.TestCase):
       omega_tp    = rng.uniform(0.0, 360.0, ntp)
       capm_tp     = rng.uniform(0.0, 360.0, ntp)
 
-      integrators= ["whm","helio","rmvs","symba"]
+      integrators= ["helio","whm","rmvs","symba"]
       sim.add_body(name=name_tp, a=a_tp, e=e_tp, inc=inc_tp, capom=capom_tp, omega=omega_tp, capm=capm_tp)
-      sim.set_parameter(tstart=0.0, tstop=0.02, dt=0.01, istep_out=1, dump_cadence=0)
-      integrators= ["whm","helio","rmvs","symba"]
+      sim.set_parameter(tstart=0.0, tstop=0.10, dt=0.01, istep_out=1, dump_cadence=0)
       for i in integrators:
-         try:
-            sim.run(integrator=i)
-         except:
-            self.fail(f"Failed with integrator {i}")
+         sim.run(integrator=i)
          
          
-   def test04_conservation(self):
+   def test03_conservation(self):
       """
       Tests that Swiftest conserves mass, energy, and momentum to within acceptable tolerances.
       """
