+++ CMakeLists.txt
@@ -60,7 +60,6 @@ SET(LIB ${CMAKE_SOURCE_DIR}/lib)
 SET(BIN ${CMAKE_SOURCE_DIR}/bin)
 SET(MOD ${CMAKE_SOURCE_DIR}/include)
 SET(PY ${CMAKE_SOURCE_DIR}/swiftest)
-#SET(TEST ${CMAKE_SOURCE_DIR}/tests)
 
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${LIB})
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${LIB})
@@ -73,9 +72,6 @@ SET(CMAKE_Fortran_MODULE_DIRECTORY ${MOD})
 ADD_SUBDIRECTORY(${SRC} ${BIN})
 
 ADD_SUBDIRECTORY(${PY})
-# # Set up test directory
-# ENABLE_TESTING()
-# ADD_SUBDIRECTORY(${TEST})
 
 # Add a distclean target to the Makefile
 ADD_CUSTOM_TARGET(distclean 
