+++ src/swiftest/swiftest_module.f90
@@ -395,6 +395,7 @@ module swiftest
       logical                         :: lbeg                 !! True if this is the beginning of a step. This is used so that test particle steps can be calculated 
                                                               !!    separately from massive bodies.  Massive body variables are saved at half steps, and passed to 
                                                               !!    the test particles
+      logical                         :: lfirst = .true.      !! Flag to indicate that this is a new run
    contains
       !> Each integrator will have its own version of the step
       procedure(abstract_step_system), deferred :: step
