### 2023-08-18 150d464
Changed save lfirst variables to nbody_system class variables
