+++ src/swiftest/swiftest_module.f90
@@ -395,7 +395,8 @@ module swiftest
       logical                         :: lbeg                 !! True if this is the beginning of a step. This is used so that test particle steps can be calculated 
                                                               !!    separately from massive bodies.  Massive body variables are saved at half steps, and passed to 
                                                               !!    the test particles
-      logical                         :: lfirst = .true.      !! Flag to indicate that this is a new run
+      logical                         :: lfirst_io   = .true.  !! Flag to indicate that this is the first time to write to a file
+      logical                         :: lfirst_peri = .true.  !! Flag to indicate that this is the first pericenter passage
    contains
       !> Each integrator will have its own version of the step
       procedure(abstract_step_system), deferred :: step
