### 2023-08-18 51440e1
Added tests to the cibuildwheel project definition
