### 2023-08-18 28ec2fd
Adjusted ignores and Mac-specific build variables in the toml
