+++ pyproject.toml
@@ -11,7 +11,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.cibuildwheel]
 dependency-versions = "latest"
-test-requires = "pytest"
+test-requires = ["pytest"]
 test-command = "pytest {package}/tests"
 
 [tool.cibuildwheel.linux]
@@ -19,5 +19,11 @@ environment = {SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF", FFLAGS="${FF
 before-all = [
     "yum install wget m4 doxygen -y",
     "buildscripts/build_dependencies.sh",
+]
+
+[tool.cibuildwheel.macos]
+environment = {MACOSX_DEPLOYMENT_TARGET="10.9", LD_LIBRARY_PATH="${TMPDIR}/lib:${LD_LIBRARY_PATH}",CPPFLAGS="${CPPFLAGS} -isystem ${TMPDIR}/include",LDFLAGS="${LDFLAGS} -L${TMPDIR}/lib -fPIE -Wl,-no_compact_unwind",CPATH="${CPATH} ${TMPDIR}/include}", CFLAGS="${CFLAGS} -Wno-unused-but-set-variable -fPIC -Wno-deprecated-non-prototype", SKBUILD_CONFIGURE_OPTIONS="-DUSE_OPENMP=OFF -DUSE_SIMD=OFF -DBUILD_SHARED_LIBS=OFF",FFLAGS="${FFLAGS} -fPIC"}
+before-all = [
+    "buildscripts/build_dependencies.sh -d ${TMPDIR} -p ${TMPDIR}"
 ]
     
