### 2023-08-18 ddbb4c5
Improved the robustness of the dependency builder. New version of Zlib dropped today, breaking the script. Now all dependencies are pulled from tagged releases on github.com. HDF5 required some adjustments due to how the main folder from github version is named.
