+++ setup.py
@@ -53,8 +53,7 @@ setup(name='swiftest',
             'astropy>=5.1',
             'astroquery>=0.4.6',
             'tqdm>=4.65.0',
-            'py>=1.11',
-            'cython>=3.0.0'
+            'cython>=3.0.0',
       ],
       packages=['swiftest'],
       test_suite="swiftest.tests",
