### 2023-08-18 c6a2ed1
Made more adjustments. Something causes the test to fail in python 3.12 and it still has some weird cython error when building some wheels
