### 2023-08-18 f523ff2
Updated the requirements.txt to match setup.py
