+++ requirements.txt
@@ -1,15 +1,13 @@
-  python>=3.8
-  numpy>=1.24.3
-  pandas>=1.5.3
-  scipy>=1.10.1
-  xarray>=2022.11.0
-  dask>=2022.1
-  bottleneck>=1.3.5
-  h5netcdf>=1.0.2
-  netcdf4>=1.6.2
-  matplotlib>=3.7.1
-  astropy>=5.1
-  astroquery>=0.4.6
-  tqdm>=4.65.0
-  x264>=1!157.20191217
-  ffmpeg>=4.3.2
+python>=3.8
+numpy>=1.24.3
+scipy>=1.10.1
+xarray>=2022.11.0
+dask>=2022.1
+bottleneck>=1.3.5
+h5netcdf>=1.0.2
+netcdf4>=1.6.2
+matplotlib>=3.7.1
+astropy>=5.1
+astroquery>=0.4.6
+tqdm>=4.65.0
+cython>=3.0.0
