### 2023-08-18 a420fb8
Added cibuildwheel artifacts to the distclean script
