+++ distclean.cmake
@@ -19,7 +19,7 @@ ELSE()
     SET(TOPDIR "${CMAKE_SOURCE_DIR}")
 ENDIF()
 
-SET(SKBUILD_DIR "_skbuild")
+SET(CIBW_DIR "_skbuild" "swiftest.egg-info" "wheelhouse")
 
 MACRO(GET_PARENT_DIRECTORIES search_string return_list grandparents)
     FILE(GLOB_RECURSE new_list ${search_string})
@@ -63,7 +63,7 @@ SET(DEL ${TOPDIRECTORIES}
         ${CMAKEFILES}
         ${CMAKETESTING}
         ${CMAKETESTFILES}
-        ${SKBUILD_DIR}
+        ${CIBW_DIR}
 )
 
 # If we are not in the build dir, delete that as well
