### 2023-08-18 cfc7705
Made some adjustments to scripts to ensure the proper gcc version for the dependency build
