### 2023-08-18 bd78cd5
Separated GCC version from MACOSX deployment target variables
