### 2023-08-18 b5d8cff
Got rid of arch definitions from project. Will set from command line
