### 2023-08-18 482bd11
Took away all arch options
