### 2023-08-19 60be996
Added test of whether the linux build system is RH or Debian.
