+++ distclean.cmake
@@ -19,7 +19,7 @@ ELSE()
     SET(TOPDIR "${CMAKE_SOURCE_DIR}")
 ENDIF()
 
-SET(CIBW_DIR "_skbuild" "swiftest.egg-info" "wheelhouse")
+SET(CIBW_DIR "_skbuild" "swiftest.egg-info")
 
 MACRO(GET_PARENT_DIRECTORIES search_string return_list grandparents)
     FILE(GLOB_RECURSE new_list ${search_string})
