### 2023-08-19 82320fb
Updated the pyproject.toml file with better cibuildwheel flags
