### 2023-08-19 7a56780
removed pip upgrade line
