+++ pyproject.toml
@@ -14,7 +14,6 @@ test-command = "pytest {package}/tests"
 test-requires = ['pytest','cython']
 test-skip = "cp312-*"
 skip = ["pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"]
-before-build = "python3 -m pip install -U pip"
 
 [tool.cibuildwheel.linux]
 environment = {SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF", FFLAGS="${FFLAGS} -fPIC", CFLAGS="${CFLAGS} -fPIC", LDFLAGS="${LDFLAGS} -fPIE", LIBS="-lgomp"}
