### 2023-08-20 e259874
Fixed typo in Development Status classifier
