### 2023-08-21 d156d95
More improvements to buildscripts. Can now build with OpenMP on M1 Mac
