+++ buildscripts/build_netcdf-fortran.sh
@@ -43,3 +43,5 @@ if [ $? -ne 0 ]; then
    exit 1
 fi
 
+make distclean
+
