+++ buildscripts/build_hdf5.sh
@@ -49,3 +49,5 @@ if [ $? -ne 0 ]; then
    exit 1
 fi
 
+make distclean
+
