+++ buildscripts/build_netcdf-c.sh
@@ -49,3 +49,4 @@ if [ $? -ne 0 ]; then
    exit 1
 fi
 
+make distclean
