### 2023-08-21 dab72a7
Improved install instructions and project description is now read from README.md
