### 2023-08-21 acafd5b
Replaced more hardcoded homebrew baths with environment variable
