### 2023-08-21 52d4acc
More improvements to build scripts for Mac
