+++ buildscripts/build_swiftest.sh
@@ -30,11 +30,8 @@ SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF"
 
 if [ $OS = "Intel" ]; then 
     FCFLAGS="${CFLAGS} -standard-semantics"
-    FFLAGS=${CFLAGS}
     SKBUILD_CONFIGURE_OPTIONS="${SKBUILD_CONFIGURE_OPTIONS} -DMACHINE_CODE_VALUE=\"SSE2\""
 else
-    FCFLAGS="${CFLAGS}"
-    FFLAGS="${CFLAGS}"
     SKBUILD_CONFIGURE_OPTIONS="${SKBUILD_CONFIGURE_OPTIONS} -DMACHINE_CODE_VALUE=\"generic\""
 fi
 
@@ -63,4 +60,9 @@ fi
 printf "*********************************************************\n"
 
 python3 -m pip install build pip
-python3 -m build
+python3 -m build --sdist
+if [ $OS = "MacOSX" ]; then
+    cibuildwheel --platform macos
+elif [ $OS = "Linux" ]; then
+    cibuildwheel --platform linux
+fi
