### 2023-08-21 2942e28
Made adjustments to CMake files to work with Windows
