+++ cmake/Modules/FindNETCDF.cmake
@@ -9,48 +9,51 @@
 
 # - Finds the NetCDF libraries 
 
-FIND_PATH(NFBIN
-NAMES nf-config
-HINTS 
-   ENV NETCDF_FORTRAN_HOME
-   ENV PATH
-PATH_SUFFIXES
-   bin
-)
+IF (NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
+   FIND_PATH(NFBIN
+   NAMES nf-config
+   HINTS 
+      ENV NETCDF_FORTRAN_HOME
+      ENV PATH
+   PATH_SUFFIXES
+      bin
+   )
 
-IF (NFBIN)
-   SET(CMD "${NFBIN}/nf-config")
-   LIST(APPEND CMD "--includedir")
-   MESSAGE(STATUS "Searching for NetCDF-Fortran include directory using ${CMD}")
-   EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFINCLUDE_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
-   IF (NFINCLUDE_DIR)
-      MESSAGE(STATUS "Found in ${NFINCLUDE_DIR}")
-   ELSE ()
-      MESSAGE(STATUS "Cannot execute ${CMD}")
-      MESSAGE(STATUS "OUTPUT: ${NFINCLUDE_DIR}")
-      MESSAGE(STATUS "RESULT: ${RES}")
-      MESSAGE(STATUS "ERROR : ${ERR}")
-   ENDIF ()
+   IF (NFBIN)
+      SET(CMD "${NFBIN}/nf-config")
+      LIST(APPEND CMD "--includedir")
+      MESSAGE(STATUS "Searching for NetCDF-Fortran include directory using ${CMD}")
+      EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFINCLUDE_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
+      IF (NFINCLUDE_DIR)
+         MESSAGE(STATUS "Found in ${NFINCLUDE_DIR}")
+      ELSE ()
+         MESSAGE(STATUS "Cannot execute ${CMD}")
+         MESSAGE(STATUS "OUTPUT: ${NFINCLUDE_DIR}")
+         MESSAGE(STATUS "RESULT: ${RES}")
+         MESSAGE(STATUS "ERROR : ${ERR}")
+      ENDIF ()
 
-   SET(CMD "${NFBIN}/nf-config")
-   LIST(APPEND CMD "--prefix")
-   MESSAGE(STATUS "Searching for NetCDF-Fortran library directory using ${CMD}")
-   EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFPREFIX_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
-   IF (NFPREFIX_DIR)
-      MESSAGE(STATUS "Found in ${NFPREFIX_DIR}")
-   ELSE ()
-      MESSAGE(STATUS "Cannot execute ${CMD}")
-      MESSAGE(STATUS "OUTPUT: ${NFPREFIX_DIR}")
-      MESSAGE(STATUS "RESULT: ${RES}")
-      MESSAGE(STATUS "ERROR : ${ERR}")
-   ENDIF ()
+      SET(CMD "${NFBIN}/nf-config")
+      LIST(APPEND CMD "--prefix")
+      MESSAGE(STATUS "Searching for NetCDF-Fortran library directory using ${CMD}")
[…]
