+++ CMakeLists.txt
@@ -55,11 +55,11 @@ ENDIF()
 LIST(APPEND CMAKE_MODULE_PATH "${CMAKE_SOURCE_DIR}/cmake/Modules/")
 
 # Define some directories that are important to the build
-SET(SRC ${CMAKE_SOURCE_DIR}/src)
-SET(LIB ${CMAKE_SOURCE_DIR}/lib)
-SET(BIN ${CMAKE_SOURCE_DIR}/bin)
-SET(MOD ${CMAKE_SOURCE_DIR}/include)
-SET(PY ${CMAKE_SOURCE_DIR}/swiftest)
+SET(SRC "${CMAKE_SOURCE_DIR}/src" CACHE FILEPATH "Location of source files")
+SET(LIB "${CMAKE_SOURCE_DIR}/lib" CACHE FILEPATH "Library file location")
+SET(BIN "${CMAKE_SOURCE_DIR}/bin" CACHE FILEPATH "Binary file location")
+SET(MOD "${CMAKE_SOURCE_DIR}/include" CACHE FILEPATH "Include directory")
+SET(PY "${CMAKE_SOURCE_DIR}/swiftest" CACHE FILEPATH "Python source directory")
 
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${LIB})
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${LIB})
@@ -75,5 +75,5 @@ ADD_SUBDIRECTORY(${PY})
 
 # Add a distclean target to the Makefile
 ADD_CUSTOM_TARGET(distclean 
-    COMMAND ${CMAKE_COMMAND} -P ${CMAKE_SOURCE_DIR}/distclean.cmake
+    COMMAND ${CMAKE_COMMAND} -P "${CMAKE_SOURCE_DIR}/distclean.cmake"
 )
