### 2023-08-21 9457f24
Improved Intel Mac version of build script
