+++ buildscripts/set_compilers.sh
@@ -89,29 +89,54 @@ case $OS in
         CPP=${CPP:-$(command -v cpp)}
         ;;
     MacOSX)
-        if $(brew --version &> /dev/null); then 
-            brew install llvm@16 libomp 
+        if [ $ARCH = "arm64" ]; then
+            if $(brew --version &> /dev/null); then 
+                brew install llvm@16 libomp 
+            else
+                echo \"Please install Homebrew first\" 
+                exit 1 
+            fi
+            COMPILER_PREFIX=${COMPILER_PREFIX:-"${HOMEBREW_PREFIX}/opt/llvm"}
+            CC=${CC:-${COMPILER_PREFIX}/bin/clang}
+            CXX=${CXX:-${COMPILER_PREFIX}/bin/clang++}
+            CPP=${CPP:-${COMPILER_PREFIX}/bin/clang-cpp}
+            AR=${AR:-${COMPILER_PREFIX}/bin/llvm-ar}
+            NM=${NM:-${COMPILER_PREFIX}/bin/llvm-nm}
+            RANLIB=${RANLIB:-${COMPILER_PREFIX}/bin/llvm-ranlib}
+            FROOT=$(realpath $(dirname $(command -v gfortran))/..) 
+            FC=$(command -v gfortran)
+            LD_LIBRARY_PATH="${COMPILER_PREFIX}/lib:${FROOT}/lib:${LD_LIBRARY_PATH}"
+            LDFLAGS="-L${HOMEBREW_PREFIX}/opt/llvm/lib/c++ -Wl,-rpath,${HOMEBREW_PREFIX}/opt/llvm/lib/c+ -L${HOMEBREW_PREFIX}/opt/libomp/lib -Wl,-no_compact_unwind"
+            CPPFLAGS="-isystem ${HOMEBREW_PREFIX}/opt/libomp/include"
+            LIBS="-lomp ${LIBS}"
+            CPATH="${FROOT}/include:${CPATH}"
+            CXXFLAGS="${CFLAGS} ${CXXFLAGS}"
+            FCFLAGS="${CFLAGS} ${FCFLAGS}"
+            CFLAGS="-mmacosx-version-min=${MACOSX_DEPLOYMENT_TARGET} -Wno-deprecated-non-prototype ${CFLAGS}"
         else
-            echo \"Please install Homebrew first\" 
-            exit 1 
-        fi
-        COMPILER_PREFIX=${COMPILER_PREFIX:-"${HOMEBREW_PREFIX}/opt/llvm"}
-        CC=${CC:-${COMPILER_PREFIX}/bin/clang}
-        CXX=${CXX:-${COMPILER_PREFIX}/bin/clang++}
-        CPP=${CPP:-${COMPILER_PREFIX}/bin/clang-cpp}
-        AR=${AR:-${COMPILER_PREFIX}/bin/llvm-ar}
-        NM=${NM:-${COMPILER_PREFIX}/bin/llvm-nm}
-        RANLIB=${RANLIB:-${COMPILER_PREFIX}/bin/llvm-ranlib}
-        FROOT=$(realpath $(dirname $(command -v gfortran))/..) 
-        FC=$(command -v gfortran)
-        LD_LIBRARY_PATH="${COMPILER_PREFIX}/lib:${FROOT}/lib:${LD_LIBRARY_PATH}"
-        LDFLAGS="-L${HOMEBREW_PREFIX}/opt/llvm/lib/c++ -Wl,-rpath,${HOMEBREW_PREFIX}/opt/llvm/lib/c+ -L${HOMEBREW_PREFIX}/opt/libomp/lib -Wl,-no_compact_unwind"
[…]
