+++ pyproject.toml
@@ -13,7 +13,7 @@ build-backend = "setuptools.build_meta"
 test-command = "pytest {package}/tests"
 test-requires = ['pytest','cython']
 test-skip = "cp312-*"
-skip = ["pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"]
+skip = "pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
 
 [tool.cibuildwheel.linux]
 environment = {SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=OFF", FFLAGS="${FFLAGS} -fPIC", CFLAGS="${CFLAGS} -fPIC", LDFLAGS="${LDFLAGS} -fPIE", LIBS="-lgomp"}
@@ -23,7 +23,6 @@ before-all = [
 ]
 
 [tool.cibuildwheel.macos]
-skip = "cp38-*"
 before-all = [
     "buildscripts/build_dependencies.sh -d ${TMPDIR} -p ${TMPDIR} -m ${MACOSX_DEPLOYMENT_TARGET}"
 ]
