+++ cmake/Modules/FindNETCDF.cmake
@@ -67,6 +67,7 @@ MESSAGE(STATUS "NetCDF-Fortran include directory: ${NETCDF_INCLUDE_DIR}")
 
 IF (BUILD_SHARED_LIBS OR CMAKE_SYSTEM_NAME STREQUAL "Windows") 
    SET(NETCDFF "netcdff")
+   SET(NETCDF "netcdf")
 ELSE ()
    SET(NETCDFF "libnetcdff.a")
    SET(NETCDF  "libnetcdf.a")
@@ -133,6 +134,11 @@ ELSE ()
       ELSE ()
          MESSAGE(FATAL_ERROR "Cannot find nc-config")
       ENDIF ()
+   ELSE ()
+      FIND_PACKAGE(NETCDF CONFIG QUIET)
+      IF (NETCDF_FOUND)
+         set(NETCDF_LIBRARY "${netCDF_LIBRARIES}")
+      ENDIF()
    ENDIF()
    
    IF (DEFINED ENV{LIBS})
