### 2023-08-22 f276bcd
Made adjustments to CMake files to handle WIndows style paths better
