### 2023-08-22 4b60068
Reverted the positioning of the environment call that was accidentally changed
