### 2023-08-22 57b084a
Improvements to Windows compatibility
