+++ swiftest/CMakeLists.txt
@@ -37,12 +37,7 @@ IF(USE_OPENMP OR USE_SIMD)
     SET_PROPERTY(TARGET ${SWIFTEST_BINDINGS} APPEND_STRING PROPERTY COMPILE_FLAGS  "${OpenMP_Fortran_FLAGS} ")
 ENDIF()
 
-IF (DEFINED ENV{LIBS})
-    STRING(STRIP "$ENV{LIBS}" LIBS)
-    SEPARATE_ARGUMENTS(LIBS NATIVE_COMMAND "$LIBS")
-ENDIF()
-
-TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} swiftest ${NETCDF_LIBRARIES} $LIBS)
+TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} swiftest ${NETCDF_LIBRARIES})
 PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_BINDINGS} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${MOD} ${NETCDF_INCLUDE_DIR})
 INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION swiftest)
