### 2023-08-22 511a366
Fixed curl command syntax to work in the Linux container
