+++ pyproject.toml
@@ -19,7 +19,7 @@ skip = "pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
 [tool.cibuildwheel.linux]
 environment = {FFLAGS="${FFLAGS} -fPIC", CFLAGS="${CFLAGS} -fPIC", LDFLAGS="${LDFLAGS} -fPIE", LIBS="-lgomp"}
 before-all = [
-    "yum install wget m4 doxygen  -y ||  apt-get install wget m4 doxygen -y",
+    "yum install doxygen  -y ||  apt-get install doxygen -y",
     "buildscripts/build_dependencies.sh -d /_dependencies -p /usr/local",
 ]
 
