### 2023-08-23 bce1ea1
Adding Spherical Harmonics flags and functionality
