### 2023-08-23 #5409 @profminton
Hello reviewers. I spent the last month+ coming up to speed on a number of tools and techniques that I hope will result in closing out the review.  One can now install Swiftest on a variety of systems (Linux and M1 Macs at the moment. Intel Mac and Windows shortly) using `pip install swiftest`.

The project was restructured in the following ways:

- The main "driver" program in Fortran was re-written to be a subroutine with C bindings such that it could be compiled into a Cython module. Now, instead of calling a standalone executable `swiftest_driver` executable using a subprocess, we now call the Cython module directly. This simplifies the installation, as one no longer has to keep track of executable paths and shell environment variables when running the code via Python. We still compile the binary executable, so it can still function as a standalone application without Python
- The Python module was restructured for packaging using the `cibuildwheel` tool. The automated tests are now called during the building of wheels. I also wrote a collection of shell scripts to help automate building all of the required dependencies as static libraries so that the final Cython module has a minimal amount of dependencies in the end.
- Python wheels were generated for Linux and Mac. I am very (very) close to getting a Windows version built. This will probably get done in the next release.
- I switched to a calendar versioning system of the form: YYYY.MM.MINOR
- The README.md has been updated to reflect the new simplified installation.
