### 2023-08-28 f8c8651
Adjusted Mac build scripts
