+++ buildscripts/build_swiftest.sh
@@ -53,6 +53,7 @@ printf "CPATH: ${CPATH}\n"
 printf "LD_LIBRARY_PATH: ${LD_LIBRARY_PATH}\n"
 printf "LDFLAGS: ${LDFLAGS}\n"
 printf "NETCDF_FORTRAN_HOME: ${NETCDF_FORTRAN_HOME}\n"
+printf "NETCDF_INCLUDE: ${NETCDF_INCLUDE}\n"
 printf "SKBUILD_CONFIGURE_OPTIONS: ${SKBUILD_CONFIGURE_OPTIONS}\n"
 if [ $OS = "MacOSX" ]; then
     printf "MACOSX_DEPLOYMENT_TARGET: ${MACOSX_DEPLOYMENT_TARGET}\n"
