### 2023-08-28 06833e0
Made adjustments to python dependency package lists
