+++ environment.yml
@@ -7,9 +7,9 @@ channels:
 dependencies:
   - python>=3.8
   - numpy>=1.24.3
-  - pandas>=1.5.3
   - scipy>=1.10.1
   - xarray>=2022.11.0
+  - distributed>=2022.1
   - dask>=2022.1
   - bottleneck>=1.3.5
   - h5netcdf>=1.0.2
@@ -20,4 +20,4 @@ dependencies:
   - tqdm>=4.65.0
   - x264>=1!157.20191217
   - ffmpeg>=4.3.2
-  - py>=1.11
+  - cython>=3.0.0
