### 2023-08-28 8dbecd6
Made a number of improvements to the Windows build
