+++ cmake/Modules/SetFortranFlags.cmake
@@ -63,311 +63,393 @@ ENDIF(CMAKE_Fortran_FLAGS_RELEASE AND CMAKE_Fortran_FLAGS_TESTING AND CMAKE_Fort
 # flag is given in the call).  This way unknown compiles are supported.
 #######################################################################
 
+IF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
+    SET(WINOPT True)
+ELSE ()
+    SET(WINOPT False)
+ENDIF ()
 #####################
 ### GENERAL FLAGS ###
 #####################
 
+
 # Free form
 IF (COMPILER_OPTIONS STREQUAL "GNU")
-        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                        Fortran "-ffree-form" # GNU
-                        ) 
+    SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+        Fortran "-ffree-form" # GNU
+        ) 
 
-        # Don't add underscores in symbols for C-compatability
-        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                        Fortran "-fno-underscoring" # GNU
-                        ) 
-        # Compile code assuming that IEEE signaling NaNs may generate user-visible traps during floating-point operations. 
-        # Setting this option disables optimizations that may change the number of exceptions visible with signaling NaNs. 
-        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                        Fortran "-fsignaling-nans " # GNU
-                        ) 
-                
-        # Allows for lines longer than 80 characters without truncation
-        SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                        Fortran "-ffree-line-length-512" # GNU (gfortran)
-                        )
+    # Don't add underscores in symbols for C-compatability
+    SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+        Fortran "-fno-underscoring" # GNU
+        ) 
+    # Compile code assuming that IEEE signaling NaNs may generate user-visible traps during floating-point operations. 
+    # Setting this option disables optimizations that may change the number of exceptions visible with signaling NaNs. 
+    SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+        Fortran "-fsignaling-nans " # GNU
+        ) 
+    
+    # Allows for lines longer than 80 characters without truncation
+    SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
+        Fortran "-ffree-line-length-512" # GNU (gfortran)
+        )
 ELSEIF (COMPILER_OPTIONS STREQUAL "Intel")
-        # Disables right margin wrapping in list-directed output
+    # Disables right margin wrapping in list-directed output
+    IF (WINOPT)
         SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
-                        Fortran  "-no-wrap-margin" # Intel
-                                "/wrap-margin-"   # Intel Windows        
-                        )
-
+            Fortran  "/wrap-margin-"   # Intel Windows    
+        )
         # Aligns a variable to a specified boundary and offset
[…]
