+++ cmake/Modules/FindNETCDF.cmake
@@ -9,7 +9,14 @@
 
 # - Finds the NetCDF libraries 
 
-IF (NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
+FILE(TO_CMAKE_PATH "${CMAKE_SOURCE_DIR}/_dependencies/${CMAKE_SYSTEM_NAME}_${CMAKE_SYSTEM_PROCESSOR}_${COMPILER_OPTIONS}" NXPREFIX_CANDIDATE)
+IF (EXISTS ${NXPREFIX_CANDIDATE})
+   SET(NCPREFIX_DIR ${NXPREFIX_CANDIDATE} CACHE PATH "Location of provided NetCDF-C dependencies")
+   SET(NFPREFIX_DIR ${NXPREFIX_CANDIDATE} CACHE PATH "Location of provided NetCDF-Fortran dependencies")
+   SET(H5PREFIX_DIR ${NXPREFIX_CANDIDATE} CACHE PATH "Location of provided HDF5 dependencies")
+   SET(ZPREFIX_DIR  ${NXPREFIX_CANDIDATE} CACHE PATH "Location of provided zlib dependencies")
+   SET(NFINCLUDE_DIR "${NFPREFIX_DIR}/include" CACHE PATH "Location of provided netcdf.mod")
+ELSEIF(NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
    FIND_PATH(NFBIN
    NAMES nf-config
    HINTS 
@@ -67,14 +74,19 @@ MESSAGE(STATUS "NetCDF-Fortran include directory: ${NETCDF_INCLUDE_DIR}")
 
 IF (BUILD_SHARED_LIBS) 
    SET(NETCDFF "netcdff")
-   SET(NETCDF "netcdf")
 ELSE ()
    IF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
       SET(NETCDFF "netcdff.lib")
       SET(NETCDF  "netcdf.lib")
+      SET(HDF5    "libhdf5.lib")
+      SET(HDF5_HL "libhdf5_hl.lib")
+      SET(ZLIB    "zlibstatic.lib")
    ELSE ()
       SET(NETCDFF "libnetcdff.a")
       SET(NETCDF  "libnetcdf.a")
+      SET(HDF5    "libhdf5.a")
+      SET(HDF5_HL "libhdf5_hl.a")
+      SET(ZLIB    "libz.a")
    ENDIF()
 ENDIF()
 
@@ -99,6 +111,7 @@ ELSE ()
    FIND_LIBRARY(NETCDF_LIBRARY 
       NAMES ${NETCDF} 
       HINTS
+         ${NCPREFIX_DIR}
          ENV NETCDF_HOME
          ENV LD_LIBRARY_PATH
       PATH_SUFFIXES
@@ -108,7 +121,46 @@ ELSE ()
    )
 
    MESSAGE(STATUS "NetCDF-C Library: ${NETCDF_LIBRARY}")
-   IF (NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
+   IF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
+      FIND_LIBRARY(HDF5_LIBRARY 
+         NAMES ${HDF5}
+         HINTS 
+            ${H5PREFIX_DIR}
+            ENV HDF5_ROOT
+            ENV LD_LIBRARY_PATH
+         PATH_SUFFIXES
+            lib
+            ${CMAKE_LIBRARY_ARCHITECTURE}
+         REQUIRED
+      )
+
+      FIND_LIBRARY(HDF5_HL_LIBRARY 
+         NAMES ${HDF5_HL}
+         HINTS 
[…]

[…]
