### 2023-08-29 1551c09
Updated version number and made a number of changes aimed at getting a stable MacOS build
