### 2023-08-29 93e69a4
Made a number of changes aimed at getting it to compile using older gfortran versions that don't support quad precision. Also fixed the version statement.
