### 2023-08-30 2c5b377
Made several changes to get a working version of the code on MacOS 11.0. Switched back to a shared library build for portability
