+++ buildscripts/build_netcdf-fortran.sh
@@ -19,7 +19,7 @@ LIBS="$(${PREFIX}/bin/nc-config --libs --static)"
 
 printf "\n"
 printf "*********************************************************\n"
-printf "*       BUILDING NETCDF-FORTRAN STATIC LIBRARY          *\n"
+printf "*          BUILDING NETCDF-FORTRAN LIBRARY              *\n"
 printf "*********************************************************\n"
 printf "LIBS: ${LIBS}\n"
 printf "CFLAGS: ${CFLAGS}\n"
@@ -30,7 +30,8 @@ printf "LDFLAGS: ${LDFLAGS}\n"
 printf "*********************************************************\n"
 
 cd ${DEPENDENCY_DIR}/netcdf-fortran-*
-./configure --disable-shared --with-pic --disable-zstandard-plugin --enable-large-file-tests=no  --enable-filter-test=no --prefix=${PREFIX}  
+#./configure --disable-shared --with-pic --disable-zstandard-plugin --enable-large-file-tests=no  --enable-filter-test=no --prefix=${PREFIX}  
+./configure --enable-large-file-tests=no  --enable-filter-test=no --prefix=${PREFIX}  
 make && make check i
 if [ -w ${PREFIX} ]; then
     make install

[…]
