### 2023-08-30 0d7c712
More changes to improve the robustness of the buildscripts
