### 2023-08-30 4ac2275
More tweaks to the build scripts to get the right libraries into the right place
