+++ buildscripts/build_netcdf-fortran.sh
@@ -30,7 +30,6 @@ printf "LDFLAGS: ${LDFLAGS}\n"
 printf "*********************************************************\n"
 
 cd ${DEPENDENCY_DIR}/netcdf-fortran-*
-#./configure --disable-shared --with-pic --disable-zstandard-plugin --enable-large-file-tests=no  --enable-filter-test=no --prefix=${PREFIX}  
 ./configure --enable-large-file-tests=no  --enable-filter-test=no --prefix=${PREFIX}  
 make && make check i
 if [ -w ${PREFIX} ]; then
@@ -38,6 +37,8 @@ if [ -w ${PREFIX} ]; then
 else
     sudo make install
 fi
+rsync -a ${PREFIX}/lib/libnetcdff* ${ROOT_DIR}/lib/
+rsync -a ${PREFIX}/include/netcdf.mod ${ROOT_DIR}/include/
 
 if [ $? -ne 0 ]; then
    printf "netcdf-fortran could not be compiled.\n"
