+++ buildscripts/build_netcdf-c.sh
@@ -30,11 +30,7 @@ printf "HDF5_ROOT: ${HDF5_ROOT}\n"
 printf "*********************************************************\n"
 
 cd ${DEPENDENCY_DIR}/netcdf-c-*
-#COPTS="--disable-shared --disable-dap --disable-byterange --disable-testsets --prefix=${PREFIX}"
 COPTS="--disable-testsets --prefix=${PREFIX}"
-#if [ !  $OS = "MacOSX" ]; then
-#    COPTS="${COPTS} --disable-libxml2"
-#fi
 printf "COPTS: ${COPTS}\n"
 ./configure $COPTS
 make && make check 
@@ -44,6 +40,7 @@ if [ -w ${PREFIX} ]; then
 else
     sudo make install
 fi
+rsync -va ${PREFIX}/lib/libnetcdf* ${ROOT_DIR}/lib/
 
 if [ $? -ne 0 ]; then
    printf "netcdf-c could not be compiled."\n
