### 2023-08-31 be7283e
More updates to build scripts for more stable Mac build
