+++ requirements.txt
@@ -12,3 +12,4 @@ astropy>=5.1
 astroquery>=0.4.6
 tqdm>=4.65.0
 cython>=3.0.0
+pyshtools
