+++ environment.yml
@@ -21,3 +21,4 @@ dependencies:
   - x264>=1!157.20191217
   - ffmpeg>=4.3.2
   - cython>=3.0.0
+  - pyshtools
