### 2023-08-31 8f8d7f4
Fixed some issues with syntax and code organization
