+++ src/swiftest/swiftest_util.f90
@@ -2455,6 +2455,7 @@ contains
          nc%file_name = param%outfile
          call nbody_system%initialize_output_file(nc, param) 
          call nc%close()
+
       end associate
 
       return
