### 2023-09-01 abe9c52
Switched back to using Homebrew, as it is a much more reliable way to get a build that actually works
