### 2023-09-02 3357312
Fixed several issues with build scripts that allow for shared library builds on Mac and Linux
