+++ pyproject.toml
@@ -19,6 +19,6 @@ skip = "pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
 [tool.cibuildwheel.linux]
 environment = {FFLAGS="${FFLAGS} -fPIC", CFLAGS="${CFLAGS} -fPIC", LDFLAGS="${LDFLAGS} -fPIE", LIBS="-lgomp"}
 before-all = [
-    "yum install doxygen  -y ||  apt-get install doxygen -y",
+    "yum install doxygen libxml2-devel -y ||  apt-get install doxygen libxml2-dev -y",
     "buildscripts/build_dependencies.sh -p /usr/local",
 ]
