### 2023-09-05 8395276
fixed a typo in the normalization for reference radius
