### 2023-09-05 1b7d9b9
added c_lm to xarray dataset in init_cond.vec2xr
