### 2023-09-06 70a7c7d
write function to convert c_lm array to numpy array " "
