### 2023-09-07 2eba89b
Major restructuring of build system. Switched to scikit-build-core for futureproofing, which means getting rid of setup.py and relying on pyproject.toml. Working on making the Linux build more efficient
