+++ buildscripts/build_netcdf-c.sh
@@ -45,5 +45,3 @@ if [ $? -ne 0 ]; then
    printf "netcdf-c could not be compiled."\n
    exit 1
 fi
-
-make distclean
