+++ buildscripts/build_hdf5.sh
@@ -48,6 +48,3 @@ if [ $? -ne 0 ]; then
    printf "hdf5 could not be compiled.\n"
    exit 1
 fi
-
-make distclean
-
