+++ buildscripts/build_zlib.sh
@@ -39,5 +39,3 @@ if [ $? -ne 0 ]; then
    printf "zlib could not be compiled.\n"
    exit 1
 fi
-
-make distclean
