+++ pyproject.toml
@@ -1,24 +1,69 @@
+[project]
+name = "swiftest"
+version = "2023.9.0"
+authors=[
+    {name = 'David A. Minton', email='daminton@purdue.edu'},
+    {name = 'Carlisle Wishard'},
+    {name = 'Jennifer Pouplin'},
+    {name = 'Jake Elliott'},
+    {name = 'Dana Singh'},
+    {name = 'Kaustub Anand'},
+]
+maintainers = [
+    {name = 'David A. Minton', email='daminton@purdue.edu'},
+]
+readme = "README.md"
+requires-python=">=3.8"
+license={file = "LICENSE.txt"}
+classifiers=[
+'Development Status :: 3 - Alpha',
+'Intended Audience :: Science/Research',
+'Topic :: Scientific/Engineering :: Astronomy',
+'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
+'Programming Language :: Python :: 3',
+]
+keywords=['astronomy','astrophysics', 'planetary', 'nbody integrator', 'symplectic', 'wisdom-holman']
+dependencies = [
+    'numpy>=1.24.3',
+    'scipy>=1.10.1',
+    'xarray>=2022.11.0',
+    'dask>=2022.1',
+    'distributed>=2022.1',
+    'bottleneck>=1.3.5',
+    'h5netcdf>=1.0.2',
+    'netcdf4>=1.6.2',
+    'matplotlib>=3.7.1',
+    'astropy>=5.1',
+    'astroquery>=0.4.6',
+    'tqdm>=4.65.0',
+    'cython>=3.0.0',
+]
+
+[project.urls]
+Repository = 'https://github.itap.purdue.edu/MintonGroup/swiftest'
+
 [build-system]
 requires = [
-    "setuptools>=42",
-    "scikit-build>=0.17",
-    "cmake>=3.5.0",
+    "scikit-build-core",
     "cython>=3.0.0",
-    "ninja",
     "pytest",
 ]
-build-backend = "setuptools.build_meta"
+build-backend = "scikit_build_core.build"
+
+[tool.scikit-build]
+cmake.args = ["-DBUILD_SHARED_LIBS=ON","-DUSE_SIMD=OFF"]
+sdist.include = ["src/globals/globals_module.f90.in","swiftest/*.py","swiftest/*.pyx","swiftest/*.h"]
 
 [tool.cibuildwheel]
-environment = {SKBUILD_CONFIGURE_OPTIONS="-DBUILD_SHARED_LIBS=ON -DUSE_SIMD=OFF"}
 test-command = "pytest {package}/tests"
-test-requires = ['pytest','cython']
+test-requires = ['pytest']
 test-skip = "cp312-*"
 skip = "pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
+build-verbosity = 1
 
 [tool.cibuildwheel.linux]
-environment = {FFLAGS="${FFLAGS} -fPIC", CFLAGS="${CFLAGS} -fPIC", LDFLAGS="${LDFLAGS} -fPIE", LIBS="-lgomp"}
+repair-wheel-command = "auditwheel repair -w {dest_dir} {wheel}"
+environment = {NETCDF_FORTRAN_HOME="/usr/local", NETCDF_INCLUDE="/usr/local/include", LD_LIBRARY_PATH="/usr/local/lib:/project/lib", CPATH="/usr/local/include:/project/include"}
 before-all = [
-    "yum install doxygen libxml2-devel -y ||  apt-get install doxygen libxml2-dev -y",
-    "buildscripts/build_dependencies.sh -p /usr/local",
+    "yum install -y epel-release  && yum install -y netcdf-fortran netcdf-devel ||  apt-get install -y netcdf-fortran",
 ]
