### 2023-09-07 37c7ca9
Fixed some remaining issues with the Linux build scripts
