+++ buildscripts/build_netcdf-c.sh
@@ -16,6 +16,19 @@ set -a
 ARGS=$@
 . ${SCRIPT_DIR}/_build_getopts.sh ${ARGS}
 
+
+NC_VER="4.9.2"
+
+printf "*********************************************************\n"
+printf "*            FETCHING NETCDF-C SOURCE                   *\n"
+printf "*********************************************************\n"
+printf "Copying files to ${DEPENDENCY_DIR}\n"
+
+if [ ! -d ${DEPENDENCY_DIR}/netcdf-c-${NC_VER} ]; then
+    [ -d ${DEPENDENCY_DIR}/netcdf-c-* ] && rm -rf ${DEPENDENCY_DIR}/netcdf-c-*
+    curl -s -L https://github.com/Unidata/netcdf-c/archive/refs/tags/v${NC_VER}.tar.gz | tar xvz -C ${DEPENDENCY_DIR}
+fi
+
 printf "\n"
 printf "*********************************************************\n"
 printf "*              BUILDING NETCDF-C LIBRARY                *\n"
