+++ buildscripts/build_netcdf-fortran.sh
@@ -15,7 +15,15 @@ set -a
 ARGS=$@
 . ${SCRIPT_DIR}/_build_getopts.sh ${ARGS}
 
-LIBS="$(${PREFIX}/bin/nc-config --libs --static)"
+NF_VER="4.6.1"
+printf "*********************************************************\n"
+printf "*          FETCHING NETCDF-FORTRAN SOURCE                  *\n"
+printf "*********************************************************\n"
+printf "Copying files to ${DEPENDENCY_DIR}\n"
+if [ ! -d ${DEPENDENCY_DIR}/netcdf-fortran-${NF_VER} ]; then
+    [ -d ${DEPENDENCY_DIR}/netcdf-fortran-* ] && rm -rf ${DEPENDENCY_DIR}/netcdf-fortran-*
+    curl -s -L https://github.com/Unidata/netcdf-fortran/archive/refs/tags/v${NF_VER}.tar.gz | tar xvz -C ${DEPENDENCY_DIR}
+fi 
 
 printf "\n"
 printf "*********************************************************\n"
