+++ pyproject.toml
@@ -57,8 +57,7 @@ sdist.include = ["src/globals/globals_module.f90.in","swiftest/*.py","swiftest/*
 [tool.cibuildwheel]
 test-command = "pytest {package}/tests"
 test-requires = ['pytest']
-test-skip = "cp312-*"
-skip = "pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
+skip = "cp312-* pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
 build-verbosity = 1
 
 [tool.cibuildwheel.linux]
