### 2023-09-07 e01eb63
Added license file to .gitignore
