### 2023-09-07 bd82507
Fixed an error with no c_lm input
