### 2023-09-07 6aad712
Restructured MacOS build scripts and prepared for GitHub actions.
