+++ .github/workflows/build_wheels.yml
@@ -1,6 +1,6 @@
 name: swiftest
 
-on: [push]
+on: push
 
 jobs:
   build_wheels:
@@ -15,6 +15,9 @@ jobs:
 
       - name: Build wheels
         uses: pypa/cibuildwheel@v2.15.0
+        env:
+          CIBW_ARCHS_MACOS: x86_64 arm64
+          CIBW_ARCHS_LINUX: x86_64 aarch64
         with:
           package-dir: .
           output-dir: wheelhouse
@@ -23,3 +26,17 @@ jobs:
       - uses: actions/upload-artifact@v3
         with:
           path: ./wheelhouse/*.whl
+
+
+  build_sdist:
+    name: Build source distribution
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Build sdist
+        run: pipx run build --sdist
+
+      - uses: actions/upload-artifact@v3
+        with:
+          path: dist/*.tar.gz
