### 2023-09-07 4efc038
Fixed option that works differently in older MacOS
