+++ pyproject.toml
@@ -63,7 +63,7 @@ build-verbosity = 1
 [tool.cibuildwheel.macos.environment]
 ROOT_DIR="$(pwd)"
 ARCH="$(uname -m)"
-MACOSX_DEPLOYMENT_TARGET="$(sw_vers --ProductVersion)"
+MACOSX_DEPLOYMENT_TARGET="$(sw_vers -productVersion)"
 PREFIX="${ROOT_DIR}"
 LD_LIBRARY_PATH="/usr/local/lib:${PREFIX}/lib:${HOMEBREW_PREFIX}/lib"
 LDFLAGS="-Wl,-rpath,${ROOT_DIR}/lib -Wl,-rpath,${PREFIX}/lib -Wl,-rpath,/usr/local/lib -Wl,-no_compact_unwind -L${PREFIX}/lib" 
