+++ pyproject.toml
@@ -91,6 +91,7 @@ RANLIB="/usr/bin/ranlib"
 
 [tool.cibuildwheel.macos]
 before-all = [
+    "brew install coreutils",
     "LIBS=\"\" buildscripts/build_dependencies.sh -p ${PREFIX} -d ${TMPDIR}/swiftest.build -m ${MACOSX_DEPLOYMENT_TARGET}"
 ]
 
