+++ pyproject.toml
@@ -65,6 +65,7 @@ ROOT_DIR="$(pwd)"
 ARCH="$(uname -m)"
 MACOSX_DEPLOYMENT_TARGET="$(sw_vers -productVersion)"
 PREFIX="${ROOT_DIR}"
+HOMEBREW_PREFIX="$(brew --prefix)"
 LD_LIBRARY_PATH="/usr/local/lib:${PREFIX}/lib:${HOMEBREW_PREFIX}/lib"
 LDFLAGS="-Wl,-rpath,${ROOT_DIR}/lib -Wl,-rpath,${PREFIX}/lib -Wl,-rpath,/usr/local/lib -Wl,-no_compact_unwind -L${PREFIX}/lib" 
 CPATH="/usr/local/include:${PREFIX}/include:${HOMEBREW_PREFIX}/include:${ROOT_DIR}/include"
@@ -81,7 +82,7 @@ NETCDF_FORTRAN_HOME="${PREFIX}"
 NETCDF_INCLUDE="${PREFIX}"
 NCDIR="${PREFIX}"
 NFDIR="${PREFIX}"
-FC="${HOMEBREW_PREFIX}/bin/gfortran-12"
+FC="$(command -v gfortran-12)"
 CC="/usr/bin/clang"
 CXX="/usr/bin/clang++"
 CPP="/usr/bin/cpp"
