### 2023-09-07 c3378eb
Added libzip to MacOS install due to failure on GitHub runner
