+++ pyproject.toml
@@ -92,7 +92,7 @@ RANLIB="/usr/bin/ranlib"
 
 [tool.cibuildwheel.macos]
 before-all = [
-    "brew install coreutils",
+    "brew install coreutils libzip",
     "LIBS=\"\" buildscripts/build_dependencies.sh -p ${PREFIX} -d ${TMPDIR}/swiftest.build -m ${MACOSX_DEPLOYMENT_TARGET}"
 ]
 
