### 2023-09-07 741bfef
Tweaked scripts to get remote Mac build to work
