### 2023-09-08 a5e2b01
More tweaks attempting to get a working build on a Mac runner
