+++ .github/workflows/build_wheels.yml
@@ -9,7 +9,8 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-20.04, macos-11, macos-12, macos-13]
-
+    env:
+      SYSTEM_VERSION_COMPAT: 0
     steps:
       - uses: actions/checkout@v3
 
