### 2023-09-08 298f1f8
Attempting to streamline the universal build on the Mac runner
