+++ .github/workflows/build_wheels.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Build wheels
         uses: pypa/cibuildwheel@v2.15.0
         env:
-          CIBW_ARCHS_MACOS: x86_64 arm64
+          CIBW_ARCHS_MACOS: universal2
           CIBW_ARCHS_LINUX: x86_64 aarch64
         with:
           package-dir: .
