### 2023-09-08 97e406e
Added missing F77 environment variable
