+++ pyproject.toml
@@ -83,6 +83,7 @@ NETCDF_INCLUDE="${PREFIX}"
 NCDIR="${PREFIX}"
 NFDIR="${PREFIX}"
 FC="$(command -v gfortran-12)"
+F77="${FC}"
 CC="/usr/bin/clang"
 CXX="/usr/bin/clang++"
 CPP="/usr/bin/cpp"
