### 2023-09-08 cbe43e2
More tweaks to scripts attempting to get arm64 build using a self-hosted runner.
