+++ .github/workflows/build_wheels.yml
@@ -9,6 +9,10 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-20.04, macos-11, macos-12, macos-13]
+        arch: [x86_64]
+        include: 
+          - os: self-hosted
+            arch: arm64
     env:
       SYSTEM_VERSION_COMPAT: 0
     steps:
@@ -17,7 +21,7 @@ jobs:
       - name: Build wheels
         uses: pypa/cibuildwheel@v2.15.0
         env:
-          CIBW_ARCHS_MACOS: universal2
+          CIBW_ARCHS_MACOS: ${{ matrix.arch }}
           CIBW_ARCHS_LINUX: x86_64 aarch64
         with:
           package-dir: .
@@ -27,7 +31,6 @@ jobs:
       - uses: actions/upload-artifact@v3
         with:
           path: ./wheelhouse/*.whl
-
 
   build_sdist:
     name: Build source distribution
