### 2023-09-08 beedb78
Removed self-hosted runner. Sticking to x86_64 for now and will do arm64 builds manually
