+++ .github/workflows/build_wheels.yml
@@ -1,6 +1,6 @@
 name: swiftest
 
-on: push
+on: [push, pull_request]
 
 jobs:
   build_wheels:
@@ -8,17 +8,25 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-20.04, macos-11, macos-12, macos-13]
+        os: [ubuntu-latest, macos-11, macos-12, macos-13]
+        cibw_archs: [x86_64]
+        include:
+          - os: ubuntu-latest
+            cibw_archs: "aarch64"
     env:
       SYSTEM_VERSION_COMPAT: 0
-    steps:
-      - uses: actions/checkout@v3
 
+    steps:
+      - name: Set up QEMU
+        if: matrix.cibw_archs == 'aarch64'
+        uses: docker/setup-qemu-action@v2
+        with:
+          platforms: arm64        
+      - uses: actions/checkout@v3
       - name: Build wheels
         uses: pypa/cibuildwheel@v2.15.0
         env:
-          CIBW_ARCHS_MACOS: x86_64
-          CIBW_ARCHS_LINUX: x86_64 aarch64
+          CIBW_ARCHS: ${{ matrix.cibw_archs }}
         with:
           package-dir: .
           output-dir: wheelhouse
