### 2023-09-11 35ae7b1
Changed kwargs clm check conditions in vec2xr. Fixed a typo in add_body
