+++ version.txt
@@ -1,1 +1,0 @@
-2023.9.0
