### 2023-09-11 ddbdf0b
Made a number of changes. Versioning is now controlled via the pyproject.toml file. The swiftest_driver executable is now packaged in the wheel. Ready for first official release!
