### 2023-09-11 331a52e
Switched to only build wheels on release to prevent wheels from being constantly built
