### 2023-09-11 #5409 @warrickball
Hello dear reviewers (@hannorein, @mikegrudic, @aatrani)! It looks like @profminton and @carlislewishard have done an enormous amount of work to make Swiftest easier to install and run, so hopefully you can try again to evaluate its functionality, testing and other remaining qualities. I just installed with `pip install swiftest --user` and the basic simulation finished running without an issue. (I did notice that though there's a file called `swiftest_driver` in the depths of `~/.local/`, it isn't in my path, but I'll check if that's a local problem before I raise an issue.)
