### 2023-09-12 cae9e38
Rearranged CMakeLists.txt files and added some lines to help deal with library resolution in the linux wheels
