+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "swiftest"
-version = "2023.9.1"
+version = "2023.9.2"
 authors=[
     {name = 'David A. Minton', email='daminton@purdue.edu'},
     {name = 'Carlisle Wishard'},
