+++ src/CMakeLists.txt
@@ -224,7 +224,6 @@ INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY}
     INCLUDES DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
 
-MESSAGE(STATUS "Setting rpaths")
 # the RPATH to be used when installing, but only if it's not a system directory
 LIST(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_INSTALL_PREFIX}/lib" isSystemDir)
 IF("${isSystemDir}" STREQUAL "-1")
@@ -237,4 +236,3 @@ IF("${isSystemDir}" STREQUAL "-1")
         INSTALL_RPATH_USE_LINK_PATH TRUE
         INSTALL_RPATH "${lib_path}")
 endif("${isSystemDir}" STREQUAL "-1")
-MESSAGE(STATUS "rpaths are set")
