+++ src/CMakeLists.txt
@@ -208,21 +208,6 @@ ELSE ()
     MESSAGE(STATUS "Quad precision real is not supported")
 ENDIF ()
 
-#####################################
-# Tell how to install this executable
-#####################################
-IF(WIN32)
-    SET(CMAKE_INSTALL_PREFIX "C:\\Program Files")
-    FILE(TO_CMAKE_PATH ${CMAKE_INSTALL_PREFIX} CMAKE_INSTALL_PREFIX)
-ELSE()
-    SET(CMAKE_INSTALL_PREFIX /usr/local)
-ENDIF(WIN32)
-INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY}
-    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
-    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
-    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
-    INCLUDES DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
-)
 
 # the RPATH to be used when installing, but only if it's not a system directory
 LIST(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_INSTALL_PREFIX}/lib" isSystemDir)
@@ -236,3 +221,10 @@ IF("${isSystemDir}" STREQUAL "-1")
         INSTALL_RPATH_USE_LINK_PATH TRUE
         INSTALL_RPATH "${lib_path}")
 endif("${isSystemDir}" STREQUAL "-1")
+
+INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY}
+    RUNTIME DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_BINDIR}
+    LIBRARY DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_LIBDIR}
+    ARCHIVE DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_LIBDIR}
+    INCLUDES DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_INCLUDEDIR}
+)
