### 2023-09-12 bb3be44
More tweaks to cmake commands
