### 2023-09-13 1c69766
Fixed install path of bindings library
