+++ swiftest/CMakeLists.txt
@@ -32,4 +32,4 @@ ENDIF()
 PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_BINDINGS} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_INSTALL_INCLUDEDIR} ${NETCDF_INCLUDE_DIR})
 
-INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION ${PY})
+INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION ${lib_path})
