### 2023-09-13 de68a73
More tweaks to build files in order to reliably get all shared libraries and executables into the right places in the wheels
