+++ src/CMakeLists.txt
@@ -137,7 +137,7 @@ SET(SWIFTEST_DRIVER swiftest_driver)
 ADD_EXECUTABLE(${SWIFTEST_DRIVER} ${DRIVER_src})
 
 # Be sure the executable gets packaged with the wheel
-SET_PROPERTY(TARGET ${SWIFTEST_DRIVER} PROPERTY RUNTIME_OUTPUT_DIRECTORY ${SKBUILD_SCRIPTS_DIR})
+#SET_PROPERTY(TARGET ${SWIFTEST_DRIVER} PROPERTY RUNTIME_OUTPUT_DIRECTORY ${SKBUILD_SCRIPTS_DIR})
 
 #####################################################
 # Add the needed libraries 
@@ -208,7 +208,6 @@ ELSE ()
     MESSAGE(STATUS "Quad precision real is not supported")
 ENDIF ()
 
-
 # the RPATH to be used when installing, but only if it's not a system directory
 LIST(FIND CMAKE_PLATFORM_IMPLICIT_LINK_DIRECTORIES "${CMAKE_INSTALL_PREFIX}/lib" isSystemDir)
 IF("${isSystemDir}" STREQUAL "-1")
@@ -222,9 +221,9 @@ IF("${isSystemDir}" STREQUAL "-1")
         INSTALL_RPATH "${lib_path}")
 endif("${isSystemDir}" STREQUAL "-1")
 
-INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY}
-    RUNTIME DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_BINDIR}
-    LIBRARY DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_LIBDIR}
-    ARCHIVE DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_LIBDIR}
-    INCLUDES DESTINATION ${CMAKE_SOURCE_DIR}/${CMAKE_INSTALL_INCLUDEDIR}
+INSTALL(TARGETS ${SWIFTEST_DRIVER} ${SWIFTEST_LIBRARY} 
+    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    INCLUDES DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
