### 2023-09-13 5980806
Simplified the cmake files for Cython and improved the ability to package the wheel
