+++ buildscripts/set_environment.sh
@@ -9,7 +9,7 @@ HOMEBREW_PREFIX="$(brew --prefix)"
 LD_LIBRARY_PATH="/usr/local/lib:${PREFIX}/lib:${HOMEBREW_PREFIX}/lib"
 LDFLAGS="-Wl,-rpath,${ROOT_DIR}/lib  -Wl,-no_compact_unwind -L${PREFIX}/lib -L${HOMEBREW_PREFIX}/lib" 
 CPATH="/usr/local/include:${PREFIX}/include:${HOMEBREW_PREFIX}/include:${ROOT_DIR}/include"
-CPPFLAGS="-isystem ${PREFIX}/include -isystem /usr/local/include"
+CPPFLAGS="-isystem ${PREFIX}/include -isystem /usr/local/include -Xclang -fopenmp"
 LIBS="-lomp"
 FCFLAGS="-mmacosx-version-min=${MACOSX_DEPLOYMENT_TARGET}"
 FFFLAGS="${FCFLAGS}"
