+++ swiftest/CMakeLists.txt
@@ -7,29 +7,25 @@
 # You should have received a copy of the GNU General Public License along with Swiftest. 
 # If not, see: https://www.gnu.org/licenses. 
 
-# Avoid Cython/Python3.8 minor incompatibility warnings, see
-# https://github.com/cython/cython/issues/3474. Note that this option is a bit
-# expansive, but it's a temporary fix and we'll be testing on other Python
-# versions concurrently so it shouldn't hide any real issues. ).
-IF(${PYTHON_VERSION_MAJOR} EQUAL 3
-   AND ${PYTHON_VERSION_MINOR} EQUAL 8
-   AND NOT WIN32)
-      ADD_COMPILE_OPTIONS("-Wno-deprecated-declarations")
-ENDIF()
-
 SET(SWIFTEST_BINDINGS _bindings)
 
-ADD_CYTHON_TARGET(${SWIFTEST_BINDINGS})
-ADD_LIBRARY(${SWIFTEST_BINDINGS} MODULE ${SWIFTEST_BINDINGS})
+FIND_PROGRAM(CYTHON "cython")
+ADD_CUSTOM_COMMAND(
+  OUTPUT "${SWIFTEST_BINDINGS}.c"
+  DEPENDS "${SWIFTEST_BINDINGS}.pyx"
+  VERBATIM
+  COMMAND "${CYTHON}" "${CMAKE_CURRENT_SOURCE_DIR}/${SWIFTEST_BINDINGS}.pyx" --output-file
+          "${CMAKE_CURRENT_BINARY_DIR}/${SWIFTEST_BINDINGS}.c")
+
+PYTHON_ADD_LIBRARY(${SWIFTEST_BINDINGS} MODULE "${CMAKE_CURRENT_BINARY_DIR}/${SWIFTEST_BINDINGS}.c" WITH_SOABI)
+
 IF (NOT BUILD_SHARED_LIBS) 
     SET_PROPERTY(TARGET ${SWIFTEST_BINDINGS} PROPERTY POSITION_INDEPENDENT_CODE)
+    TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES})
+ELSE ()
+    TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} PUBLIC ${SWIFTEST_LIBRARY})
 ENDIF ()
+TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_BINDINGS} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
-TARGET_LINK_LIBRARIES(${SWIFTEST_BINDINGS} ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES})
-IF(USE_OPENMP OR USE_SIMD)
-    SET_PROPERTY(TARGET ${SWIFTEST_BINDINGS} APPEND_STRING PROPERTY LINK_FLAGS  "${OpenMP_Fortran_FLAGS} ")
-ENDIF()
-PYTHON_EXTENSION_MODULE(${SWIFTEST_BINDINGS})
-TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_BINDINGS} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_INSTALL_INCLUDEDIR} ${NETCDF_INCLUDE_DIR})
-
-INSTALL(TARGETS ${SWIFTEST_BINDINGS} LIBRARY DESTINATION ${PYTHON_RELATIVE_SITE_PACKAGES_DIR}/${SKBUILD_PROJECT_NAME})
[…]
