### 2023-09-13 a944ce8
added preliminary c_lm parameters for netcdf
