### 2023-09-13 bf4fd8c
Finally got the rpath to be something workable
