+++ CMakeLists.txt
@@ -29,12 +29,16 @@ IF (SKBUILD)
     SET(INSTALL_BINDIR ${SKBUILD_PLATLIB_DIR}/${SKBUILD_PROJECT_NAME}) 
     SET(INSTALL_LIBDIR ${SKBUILD_PLATLIB_DIR}/${SKBUILD_PROJECT_NAME}) 
     SET(INSTALL_INCLUDEDIR ${INSTALL_LIBDIR}) 
+    IF (APPLE)
+        SET(CMAKE_INSTALL_RPATH "@loader_path")
+    ELSEIF (LINUX)
+        SET(CMAKE_INSTALL_RPATH $ORIGIN)
+    ENDIF ()
 ELSE ()
     SET(INSTALL_BINDIR ${CMAKE_INSTALL_BINDIR})
     SET(INSTALL_LIBDIR ${CMAKE_INSTALL_LIBDIR})
     SET(INSTALL_INCLUDEDIR ${CMAKE_INSTALL_INCLUDEDIR})
 ENDIF ()
-
 
 # Have the .mod files placed in the include folder
 SET(CMAKE_Fortran_MODULE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/mod)
