### 2023-09-14 f0a0c10
Extended CB class to include c_lm. Added c_lm to netcdf
