### 2023-09-15 e0dcf9b
Converted CMAKE_INSTALL_PREFIX to be a cache variable so that it gets set properly via flags
