### 2023-09-15 8adb5a2
Made a number of formatting changes to better conform to standard line lengths. Also now include the version number in the log output
