+++ src/swiftest/swiftest_io.f90
@@ -22,14 +22,16 @@ contains
       interface fmt
          !! author: David Minton
          !!
-         !! Formats a pair of variables and corresponding values for the compact display output. Generic interface for different variable types to format.
+         !! Formats a pair of variables and corresponding values for the compact display output. Generic interface for different 
+         !! variable types to format.
          procedure :: fmt_I4B, fmt_I8B, fmt_DP
       end interface
 
       ! Arguments
       class(swiftest_nbody_system), intent(in) :: self  !! Swiftest nbody system object   
       class(swiftest_parameters),   intent(in) :: param !! Input colleciton of user-defined parameters
-      class(*),                     intent(in) :: timer !! Object used for computing elapsed wall time  (must be unlimited polymorphic because the walltimer module requires base)
+      class(*),                     intent(in) :: timer !! Object used for computing elapsed wall time  (must be unlimited 
+                                                        !! polymorphic because the walltimer module requires base)
       ! Internals
       character(len=:), allocatable :: formatted_output
 
@@ -41,17 +43,20 @@ contains
             formatted_output = formatted_output // fmt("NPLM",pl%nplm)
          end select
          if (param%lenergy) then
-            formatted_output = formatted_output // fmt("LTOTERR",self%L_total_error) // fmt("ETOTERR",self%te_error) // fmt("MTOTERR",self%Mtot_error) &
-                             // fmt("KEOERR",self%ke_orbit_error) // fmt("PEERR",self%pe_error) // fmt("EORBERR",self%E_orbit_error) &
-                             // fmt("EUNTRERR",self%E_untracked_error) // fmt("LESCERR",self%L_escape_error) // fmt("MESCERR",self%Mescape_error)
+            formatted_output = formatted_output // fmt("LTOTERR",self%L_total_error) // fmt("ETOTERR",self%te_error) &
+                             // fmt("MTOTERR",self%Mtot_error) // fmt("KEOERR",self%ke_orbit_error) // fmt("PEERR",self%pe_error) &
+                             // fmt("EORBERR",self%E_orbit_error) // fmt("EUNTRERR",self%E_untracked_error) &
+                             // fmt("LESCERR",self%L_escape_error) // fmt("MESCERR",self%Mescape_error)
             if (param%lclose) formatted_output = formatted_output // fmt("ECOLLERR",self%Ecoll_error)
-            if (param%lrotation) formatted_output = formatted_output // fmt("KESPINERR",self%ke_spin_error) // fmt("LSPINERR",self%L_spin_error) 
+            if (param%lrotation) formatted_output = formatted_output // fmt("KESPINERR",self%ke_spin_error) &
+                             // fmt("LSPINERR",self%L_spin_error) 
          end if
 
          if (.not. timer%main_is_started) then ! This is the start of a new run
             formatted_output =  formatted_output // fmt("WT",0.0_DP) // fmt("IWT",0.0_DP) // fmt("WTPS",0.0_DP) 
          else
-            formatted_output = formatted_output // fmt("WT",timer%wall_main) // fmt("IWT",timer%wall_step) // fmt("WTPS",timer%wall_per_substep)
+            formatted_output = formatted_output // fmt("WT",timer%wall_main) // fmt("IWT",timer%wall_step) &
+                                               // fmt("WTPS",timer%wall_per_substep)
          end if
          write(*,*) formatted_output
       end select
@@ -123,12 +128,13 @@ contains
[…]
