### 2023-09-15 7aed38d
Added a missing close for the log file at the end of a run
