### 2023-09-15 a2369d4
Refactored test and added gr test
