+++ tests/test_swiftest.py
@@ -1,0 +1,380 @@
+#!/usr/bin/env python3
+
+"""
+ Copyright 2023 - David Minton, Carlisle Wishard, Jennifer Pouplin, Jake Elliott, & Dana Singh
+ This file is part of Swiftest.
+ Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
+ as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ Swiftest is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
+ of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ You should have received a copy of the GNU General Public License along with Swiftest. 
+ If not, see: https://www.gnu.org/licenses. 
+"""
+
+"""
+Tests that energy and momentum errors are within tolerances in a Swiftest simulation
+
+Input
+------
+
+Output
+------
+None
+"""
+
+import swiftest
+import unittest
+import os
+import shutil
+import numpy as np
+from numpy.random import default_rng
+from astroquery.jplhorizons import Horizons
+import datetime
+
+
+rng = default_rng(seed=123)
+
+major_bodies = ["Sun","Mercury","Venus","Earth","Mars","Jupiter","Saturn","Uranus","Neptune"]
+param = {}
+
+
+class TestSwiftest(unittest.TestCase):
+    
+    def test01_gen_ic(self):
+        """
+        Tests that Swiftest is able to successfully generate a set of initial conditions in a file without any exceptions being raised
+        """
+        print("\ntest_gen_ic: Test whether we can generate simulation initial conditions test")
+        # Files that are expected to be generated:
+        simdir = "simdata"
+        file_list = [simdir, os.path.join(simdir,"param.in"), os.path.join(simdir,"init_cond.nc")]
+        
+        sim = swiftest.Simulation()
+        sim.clean()
+#!/usr/bin/env python3
+
+"""
+ Copyright 2023 - David Minton, Carlisle Wishard, Jennifer Pouplin, Jake Elliott, & Dana Singh
+ This file is part of Swiftest.
+ Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
+ as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ Swiftest is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
+ of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ You should have received a copy of the GNU General Public License along with Swiftest. 
+ If not, see: https://www.gnu.org/licenses. 
+"""
+
+"""
+Tests that energy and momentum errors are within tolerances in a Swiftest simulation
+
+Input
+------
+
+Output
+------
+None
+"""
+
+import swiftest
+import unittest
+import os
+import shutil
+import numpy as np
+from numpy.random import default_rng
+
+rng = default_rng(seed=123)
+
+major_bodies = ["Sun","Mercury","Venus","Earth","Mars","Jupiter","Saturn","Uranus","Neptune"]
+param = {}
+
+
+class TestSwiftest(unittest.TestCase):
+    
+    def test01_gen_ic(self):
+        """
+        Tests that Swiftest is able to successfully generate a set of initial conditions in a file without any exceptions being raised
+        """
+        print("\ntest_gen_ic: Test whether we can generate simulation initial conditions test")
[…]
