### 2023-09-17 9ea4b06
Added nf90 inquire dimensions for read_frame
