### 2023-09-18 6889de2
Added C flags to the compiler flags so that the cython code can get debugging flags
