+++ cmake/Modules/SetSwiftestFlags.cmake
@@ -46,14 +46,13 @@ ELSE()
     MESSAGE(FATAL_ERROR "CMAKE_BUILD_TYPE not valid! ${BUILD_TYPE_MSG}")
 ENDIF(BT STREQUAL "RELEASE")
 
-
 #########################################################
 # If the compiler flags have already been set, return now
 #########################################################
 
 IF(CMAKE_Fortran_FLAGS_RELEASE AND CMAKE_Fortran_FLAGS_TESTING AND CMAKE_Fortran_FLAGS_DEBUG AND CMAKE_Fortran_FLAGS_PROFILE )
     RETURN ()
-ENDIF(CMAKE_Fortran_FLAGS_RELEASE AND CMAKE_Fortran_FLAGS_TESTING AND CMAKE_Fortran_FLAGS_DEBUG AND CMAKE_Fortran_FLAGS_PROFILE)
+ENDIF()
 
 ########################################################################
 # Determine the appropriate flags for this compiler for each build type.
@@ -72,13 +71,11 @@ ENDIF ()
 ### GENERAL FLAGS ###
 #####################
 
-
 # Free form
 IF (COMPILER_OPTIONS STREQUAL "GNU")
     SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
         Fortran "-ffree-form" # GNU
         ) 
-
     # Don't add underscores in symbols for C-compatability
     SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
         Fortran "-fno-underscoring" # GNU
@@ -88,12 +85,10 @@ IF (COMPILER_OPTIONS STREQUAL "GNU")
     SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
         Fortran "-fsignaling-nans " # GNU
         ) 
-    
     # Allows for lines longer than 80 characters without truncation
     SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
         Fortran "-ffree-line-length-512" # GNU (gfortran)
         )
-
     # Sets the dialect standard
     SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
         Fortran "-std=f2018" 
@@ -131,7 +126,6 @@ IF (NOT BUILD_SHARED_LIBS AND NOT WINOPT)
     SET_COMPILE_FLAG(CMAKE_FORTRAN_FLAGS "${CMAKE_FORTRAN_FLAGS}"
         Fortran "-fPIC"
         )
-
     IF (COMPILER_OPTIONS STREQUAL "Intel")
         # Use static Intel libraries
         SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
@@ -141,27 +135,27 @@ IF (NOT BUILD_SHARED_LIBS AND NOT WINOPT)
         SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
             Fortran "-static_mpi"  # Intel
         )
-
         IF (USE_OPENMP)
             SET_COMPILE_FLAG(CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS}"
                 Fortran "-qopenmp-link=static"  # Intel
             )
         ENDIF (USE_OPENMP)
-
     ELSEIF (COMPILER_OPTIONS STREQUAL "GNU") 
[…]
