+++ buildscripts/install_editable_debug.sh
@@ -3,13 +3,16 @@
 set -a
 SCRIPT_DIR=$(realpath $(dirname $0))
 ROOT_DIR=$(realpath ${SCRIPT_DIR}/..)
+VENV_DIR=${ROOT_DIR}/env
 cd ${ROOT_DIR}
-python3 -m pip install --upgrade pip --user
-pip install scikit-build-core pyproject-metadata pathspec ninja --user
+python3 -m venv ${VENV_DIR}
+. ${VENV_DIR}/bin/activate
+python3 -m pip install --upgrade pip 
+pip install scikit-build-core pyproject-metadata pathspec ninja 
 pip install --config-settings=editable.rebuild=true \
             --config-settings=build-dir="build/{wheel_tag}" \
             --config-settings=cmake.build-type="Debug" \
             --config-settings=cmake.args="-DUSE_SIMD=ON" \
             --config-settings=cmake.args="-DUSE_OPENMP=ON" \
             --no-build-isolation \
-            -ve . --user
+            -ve . 
