+++ buildscripts/install_editable_debug.sh
@@ -8,7 +8,7 @@ cd ${ROOT_DIR}
 python3 -m venv ${VENV_DIR}
 . ${VENV_DIR}/bin/activate
 python3 -m pip install --upgrade pip 
-pip install scikit-build-core pyproject-metadata pathspec ninja 
+pip install scikit-build-core pyproject-metadata pathspec ninja cython
 pip install --config-settings=editable.rebuild=true \
             --config-settings=build-dir="build/{wheel_tag}" \
             --config-settings=cmake.build-type="Debug" \
