+++ src/swiftest/swiftest_io.f90
@@ -982,6 +982,12 @@ contains
          call netcdf_io_check( nf90_def_var(nc%id, nc%j4rp4_varname, nc%out_type, nc%time_dimid, nc%j4rp4_varid), &
                                   "netcdf_io_initialize_output nf90_def_var j4rp4_varid"  )
 
+         status = nf90_inq_varid(nc%id, nc%c_lm_varname, nc%c_lm_varid)
+         if (status == NF90_NOERR) then
+            call netcdf_io_check( nf90_def_var(nc%id, nc%c_lm_varname, nc%out_type, [2, nc%l_dimid, nc%m_dimid], nc%c_lm_varid), & 
+                                    "netcdf_io_initialize_output nf90_def_var c_lm_varid" )
+         end if
+
          ! Set fill mode to NaN for all variables
          call netcdf_io_check( nf90_inquire(nc%id, nVariables=nvar), "netcdf_io_initialize_output nf90_inquire nVariables" )
          do varid = 1, nvar
@@ -1072,6 +1078,14 @@ contains
          call netcdf_io_check( nf90_inq_dimid(nc%id, nc%str_dimname, nc%str_dimid), &
                                   "swiftest_io_netcdf_open nf90_inq_dimid str_dimid"  )
 
+         status = nf90_inq_varid(nc%id, nc%c_lm_varname, nc%c_lm_varid)
+         if (status == NF90_NOERR) then
+            call netcdf_io_check( nf90_inq_dimid(nc%id, nc%l_dimname, cs%l_dimid), &
+                                    "swiftest_io_netcdf_open nf90_inq_dimid l_dimid")
+            call netcdf_io_check( nf90_inq_dimid(nc%id, nc%m_dimname, cs%m_dimid), &
+                                    "swiftest_io_netcdf_open nf90_inq_dimid m_dimid")
+         end if
+
          ! Dimension coordinates
          call netcdf_io_check( nf90_inq_varid(nc%id, nc%time_dimname, nc%time_varid), &
                                   "swiftest_io_netcdf_open nf90_inq_varid time_varid" )
@@ -1079,6 +1093,13 @@ contains
                                   "swiftest_io_netcdf_open nf90_inq_varid space_varid" )
          call netcdf_io_check( nf90_inq_varid(nc%id, nc%name_dimname, nc%name_varid), &
                                   "swiftest_io_netcdf_open nf90_inq_varid name_varid" )
+         status = nf90_inq_varid(nc%id, nc%c_lm_varname, nc%c_lm_varid)
+         if (status == NF90_NOERR) then
+            call netcdf_io_check( nf90_inq_dimid(nc%id, nc%l_dimname, cs%l_varid), &
+                                    "swiftest_io_netcdf_open nf90_inq_dimid l_varid")
+            call netcdf_io_check( nf90_inq_dimid(nc%id, nc%m_dimname, cs%m_varid), &
+                                    "swiftest_io_netcdf_open nf90_inq_dimid m_varid")
+         end if
 
          ! Required Variables
          call netcdf_io_check( nf90_inq_varid(nc%id, nc%id_varname, nc%id_varid), &
@@ -1180,6 +1201,8 @@ contains
             status = nf90_inq_varid(nc%id, nc%E_untracked_varname, nc%E_untracked_varid)
             status = nf90_inq_varid(nc%id, nc%GMescape_varname, nc%GMescape_varid)
          end if
+
+         status = nf90_inq_varid(nc%id, nc%c_lm_varname, nc%c_lm_varid)
 
       end associate
 
@@ -2019,6 +2042,7 @@ contains
[…]
