### 2023-09-18 c7d4942
Added debug arguments to cython call when in debug build type
