+++ swiftest/CMakeLists.txt
@@ -18,12 +18,16 @@ FIND_PROGRAM(CYTHON
             NO_CMAKE_FIND_ROOT_PATH
             )
 MESSAGE(STATUS "Cython executable path: ${CYTHON}")
+SET(CYTHON_ARGS "${CMAKE_CURRENT_SOURCE_DIR}/${SWIFTEST_BINDINGS}.pyx" "--output-file" "${CMAKE_CURRENT_BINARY_DIR}/${SWIFTEST_BINDINGS}.c")
+STRING(TOUPPER "${CMAKE_BUILD_TYPE}" BT)
+IF (BT STREQUAL "DEBUG")
+    LIST(APPEND CYTHON_ARGS "--gdb")
+endif ()
 ADD_CUSTOM_COMMAND(
   OUTPUT "${SWIFTEST_BINDINGS}.c"
   DEPENDS "${SWIFTEST_BINDINGS}.pyx"
   VERBATIM
-  COMMAND "${CYTHON}" "${CMAKE_CURRENT_SOURCE_DIR}/${SWIFTEST_BINDINGS}.pyx" --output-file
-          "${CMAKE_CURRENT_BINARY_DIR}/${SWIFTEST_BINDINGS}.c")
+  COMMAND "${CYTHON}" ${CYTHON_ARGS} )
 
 PYTHON_ADD_LIBRARY(${SWIFTEST_BINDINGS} MODULE "${CMAKE_CURRENT_BINARY_DIR}/${SWIFTEST_BINDINGS}.c" WITH_SOABI)
 
