### 2023-09-18 c0228df
fixed allocation error for c_lm
