+++ src/CMakeLists.txt
@@ -21,6 +21,8 @@ INCLUDE(SetUpNetCDF)
 IF (COMPILER_OPTIONS STREQUAL "Intel" AND NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
     INCLUDE(SetMKL)
 ENDIF ()
+FIND_PACKAGE(SHTOOLS REQUIRED)
+FIND_PACKAGE(FFTW3 REQUIRED)
 
 # Add the source files
 SET(STRICT_MATH_FILES
@@ -141,14 +143,20 @@ IF (NOT BUILD_SHARED_LIBS)
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} PROPERTY POSITION_INDEPENDENT_CODE)
 ENDIF ()
 
-TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_INCLUDE_DIR})
-TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR})
-TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES})
-TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES})
+TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_INCLUDE_DIR} ${SHTOOLS_INCLUDE_DIR} ${FFTW3_INCLUDE_DIR})
+TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR} ${SHTOOLS_INCLUDE_DIR} ${FFTW3_INCLUDE_DIR})
+TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES} ${FFTW3_LIBRARY})
+TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES} ${FFTW3_LIBRARY})
 
 IF(USE_OPENMP OR USE_SIMD)
+    TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${SHTOOLS_LIBRARY_MP})
+    TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SHTOOLS_LIBRARY_MP})
+
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY COMPILE_FLAGS  "${OpenMP_Fortran_FLAGS} ")
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} ${SWIFTEST_DRIVER} APPEND_STRING PROPERTY LINK_FLAGS  "${OpenMP_Fortran_FLAGS} ")
+ELSE ()
+    TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${SHTOOLS_LIBRARY})
+    TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SHTOOLS_LIBRARY})
 ENDIF()
 
 IF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
