### 2023-09-18 89a3f68
Got rid of dedicated FFTW finder
