+++ src/CMakeLists.txt
@@ -22,7 +22,6 @@ IF (COMPILER_OPTIONS STREQUAL "Intel" AND NOT CMAKE_SYSTEM_NAME STREQUAL "Window
     INCLUDE(SetMKL)
 ENDIF ()
 FIND_PACKAGE(SHTOOLS REQUIRED)
-FIND_PACKAGE(FFTW3 REQUIRED)
 
 # Add the source files
 SET(STRICT_MATH_FILES
@@ -145,8 +144,8 @@ ENDIF ()
 
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_INCLUDE_DIR} ${SHTOOLS_INCLUDE_DIR} ${FFTW3_INCLUDE_DIR})
 TARGET_INCLUDE_DIRECTORIES(${SWIFTEST_DRIVER} PUBLIC ${NETCDF_INCLUDE_DIR} ${SHTOOLS_INCLUDE_DIR} ${FFTW3_INCLUDE_DIR})
-TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES} ${FFTW3_LIBRARY})
-TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES} ${FFTW3_LIBRARY})
+TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${NETCDF_LIBRARIES})
+TARGET_LINK_LIBRARIES(${SWIFTEST_DRIVER} PUBLIC ${SWIFTEST_LIBRARY} ${NETCDF_LIBRARIES})
 
 IF(USE_OPENMP OR USE_SIMD)
     TARGET_LINK_LIBRARIES(${SWIFTEST_LIBRARY} PUBLIC ${SHTOOLS_LIBRARY_MP})
