### 2023-09-19 2187347
Added SHTOOLS to the CI dependency build scripts
