### 2023-09-19 b6e85fc
Cleaned up build scripts and made the cmake script outputs more useful and consistent
