### 2023-09-20 2446448
Updated cmake module for Windows
