### 2023-09-20 6cfa241
Updated compiler flag checks. Now use built in check_fortran_compiler_flag function and fixed Windows options
