+++ src/CMakeLists.txt
@@ -12,16 +12,6 @@ set(GLOBAL_MODULE_IN ${SRC}/globals/globals_module.f90.in)
 set(GLOBAL_MODULE_OUT ${SRC}/globals/globals_module.f90)
 CONFIGURE_FILE(${GLOBAL_MODULE_IN} ${GLOBAL_MODULE_OUT})
 
-IF(NOT CMAKE_Fortran_COMPILER_SUPPORTS_F90)
-    MESSAGE(FATAL_ERROR "Fortran compiler does not support F90")
-ENDIF(NOT CMAKE_Fortran_COMPILER_SUPPORTS_F90)
-
-INCLUDE(SetParallelizationLibrary)
-INCLUDE(SetUpNetCDF)
-IF (COMPILER_OPTIONS STREQUAL "Intel" AND NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
-    INCLUDE(SetMKL)
-ENDIF ()
-
 # Add the source files
 SET(STRICT_MATH_FILES
             ${SRC}/collision/collision_generate.f90
@@ -136,6 +126,7 @@ ADD_EXECUTABLE(${SWIFTEST_DRIVER} ${DRIVER_src})
 # Add the needed libraries 
 #####################################################
 # Create a library from the source files, except the driver
+INCLUDE(SetUpNetCDF)
 ADD_LIBRARY(${SWIFTEST_LIBRARY} ${SWIFTEST_src})
 IF (NOT BUILD_SHARED_LIBS) 
     SET_PROPERTY(TARGET ${SWIFTEST_LIBRARY} PROPERTY POSITION_INDEPENDENT_CODE)
