+++ src/swiftest/swiftest_io.f90
@@ -2077,11 +2077,11 @@ contains
 
          status = nf90_inq_varid(nc%id, nc%c_lm_varname, nc%c_lm_varid)
          if (status == NF90_NOERR) then
-            call netcdf_io_check( nf90_inquire_dimension(nc%id, nc%l_dimid, len = l_dim_max), "netcdf_io_read_frame_system nf90_inquire_dimension l_dimid"  )
-            call netcdf_io_check( nf90_inquire_dimension(nc%id, nc%m_dimid, len = m_dim_max), "netcdf_io_read_frame_system nf90_inquire_dimension m_dimid")
+            call netcdf_io_check( nf90_inquire_dimension(nc%id, nc%l_dimid, len = l_dim_max), "netcdf_io_write_frame_cb nf90_inquire_dimension l_dimid")
+            call netcdf_io_check( nf90_inquire_dimension(nc%id, nc%m_dimid, len = m_dim_max), "netcdf_io_write_frame_cb nf90_inquire_dimension m_dimid")
             
             allocate(self%c_lm(2, l_dim_max, m_dim_max))
-            call netcdf_io_check( nf90_put_var(nc%id, nc%c_lm_varid, self%c_lm, count = [2, l_dim_max, m_dim_max]), "netcdf_io_read_frame_system nf90_getvar c_lm_varid")
+            call netcdf_io_check( nf90_put_var(nc%id, nc%c_lm_varid, self%c_lm, count = [2, l_dim_max, m_dim_max]), "netcdf_io_write_frame_cb nf90_getvar c_lm_varid")
          end if
 
          call netcdf_io_check( nf90_set_fill(nc%id, old_mode, tmp), &
