### 2023-09-20 4e4b156
Added accel_sph (spherical harmonicsc acceleration) to whm_module for pl and tps
