+++ src/whm/whm_kick.f90
@@ -59,7 +59,8 @@ contains
          end if
 
          if(param%lshgrav) then
-            call pl%swiftest_sph_g_acc_pl_all(nbody_system)
+            call pl%accel_sph(nbody_system)
+         end if
 
          if (param%lgr) call pl%accel_gr(param) 
 
@@ -117,7 +118,7 @@ contains
          end if
 
          if (param%loblatecb) call tp%accel_obl(nbody_system)
-         if (param%lshgrav) call tp%swiftest_sph_g_acc_tp_all(nbody_system)
+         if (param%lshgrav) call tp%accel_sph(nbody_system)
          if (param%lextra_force) call tp%accel_user(nbody_system, param, t, lbeg)
          if (param%lgr) call tp%accel_gr(param) 
       end associate
