### 2023-09-20 2c62fed
Moved accel_sph to swiftest_module from whm_module
