### 2023-09-20 0cb455f
Made more changes to get the NetCDF-Fortran library finding to work without too much user intervention
