+++ cmake/Modules/FindNETCDF.cmake
@@ -8,9 +8,48 @@
 # If not, see: https://www.gnu.org/licenses. 
 
 # - Finds the NetCDF libraries 
+# Begin searches with "typical" install locations of dependent libraries. These can be overrided in the cache or supplemented
+# with environment variables
+IF (CMAKE_SYSTEM_NAME STREQUAL "Linux")
+   SET(NFPREFIX_DIR "/usr/local" CACHE PATH "Location of provided NetCDF-Fortran dependencies")
+   SET(NFINCLUDE_DIR "/usr/local/include" CACHE PATH "Location of provided netcdf.mod")
+   IF (NOT BUILD_SHARED_LIBS)
+      SET(NCPREFIX_DIR "/usr/local" CACHE PATH "Location of provided NetCDF-C dependencies")
+      SET(H5PREFIX_DIR "/usr/local" CACHE PATH "Location of provided HDF5 dependencies")
+      SET(ZPREFIX_DIR  "/usr/local" CACHE PATH "Location of provided zlib dependencies")
+   ENDIF ()
+ELSEIF (CMAKE_SYSTEM_NAME STREQUAL "Darwin")
+   IF (DEFINED ENV{HOMEBREW_PREFIX})
+      SET(LIBPREFIX "$ENV{HOMEBREW_PREFIX}")
+   ELSE ()
+      SET(LIBPREFIX "/usr/local")
+   ENDIF()
 
-SET(NFPREFIX_DIR ${CMAKE_SOURCE_DIR} CACHE PATH "Location of provided NetCDF-Fortran dependencies")
-SET(NFINCLUDE_DIR "${NFPREFIX_DIR}/include" CACHE PATH "Location of provided netcdf.mod")
+   SET(NFPREFIX_DIR "${LIBPREFIX}" CACHE PATH "Location of provided NetCDF-Fortran dependencies")
+   SET(NFINCLUDE_DIR "${LIBPREFIX}/include" CACHE PATH "Location of provided netcdf.mod")
+   IF (NOT BUILD_SHARED_LIBS)
+      SET(NCPREFIX_DIR "${LIBPREFIX}" CACHE PATH "Location of provided NetCDF-C dependencies")
+      SET(H5PREFIX_DIR "${LIBPREFIX}" CACHE PATH "Location of provided HDF5 dependencies")
+      SET(ZPREFIX_DIR  "${LIBPREFIX}" CACHE PATH "Location of provided zlib dependencies")
+   ENDIF ()
+ELSEIF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
+   FILE(GLOB LIBDIRS "C:/Program Files*/NC4F")
+   LIST(SORT LIBDIRS)
+   LIST(GET LIBDIRS -1 LIBPREFIX)
+   SET(NFPREFIX_DIR "${LIBPREFIX}" CACHE PATH "Location of provided NetCDF-Fortran dependencies")
+   SET(NFINCLUDE_DIR "${LIBPREFIX}/include" CACHE PATH "Location of provided netcdf.mod")
+   IF (NOT BUILD_SHARED_LIBS)
+      # Assumes that the dependency libraries are packaged with NetCDF-C.
+      FILE(GLOB LIBDIRS "C:/Program Files*/netCDF*")
+      LIST(SORT LIBDIRS)
+      LIST(GET LIBDIRS -1 LIBPREFIX)
[…]
