+++ cmake/Modules/FindNETCDF.cmake
@@ -48,7 +48,6 @@ ELSEIF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
       SET(ZPREFIX_DIR  "${LIBPREFIX}" CACHE PATH "Location of provided zlib dependencies")
    ENDIF ()
 ENDIF ()
-MESSAGE(STATUS "NFPREFIX_DIR: ${NFPREFIX_DIR}")
 
 IF(NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
    FIND_PATH(NFBIN
