### 2023-09-20 9274cec
debugging pt 3; changed cmakelists.txt
