### 2023-09-20 9a628b5
Rearranged cmake files
