+++ cmake/Modules/FindNETCDF_Fortran.cmake
@@ -50,7 +50,7 @@ ELSEIF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
 ENDIF ()
 
 IF(NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
-   FIND_PATH(NFBIN
+   FIND_FILE(NFBIN
    NAMES nf-config
    HINTS 
       NFPREFIX_DIR
@@ -61,7 +61,7 @@ IF(NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
    )
 
    IF (NFBIN)
-      SET(CMD "${NFBIN}/nf-config")
+      SET(CMD "${NFBIN}")
       LIST(APPEND CMD "--includedir")
       MESSAGE(STATUS "Searching for NetCDF-Fortran include directory using ${CMD}")
       EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFINCLUDE_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
@@ -74,7 +74,7 @@ IF(NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
          MESSAGE(STATUS "ERROR : ${ERR}")
       ENDIF ()
 
-      SET(CMD "${NFBIN}/nf-config")
+      SET(CMD "${NFBIN}")
       LIST(APPEND CMD "--prefix")
       MESSAGE(STATUS "Searching for NetCDF-Fortran library directory using ${CMD}")
       EXECUTE_PROCESS(COMMAND ${CMD} OUTPUT_VARIABLE NFPREFIX_DIR ERROR_VARIABLE ERR RESULT_VARIABLE RES OUTPUT_STRIP_TRAILING_WHITESPACE)
@@ -89,7 +89,7 @@ IF(NOT CMAKE_SYSTEM_NAME STREQUAL "Windows")
    ENDIF()
 ENDIF()
 
-FIND_PATH(NETCDF_INCLUDE_DIR 
+FIND_PATH(NETCDF_FORTRAN_INCLUDE_DIR 
    NAMES netcdf.mod 
    HINTS 
       ${NFINCLUDE_DIR}
@@ -121,7 +121,7 @@ ELSE ()
    ENDIF()
 ENDIF()
 
-FIND_LIBRARY(NETCDF_FORTRAN_LIBRARY 
+FIND_LIBRARY(NFLIB 
    NAMES ${NETCDFF} 
    PATHS
       ${NFPREFIX_DIR}
@@ -133,105 +133,33 @@ FIND_LIBRARY(NETCDF_FORTRAN_LIBRARY
       ${CMAKE_LIBRARY_ARCHITECTURE} 
    REQUIRED
 )
-
-IF (BUILD_SHARED_LIBS)
-   SET(NETCDF_LIBRARIES ${NETCDF_FORTRAN_LIBRARY} CACHE STRING "NetCDF Fortran library")
-ELSE ()
-   FIND_LIBRARY(NETCDF_LIBRARY 
-      NAMES ${NETCDF} 
-      HINTS
-         ${NCPREFIX_DIR}
-         ENV NETCDF_HOME
-         ENV LD_LIBRARY_PATH
+ADD_LIBRARY(NETCDF_FORTRAN_LIBRARY SHARED IMPORTED)
+IF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
+   # Get the DLL added in
+   FIND_FILE(NFDLL
+      NAMES "netcdff.dll"
+      HINTS 
+         NFPREFIX_DIR
+         ENV NETCDF_FORTRAN_HOME
+         ENV PATH
[…]
