+++ CMakeLists.txt
@@ -24,6 +24,12 @@ IF (${CMAKE_VERSION} VERSION_GREATER_EQUAL "3.27")
     CMAKE_POLICY(SET CMP0148 OLD)
 ENDIF ()
 
+# Set some options the user may choose
+OPTION(USE_COARRAY "Use Coarray Fortran for parallelization of test particles" OFF)
+OPTION(USE_OPENMP "Use OpenMP for parallelization" ON)
+OPTION(USE_SIMD "Use SIMD vectorization" ON)
+OPTION(BUILD_SHARED_LIBS "Build using shared libraries" ON)
+
 INCLUDE(GNUInstallDirs)
 IF (SKBUILD)
     SET(INSTALL_BINDIR ${SKBUILD_PLATLIB_DIR}/${SKBUILD_PROJECT_NAME}) 
@@ -55,16 +61,12 @@ ELSE ()
     MESSAGE(FATAL_ERROR "Compiler ${CMAKE_Fortran_COMPILER_ID} not recognized!") 
 ENDIF ()
 
-# Set some options the user may choose
-OPTION(USE_COARRAY "Use Coarray Fortran for parallelization of test particles" OFF)
-OPTION(USE_OPENMP "Use OpenMP for parallelization" ON)
-OPTION(USE_SIMD "Use SIMD vectorization" ON)
-OPTION(BUILD_SHARED_LIBS "Build using shared libraries" ON)
-
 # The following section is modified from Numpy f2py documentation
 IF(PROJECT_SOURCE_DIR STREQUAL PROJECT_BINARY_DIR)
   MESSAGE(FATAL_ERROR "In-source builds not allowed. Please make a new directory (called a build directory) and run CMake from there.\n")
 ENDIF()
+
+FIND_PACKAGE(NETCDF_FORTRAN REQUIRED)
 
 # Ensure scikit-build modules
 FIND_PACKAGE(Python COMPONENTS Interpreter Development.Module REQUIRED)
@@ -76,13 +78,14 @@ SET(PY "${CMAKE_SOURCE_DIR}/swiftest")
 #####################################
 # Tell how to install this executable
 #####################################
-IF(WIN32)
+IF(CMAKE_SYSTEM_NAME STREQUAL "Windows")
     SET(CMAKE_INSTALL_PREFIX "C:\\Program Files\\swiftest")
     FILE(TO_CMAKE_PATH ${CMAKE_INSTALL_PREFIX} CMAKE_INSTALL_PREFIX)
     SET(CMAKE_INSTALL_PREFIX ${CMAKE_INSTALL_PREFIX} CACHE PATH "Path for install")
 ELSE()
     SET(CMAKE_INSTALL_PREFIX /usr/local CACHE PATH "Path for install")
-ENDIF(WIN32)
+ENDIF()
+
 
 # Make sure paths are correct for Unix or Windows style
 FILE(TO_CMAKE_PATH ${SRC} SRC)
