### 2023-09-21 4611302
More tweaks to cmake files to try to get cross-platform consistency
