+++ cmake/Modules/FindNETCDF_Fortran.cmake
@@ -121,7 +121,7 @@ ELSE ()
    ENDIF()
 ENDIF()
 
-FIND_LIBRARY(NFLIB 
+FIND_LIBRARY(NETCDF_FORTRAN_LIBRARY
    NAMES ${NETCDFF} 
    PATHS
       ${NFPREFIX_DIR}
@@ -133,7 +133,7 @@ FIND_LIBRARY(NFLIB
       ${CMAKE_LIBRARY_ARCHITECTURE} 
    REQUIRED
 )
-ADD_LIBRARY(NETCDF_FORTRAN_LIBRARY SHARED IMPORTED)
+ADD_LIBRARY(NETCDF_FORTRAN_LIBRARY UNKNOWN IMPORTED PUBLIC)
 IF (CMAKE_SYSTEM_NAME STREQUAL "Windows")
    # Get the DLL added in
    FIND_FILE(NFDLL
@@ -158,7 +158,7 @@ ELSE ()
                         )
 ENDIF()
  
-MESSAGE(STATUS "NetCDF-Fortran library: ${NFLIB}")
+MESSAGE(STATUS "NetCDF-Fortran library: ${NETCDF_FORTRAN_LIBRARY}")
 MESSAGE(STATUS "NetCDF-Fortran include directory: ${NETCDF_FORTRAN_INCLUDE_DIR}")
 
 SET(NETCDF_FORTRAN_FOUND TRUE)
