### 2023-09-21 d43cab7
Started to simplify the NetCDF library methods using cmake packages instead of hand-coded library finders
