+++ src/swiftest/swiftest_sph.f90
@@ -29,41 +29,45 @@ contains
         real(DP), intent(in)        :: r_0                         !! radius of the central body
         real(DP), intent(in)        :: phi                         !! Azimuthal/Phase angle (radians)
         real(DP), intent(in)        :: theta                       !! Inclination/Zenith angle (radians)
-        real(DP), intent(in), dimension(:)       :: rh          !! distance vector of body
+        real(DP), intent(in), dimension(:)          :: rh          !! distance vector of body
         real(DP), intent(in), dimension(:, :, :)    :: c_lm        !! Spherical Harmonic coefficients
-        real(DP), intent(out), dimension(:)      :: g_sph       !! acceleration vector
+        real(DP), intent(out), dimension(:)         :: g_sph       !! acceleration vector
         real(DP), dimension(:),   intent(in),  optional :: GMpl    !! Masses of input bodies if they are not test particles
-        real(DP), dimension(:),   intent(inout), optional :: aoblcb  !! Barycentric acceleration of central body (only needed if input bodies are massive)
+        real(DP), dimension(:),   intent(inout), optional :: aoblcb  !! Barycentric acceleration of central body (only for massive input bodies)
      
         ! Internals
         integer        :: l, m              !! SPH coefficients
-        real(DP)       :: r_mag             !! magnitude of rh
-        real(DP), dimension(:), allocatable  :: p, p_deriv             !! Associated Lengendre Polynomials at a given cos(theta)
         integer        :: l_max             !! max Spherical Harmonic l order value
+        integer(I4B)   :: N, lmindex        !! Length of Legendre polynomials and index at a given l, m
+        real(DP)       :: r_mag             !! magnitude of rh
         real(DP)       :: plm, dplm         !! Associated Legendre polynomials at a given l, m
-        integer(I4B)   :: N
+        real(DP)       :: ccss, cssc        !! See definition in source code
+        real(DP), dimension(:), allocatable  :: p, p_deriv   !! Associated Lengendre Polynomials at a given cos(theta)
 
         g_sph(:) = 0.0_DP
         r_mag = sqrt(dot_product(rh(:), rh(:)))
         l_max = size(c_lm, 2) - 1
         N = (l_max + 1) * (l_max + 2) / 2
         allocate(p(N),p_deriv(N))
-        PlmON_d1(p, p_deriv, l_max, cos(theta))      ! Orthonormalized Associated Legendre Polynomials and the 1st Derivative
+        call PlmON_d1(p, p_deriv, l_max, cos(theta))      ! Orthonormalized Associated Legendre Polynomials and the 1st Derivative
 
         do l = 1, l_max
             do m = 0, l
 
-                ! Associated Legendre Polynomials   
-                plm = p(PlmIndex(l, m))         ! p_l,m
-                dplm = p_deriv(PlmIndex(l, m))       ! d(p_l,m)
+                ! Associated Legendre Polynomials 
+                lmindex = PlmIndex(l, m)  
+                plm = p(lmindex)              ! p_l,m
+                dplm = p_deriv(lmindex)       ! d(p_l,m)
 
                 ! C_lm and S_lm with Cos and Sin of m * phi
-                ccss = c_lm(1, l+1, m+1) * cos(m * phi) + c_lm(2, l+1, m+1) * sin(m * phi)          ! C_lm * cos(m * phi) + S_lm * sin(m * phi)
[…]
