### 2023-09-21 e3153bf
debugging pt 5. In swiftest_sph.f90
