### 2023-09-21 e7b86b4
Made significant changes to build scripts. Now use CMake to build dependencies for speed
