+++ buildscripts/_build_getopts.sh
@@ -13,9 +13,14 @@ set -a
 SCRIPT_DIR=$(realpath $(dirname $0))
 ROOT_DIR=$(realpath ${SCRIPT_DIR}/..)
 
+# Get platform and architecture
+read -r OS ARCH < <($SCRIPT_DIR/get_platform.sh)
+
 # Parse arguments
 USTMT="Usage: ${0} [-d /path/to/dependency/source] [-p /prefix/path] [-m MACOSX_DEPLOYMENT_TARGET]"
-MACOSX_DEPLOYMENT_TARGET=${MACOSX_DEPLOYMENT_TARGET:-"$(sw_vers --ProductVersion)"}
+if [ $OS = "MacOSX" ]; then
+    MACOSX_DEPLOYMENT_TARGET=${MACOSX_DEPLOYMENT_TARGET:-"$(sw_vers --ProductVersion)"}
+fi
 
 while getopts ":d:p:m:h" ARG; do
     case "${ARG}" in
@@ -44,7 +49,6 @@ while getopts ":d:p:m:h" ARG; do
     esac
 done
 
-read -r OS ARCH < <($SCRIPT_DIR/get_platform.sh)
 BUILD_DIR=${BUILD_DIR:-$(mktemp -ut swiftest_build.XXXXXXXX)}
 PREFIX=${PREFIX:-${ROOT_DIR}}
 DEPENDENCY_DIR=${DEPENDENCY_DIR:-${BUILD_DIR}}
