### 2023-09-21 00a4985
Fixed the prefix path in the build scripts
