+++ buildscripts/build_hdf5.sh
@@ -64,12 +64,12 @@ printf "*********************************************************\n"
 
 cd ${DEPENDENCY_DIR}/hdfsrc
 ZLIB_ROOT=${PREFIX}
-cmake -B build -S . -G Ninja 
+cmake -B build -S . -G Ninja  -DCMAKE_INSTALL_PREFIX=${PREFIX}
 cmake --build build -j${NPROC}
 if [ -w ${PREFIX} ]; then
-    cmake --install build --prefix ${PREFIX}
+    cmake --install build 
 else
-    sudo cmake --install build --prefix ${PREFIX}
+    sudo cmake --install build 
 fi
 
 if [ $? -ne 0 ]; then
