+++ buildscripts/build_netcdf-c.sh
@@ -53,12 +53,12 @@ printf "HDF5_ROOT: ${HDF5_ROOT}\n"
 printf "*********************************************************\n"
 
 cd ${DEPENDENCY_DIR}/netcdf-c-*
-cmake -B build -S . -G Ninja 
+cmake -B build -S . -G Ninja -DCMAKE_INSTALL_PREFIX=${PREFIX}
 cmake --build build -j${NPROC}
 if [ -w ${PREFIX} ]; then
-    cmake --install build --prefix ${PREFIX}
+    cmake --install build
 else
-    sudo cmake --install build --prefix ${PREFIX}
+    sudo cmake --install build
 fi
 
 if [ $? -ne 0 ]; then
