### 2023-09-21 58b7997
Added Ninja to the linux build
