+++ buildscripts/build_dependencies.sh
@@ -19,6 +19,25 @@ ARGS=$@
 NPROC=$(nproc)
 
 cd $ROOT_DIR
+NINJA_VER="1.11.1"
+
+printf "*********************************************************\n"
+printf "*             FETCHING NINJA SOURCE                      *\n"
+printf "*********************************************************\n"
+printf "Copying files to ${DEPENDENCY_DIR}\n"
+mkdir -p ${DEPENDENCY_DIR}
+if [ ! -d ${DEPENDENCY_DIR}/ninja-${NINJA_VER} ]; then
+    [ -d ${DEPENDENCY_DIR}/zlib-* ] && rm -rf ${DEPENDENCY_DIR}/zlib-*
+    curl -L https://github.com/ninja-build/ninja/archive/refs/tags/v${NINJA_VER}.tar.gz | tar xvz -C ${DEPENDENCY_DIR}
+fi
+cd ${DEPENDENCY_DIR}/ninja-*
+cmake -B build -S . -DCMAKE_INSTALL_PREFIX=${PREFIX}
+cmake --build build 
+if [ -w ${PREFIX} ]; then
+    cmake --install build 
+else
+    sudo cmake --install build
+fi
 
 set -e
 ${SCRIPT_DIR}/build_zlib.sh ${ARGS}
