+++ pyproject.toml
@@ -98,12 +98,14 @@ netCDF-Fortran_DIR="${PREFIX}/lib/cmake/netCDF"
 [tool.cibuildwheel.macos]
 before-all = [
     "brew install coreutils",
-    "LIBS=\"\" buildscripts/build_dependencies.sh -p ${PREFIX} -d ${PREFIX}/build -m ${MACOSX_DEPLOYMENT_TARGET}"
+    "LIBS=\"\" buildscripts/build_dependencies.sh -p ${TMPDIR}/build -d ${HOME}/Downloads -m ${MACOSX_DEPLOYMENT_TARGET}"
 ]
 
 [tool.cibuildwheel.linux]
 before-all = [
     "yum install doxygen libxml2-devel libcurl-devel -y",
+    "wget https://github.com/ninja-build/ninja/releases/download/v1.11.1/ninja-linux.zip",
+    "unzip ninja-linux.zip -d /usr/local/bin",
     "buildscripts/build_dependencies.sh -p /usr/local"
 ]
 [tool.cibuildwheel.linux.environment]
