### 2023-09-21 e16b9c8
Adjustments to netcdf-fortran build script to to get working in cibuildwheel
