+++ pyproject.toml
@@ -104,8 +104,6 @@ before-all = [
 [tool.cibuildwheel.linux]
 before-all = [
     "yum install doxygen libxml2-devel libcurl-devel -y",
-    "wget https://github.com/ninja-build/ninja/releases/download/v1.11.1/ninja-linux.zip",
-    "unzip ninja-linux.zip -d /usr/local/bin",
     "buildscripts/build_dependencies.sh -p /usr/local"
 ]
 [tool.cibuildwheel.linux.environment]
