### 2023-09-22 311e667
fixed typos in build scripts
