### 2023-09-22 d4c8dcb
Major update to build scripts. Now use CMake for all dependencies and attempted to simplify the FindNETCDF_Fortran package finder
