+++ buildscripts/set_environment.sh
@@ -11,18 +11,19 @@ LDFLAGS="-Wl,-rpath,${ROOT_DIR}/lib  -Wl,-no_compact_unwind -L${PREFIX}/lib -L${
 CPATH="/usr/local/include:${PREFIX}/include:${HOMEBREW_PREFIX}/include:${ROOT_DIR}/include"
 CPPFLAGS="-isystem ${PREFIX}/include -isystem /usr/local/include -Xclang -fopenmp"
 LIBS="-lomp"
+CFLAGS="-mmacosx-version-min=${MACOSX_DEPLOYMENT_TARGET} -Wno-deprecated-non-prototype -arch ${ARCH}"
 FCFLAGS="-mmacosx-version-min=${MACOSX_DEPLOYMENT_TARGET}"
 FFFLAGS="${FCFLAGS}"
 CFLAGS="${FCFLAGS} -Wno-deprecated-non-prototype"
 CXXFLAGS="${CFLAGS}"
-HDF5_ROOT="${PREFIX}"
+HDF5_ROOT="${PREFIX}/HDF_Group/HDF5/1.14.2"
 HDF5_LIBDIR="${HDF5_ROOT}/lib"
 HDF5_INCLUDE_DIR="${HDF5_ROOT}/include"
 HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
-NETCDF_FORTRAN_HOME="${PREFIX}"
-NETCDF_INCLUDE="${PREFIX}"
 NCDIR="${PREFIX}"
 NFDIR="${PREFIX}"
+NETCDF_FORTRAN_HOME="${NFDIR}"
+NETCDF_FORTRAN_INCLUDE="${NFDIR}/include"
 FC="$(command -v gfortran-12)"
 F77="${FC}"
 CC="/usr/bin/clang"
