### 2023-09-22 2aab087
Fixed a few lingering issues with the Mac build
