+++ buildscripts/build_hdf5.sh
@@ -84,22 +84,28 @@ else
     SZIP_LIBRARY="${SZIP_ROOT}/lib/libsz.so"
 fi
 
-cmake -B build -C ./config/cmake/cacheinit.cmake -G Ninja \
-    -DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT} \
-    -DHDF5_ALLOW_EXTERNAL_SUPPORT:STRING="TGZ" \
+ARGLIST="-DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT} \
+    -DHDF5_ALLOW_EXTERNAL_SUPPORT:STRING=TGZ \
     -DBUILD_ZLIB_WITH_FETCHCONTENT:BOOL=ON \
     -DZLIB_USE_LOCALCONTENT:BOOL=OFF \
-    -DZLIB_TGZ_ORIGNAME:STRING="${ZLIB_TGZ_NAME}" \
-    -DZLIB_TGZ_ORIGPATH:STRING="${ZLIB_TGZ_ORIGPATH}" \
+    -DZLIB_TGZ_ORIGNAME:STRING=${ZLIB_TGZ_NAME} \
+    -DZLIB_TGZ_ORIGPATH:STRING=${ZLIB_TGZ_ORIGPATH} \
     -DHDF5_ENABLE_SZIP_SUPPORT:BOOL=ON \
     -DBUILD_SZIP_WITH_FETCHCONTENT:BOOL=ON \
     -DLIBAEC_USE_LOCALCONTENT:BOOL=OFF \
-    -DLIBAEC_TGZ_ORIGNAME:STRING="${LIBAEC_TGZ_NAME}" \
-    -DLIBAEC_TGZ_ORIGPATH:STRING="${LIBAEC_TGZ_ORIGPATH}" \
+    -DLIBAEC_TGZ_ORIGNAME:STRING=${LIBAEC_TGZ_NAME} \
+    -DLIBAEC_TGZ_ORIGPATH:STRING=${LIBAEC_TGZ_ORIGPATH} \
     -DHDF5_ENABLE_PLUGIN_SUPPORT:BOOL=OFF \
     -DHDF5_BUILD_CPP_LIB:BOOL=OFF \
     -DHDF5_BUILD_FORTRAN:BOOL=OFF \
-    -DHDF5_BUILD_JAVA:BOOL=OFF  .
+    -DHDF5_BUILD_JAVA:BOOL=OFF"
+
+
+if [ $OS = "MacOSX" ]; then
+    ARGLIST="${ARGLIST} -DCMAKE_BUILD_WITH_INSTALL_RPATH:BOOL=OFF"
+fi
+
+cmake -B build -C ./config/cmake/cacheinit.cmake -G Ninja ${ARGLIST} .
 
 cmake --build build -j${NPROC}
 if [ -w ${PREFIX} ]; then
