+++ pyproject.toml
@@ -81,8 +81,8 @@ HDF5_ROOT="${PREFIX}/HDF_Group/HDF5/1.14.2"
 HDF5_LIBDIR="${HDF5_ROOT}/lib"
 HDF5_INCLUDE_DIR="${HDF5_ROOT}/include"
 HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
-NCDIR="${PREFIX}/netCDF-C"
-NFDIR="${PREFIX}/netCDF-Fortran"
+NCDIR="${PREFIX}"
+NFDIR="${PREFIX}"
 NETCDF_FORTRAN_HOME="${NFDIR}"
 NETCDF_FORTRAN_INCLUDE="${NFDIR}/include"
 FC="$(command -v gfortran-12)"
