### 2023-09-22 cb2ad73
More changes to library linking.  wheel fails to find all libraries still
