+++ pyproject.toml
@@ -77,7 +77,7 @@ FCFLAGS="-mmacos-version-min=${MACOSX_DEPLOYMENT_TARGET} -arch ${ARCH}"
 FFFLAGS="${FCFLAGS}"
 CFLAGS="${FCFLAGS} -Wno-deprecated-non-prototype -arch ${ARCH}"
 CXXFLAGS="${CFLAGS}"
-HDF5_ROOT="${PREFIX}/HDF_Group/HDF5/1.14.2"
+HDF5_ROOT="${PREFIX}"
 HDF5_LIBDIR="${HDF5_ROOT}/lib"
 HDF5_INCLUDE_DIR="${HDF5_ROOT}/include"
 HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
@@ -107,11 +107,21 @@ before-all = [
     "buildscripts/build_dependencies.sh -p /usr/local"
 ]
 [tool.cibuildwheel.linux.environment]
-NETCDF_FORTRAN_HOME="/usr/local"
-NETCDF_INCLUDE="/usr/local/include"
-LD_LIBRARY_PATH="/usr/local/lib:/usr/local/lib64:/project/lib"
-CPATH="/usr/local/include:/project/include"
-netCDF-Fortran_DIR="/usr/local/lib64/cmake/netCDF"
+PREFIX="/usr/local"
+NCDIR="${PREFIX}"
+NFDIR="${PREFIX}"
+NETCDF_FORTRAN_HOME="${NFDIR}"
+NETCDF_FORTRAN_INCLUDE="${NFDIR}/include"
+HDF5_ROOT="${PREFIX}"
+HDF5_LIBDIR="${HDF5_ROOT}/lib"
+HDF5_INCLUDE_DIR="${HDF5_ROOT}/include"
+HDF5_PLUGIN_PATH="${HDF5_LIBDIR}/plugin"
+LD_LIBRARY_PATH="${PREFIX}/lib:/project/lib:${HDF5_LIBDIR}"
+CPATH="${PREFIX}/include:/project/include:${HDF5_INCLUDE_DIR}"
+netCDF-Fortran_DIR="${PREFIX}/lib/cmake/netCDF"
+HDF5_DIR="${HDF5_ROOT}/cmake"
+PATH="${HDF5_ROOT}/bin:${PATH}"
+CMAKE_INSTALL_LIBDIR="lib"
 
 [[tool.scikit-build.generate]]
 path = "version.txt"
