+++ buildscripts/build_hdf5.sh
@@ -85,6 +85,7 @@ else
 fi
 
 ARGLIST="-DCMAKE_INSTALL_PREFIX:PATH=${HDF5_ROOT} \
+    -DCMAKE_BUILD_TYPE:STRING="Release" \
     -DHDF5_ALLOW_EXTERNAL_SUPPORT:STRING=TGZ \
     -DBUILD_ZLIB_WITH_FETCHCONTENT:BOOL=ON \
     -DZLIB_USE_LOCALCONTENT:BOOL=OFF \
