+++ buildscripts/build_netcdf-c.sh
@@ -58,6 +58,7 @@ ZLIB_ROOT=${PREFIX}
 cmake -B build -S . -G Ninja \
     -DCMAKE_BUILD_TYPE:STRING="Release" \
     -DHDF5_DIR:PATH=${HDF5_ROOT}/cmake \
+    -DHDF5_ROOT:PATH=${HDF5_ROOT} \
     -DCMAKE_INSTALL_PREFIX:STRING="${NCDIR}" \
     -DENABLE_DAP:BOOL=OFF \
     -DENABLE_BYTERANGE:BOOL=OFF \
