### 2023-09-24 cb2faa4
Made sph_harmonnics_gravity a class for users to call it
