+++ swiftest/sph_harmonics_gravity.py
@@ -28,108 +28,109 @@ from typing import (
     Any
 )
 
-def clm_from_ellipsoid(mass, density, a, b = None, c = None, lmax = 6, ref_radius = False):
-    """
-    Creates and returns the gravity coefficients for an ellipsoid with principal axes a, b, c upto a certain maximum degree lmax. 
-    Uses pyshtools. No units necessary for a, b, & c. However, they need to be in the same units (DU).
+class Sph_Harmonics(object):
+    def clm_from_ellipsoid(mass, density, a, b = None, c = None, lmax = 6, ref_radius = False):
+        """
+        Creates and returns the gravity coefficients for an ellipsoid with principal axes a, b, c upto a certain maximum degree lmax. 
+        Uses pyshtools. No units necessary for a, b, & c. However, they need to be in the same units (DU).
 
-    Parameters
-    ----------
-    mass : float
-        mass of the central body
-    density : float
-        density of the central body
-    a : float 
-        length of the pricipal axis aligned with the x axis
-    b : float, optional, default = a
-        length of the pricipal axis aligned with the y axis
-    c : float, optional, default = b
-        length of the pricipal axis aligned with the z axis
-    lmax : int, optional, default = 6
-        The maximum spherical harmonic degree resolvable by the grid.
-    ref_radius : boolean, optional, default = False
-        Boolean value to return the reference radius calculated by SHTOOLS
+        Parameters
+        ----------
+        mass : float
+            mass of the central body
+        density : float
+            density of the central body
+        a : float 
+            length of the pricipal axis aligned with the x axis
+        b : float, optional, default = a
+            length of the pricipal axis aligned with the y axis
+        c : float, optional, default = b
+            length of the pricipal axis aligned with the z axis
+        lmax : int, optional, default = 6
+            The maximum spherical harmonic degree resolvable by the grid.
+        ref_radius : boolean, optional, default = False
+            Boolean value to return the reference radius calculated by SHTOOLS
 
-    Returns
-    -------
-    clm : ndarry, shape (2, lmax+1, lmax+1)
-        numpy ndarray of the gravitational potential spherical harmonic coefficients. 
-        This is a three-dimensional array formatted as coeffs[i, degree, order], 
-        where i=0 corresponds to positive orders and i=1 to negative orders.
+        Returns
+        -------
+        clm : ndarry, shape (2, lmax+1, lmax+1)
+            numpy ndarray of the gravitational potential spherical harmonic coefficients. 
+            This is a three-dimensional array formatted as coeffs[i, degree, order], 
+            where i=0 corresponds to positive orders and i=1 to negative orders.
 
-    """
-    Gmass = swiftest.constants.GC * mass # SHTOOLS uses an SI G value, and divides it before using the mass; NO NEED TO CHANGE UNITS
+        """
+        Gmass = swiftest.constants.GC * mass # SHTOOLS uses an SI G value, and divides it before using the mass; NO NEED TO CHANGE UNITS
 
-    # cap lmax to ensure fast performance without giving up accuracy
-    lmax_limit = 6              # lmax_limit = 6 derived from Jean's Law; characteristic wavelength = the radius of the CB
-    if(lmax > lmax_limit):                           
-        lmax = lmax_limit
-        print(f'Setting maximum spherical harmonic degree to {lmax_limit}')
+        # cap lmax to ensure fast performance without giving up accuracy
+        lmax_limit = 6              # lmax_limit = 6 derived from Jean's Law; characteristic wavelength = the radius of the CB
[…]
