### 2023-09-25 9d6fc7a
Additions made while troubleshooting the build_dependencies script on Linux
