### 2023-09-25 3a5cddb
Removed debug flag from the netcdf-c build script
