### 2023-09-25 1b1186c
Added the setting of environment variables to the set_compiler script
