### 2023-09-25 e13383f
Removed libaec from build dependencies as it caused issues in mac builds
