### 2023-09-25 c3e3f39
Set default dependency build location and prefix to ~/Downloads and /usr/local for simplicity
