### 2023-09-25 830235a
added back the cp312 build to the wheels
