### 2023-09-25 f8299c2
Added sign as a dimension in netcdf; debugging to see why c_lm isn t read in fortran
