+++ swiftest/init_cond.py
@@ -505,7 +505,7 @@ def vec2xr(param: Dict, **kwargs: Any):
     # create a C_lm Dataset and combine
 
     if "c_lm" in kwargs:
-        clm_xr = xr.Dataset(data_vars = {k:(sph_dims, v) for k,v in kwargs.item() if k in sph_vars}, 
+        clm_xr = xr.Dataset(data_vars = {k:(sph_dims, v) for k,v in kwargs.items() if k in sph_vars}, 
                             coords = {
                                     'sign':(['sign'], [1, -1]),
                                     'l': (['l'], range(0, kwargs['c_lm'].shape[1])),
