### 2023-09-25 f96817c
Skip the cp312 Linux version because h5py can't be built
