### 2023-09-25 cdbe4a6
edited netcdf calls to c_lm dimensions
