### 2023-09-25 #5409 @warrickball
Hello again, dear reviewers! I think this is close to publication so would appreciate a last push on the few remaining checklist items.

@hannorein, your only outstanding item appears to be the presence of an open source license. @carlislewishard and @profminton, I notice that GitHub doesn't seem to automatically pick up on it but you do have a license file that says GPLv3 and the text of GPLv3 is also present (in `COPYING.txt`). I think this should be satisfactory, though the `LICENSE.txt` file might be redundant.

@mikegrudic, your only unchecked item is the present of automatic (or at least easy-to-run) tests. Does the latest version of Swiftest satisfy this requirement? There are tests under `tests/`, which aren't automatic but that isn't a strict requirement for JOSS.

Finally, @aatrani you also seem to be waiting to confirm the code's functionality. Could you give this another try? Swiftest is now much easier to install and I was myself able to test some functionality.
