### 2023-09-25 #5409 @profminton
FYI, the tests are actually run automatically by the continuous integration system, `cibuildwheels`. The tests are quite basic at the moment, and I hope to expand them to include more test conditions (like the precession of Mercury with GR and also various conserved quantities during collisions, etc.). But the current test suite is a good platform for us to build on as the Python wheel building will fail if any test condition is not met.

I will look into the license issue when I have a chance.
