### 2023-09-25 1735a5a
added a cacheinit configuration file for the netcdf-fortran build on Windows (rough draft)
