### 2023-09-25 #5409 @hannorein
Very minor: I think the license file [needs](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository) to be named LICENSE or LICENSE.md not LICENSE.txt.

I agree that more tests would be nice. It's really hard to have too many tests, even if they are very short/simple.
