### 2023-09-26 fa952e1
Allow all but the linux build to make the python 3.12 version
