+++ pyproject.toml
@@ -59,7 +59,7 @@ cmake.verbose = true
 [tool.cibuildwheel]
 test-command = "pytest {package}/tests"
 test-requires = ['pytest']
-skip = "cp312-cp312-manylinux* pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
+skip = "pp* *i686 *-manylinux_i686 *_ppc64le *_s390x *-musllinux* *-win32"
 build-verbosity = 1
 
 [tool.cibuildwheel.macos.environment]
@@ -102,6 +102,7 @@ before-all = [
 ]
 
 [tool.cibuildwheel.linux]
+skip = "cp312-*"
 before-all = [
     "yum install doxygen libxml2-devel libcurl-devel -y",
     "buildscripts/build_dependencies.sh -p /usr/local"
