### 2023-09-26 40e683b
Added pyshtools to the list of requirements
