### 2023-09-26 1cb9c92
Made a number of changes to get the new SHTOOLS into the build and get the libraries found
