### 2023-09-26 719b06d
Made adjustments to build scripts
