### 2023-09-27 54a9f95
Added new windows cmake config files
