### 2023-09-27 a947007
Updated windows build cmake settings files
