+++ CMakeLists.txt
@@ -67,12 +67,20 @@ IF(PROJECT_SOURCE_DIR STREQUAL PROJECT_BINARY_DIR)
 ENDIF()
 
 FIND_PACKAGE(NETCDF_Fortran REQUIRED)
-FIND_PACKAGE(HDF5 COMPONENTS C HL REQUIRED)
+IF (MSVC)
+    FIND_PACKAGE(HDF5 NAMES hdf5 COMPONENTS C HL REQUIRED CONFIG)
+    MESSAGE(STATUS "HDF5_FOUND: ${HDF5_FOUND}")
+    MESSAGE(STATUS "HDF5_VERSION: ${HDF5_VERSION}")
+    MESSAGE(STATUS "HDF5_LIBRARIES: ${HDF5_LIBRARIES}")
+    MESSAGE(STATUS "HDF5_INCLUDE_DIRS: ${HDF5_INCLUDE_DIRS}")
+ELSE ()
+    FIND_PACKAGE(HDF5 COMPONENTS C HL REQUIRED)
+ENDIF ()
 
 # Ensure scikit-build modules
 FIND_PACKAGE(Python COMPONENTS Interpreter Development.Module REQUIRED)
 
-# Define some directories that are important to the build
+
 SET(SRC "${CMAKE_SOURCE_DIR}/src")
 SET(PY "${CMAKE_SOURCE_DIR}/swiftest")
 
