### 2023-09-28 cbe7f9f
avoided allocation error for a pre-allocated c_lm array
