### 2023-09-28 31364df
rearranged allocation order in write_frame_cb; Changed order of indices in gravity calculation
