### 2023-09-28 09515b3
cchanged indexing of "ah" and "aobl" to prevent seg fault
