### 2023-09-28 1da0860
added lshgrav to allocation criteria to aobl
