### 2023-09-28 f859be1
Converted while loop into a do loop
