### 2023-09-28 4a6b2f8
Converted another while loop to a do
