### 2023-09-28 6b3f3a7
fixed a typo. Changed dimension of GMpl
