### 2023-09-28 bfbb41e
fixed typo with g_Sph dimensions
