### 2023-09-29 4efa637
fixed typo that was leading pl%aobl to not being allocated
