+++ tests/test_swiftest.py
@@ -260,122 +260,3 @@ class TestSwiftest(unittest.TestCase):
         
 if __name__ == '__main__':
     unittest.main()
-    if os.path.exists("simdir"):
-        shutil.rmtree("simdir")
-
-        # Add the modern planets and the Sun using the JPL Horizons Database.
-        # Add the modern planets and the Sun using the JPL Horizons Database.
-        sim.add_solar_system_body(major_bodies)
-    
-        # Display the run configuration parameters.
-        param = sim.get_parameter(verbose=False)
-        sim.save()
-        
-        for f in file_list:
-            self.assertTrue(os.path.exists(f))
-            
-        print("\ntest_read_ic: Test whether we can read back initial conditions files created by test_gen_ic")
-        sim = swiftest.Simulation(read_param=True)
-        # Check if all names in Dataset read in from file match the expected list of names
-        self.assertTrue((major_bodies == sim.init_cond['name']).all(), msg="Name mismatch in Dataset")
-        
-        # Check to see if all parameter values read in from file match the expected parameters saved when generating the file
-        self.assertTrue(all([v == param[k] for k,v in sim.param.items() if k in param]))
-      
-    def test02_integrators(self):
-        """
-        Tests that Swiftest is able to integrate a collection of massive bodies and test particles with all available integrators
-        """ 
-        print("\ntest_integrators: Tests that Swiftest is able to integrate a collection of massive bodies and test particles with all available integrators")
-        sim = swiftest.Simulation()
-        sim.clean()
-
-        # Add the modern planets and the Sun using the JPL Horizons Database.
-        # Add the modern planets and the Sun using the JPL Horizons Database.
-        sim.add_solar_system_body(major_bodies)
-    
-        # Display the run configuration parameters.
-        param = sim.get_parameter(verbose=False)
-        sim.save()
-        
-        # Add 10 user-defined test particles.
-        ntp = 10
-
-        name_tp      = [f"TestParticle_{i:02}" for i in range(1,ntp+1)]
-        a_tp          = rng.uniform(0.3, 1.5, ntp)
-        e_tp          = rng.uniform(0.0, 0.2, ntp)
-        inc_tp        = rng.uniform(0.0, 10, ntp)
-        capom_tp     = rng.uniform(0.0, 360.0, ntp)
-        omega_tp     = rng.uniform(0.0, 360.0, ntp)
-        capm_tp      = rng.uniform(0.0, 360.0, ntp)
-
-        integrators= ["helio","whm","rmvs","symba"]
-        sim.add_body(name=name_tp, a=a_tp, e=e_tp, inc=inc_tp, capom=capom_tp, omega=omega_tp, capm=capm_tp)
-        sim.set_parameter(tstart=0.0, tstop=0.10, dt=0.01, istep_out=1, dump_cadence=0)
-        for i in integrators:
-            sim.run(integrator=i)
-            
-            
-    def test03_conservation(self):
-        """
-        Tests that Swiftest conserves mass, energy, and momentum to within acceptable tolerances.
-        """
-        print("\ntest_conservation: Tests that Swiftest conserves mass, energy, and momentum to within acceptable tolerances.")
-        
-        # Error limits
-        L_slope_limit = 1e-10
-        E_slope_limit = 1e-8
-        GM_limit = 1e-14
- 
-        sim = swiftest.Simulation()
-        sim.clean()
-        
-        sim.add_solar_system_body(major_bodies)
-        
-        dt = 0.01
-        nout = 1000
[…]
