### 2023-10-02 da736f1
Fixed bad merge of test script
