+++ version.txt
@@ -1,1 +1,1 @@
-2023.9.4
+2023.10.0
