### 2023-10-02 5a11574
Streamlined dependency build on MacOS
