### 2023-10-02 f539e26
Fixed issue that was causing initial conditions from being messed up when calling .run() multiple times in the same session, which screwed up the tests. Tests work again
