+++ pyproject.toml
@@ -102,7 +102,6 @@ before-all = [
 ]
 
 [tool.cibuildwheel.linux]
-skip = "cp312-*"
 before-all = [
     "yum install doxygen libxml2-devel libcurl-devel -y",
     "buildscripts/build_dependencies.sh -p /usr/local"
