### 2023-10-02 f55fa15
Updated versioning of dependency python packages and got rid of some cruft in the hdf5 build script
