+++ pyproject.toml
@@ -30,8 +30,9 @@ dependencies = [
     'dask>=2022.1',
     'distributed>=2022.1',
     'bottleneck>=1.3.5',
-    'h5netcdf>=1.0.2',
-    'netcdf4>=1.6.2',
+    'h5netcdf>=1.2.0',
+    'h5py>=3.9.0',
+    'netcdf4>=1.6.4',
     'matplotlib>=3.7.1',
     'astropy>=5.1',
     'astroquery>=0.4.6',
