### 2023-10-02 96429ee
Relaxing some packaging version requirements
