### 2023-10-03 0e86eaf
Put back the missing download of hdf5
