+++ pyproject.toml
@@ -103,6 +103,7 @@ before-all = [
 ]
 
 [tool.cibuildwheel.linux]
+skip = "cp312-*"
 before-all = [
     "yum install doxygen libxml2-devel libcurl-devel -y",
     "buildscripts/build_dependencies.sh -p /usr/local"
