### 2023-10-03 47e3c88
Fixed bug that was preventing initial conditions file from being saved when new bodys are added in multiple add_solar_system_body calls
