### 2023-10-04 592bba8
Updated dependency version minimums
