+++ pyproject.toml
@@ -26,17 +26,17 @@ keywords=['astronomy','astrophysics', 'planetary', 'nbody integrator', 'symplect
 dependencies = [
     'numpy>=1.24.3',
     'scipy>=1.10.1',
-    'xarray>=2022.11.0',
-    'dask>=2022.1',
-    'distributed>=2022.1',
-    'bottleneck>=1.3.5',
-    'h5netcdf',
-    'h5py',
-    'netcdf4',
-    'matplotlib>=3.7.1',
-    'astropy>=5.1',
+    'xarray>=2023.1',
+    'dask>=2023.5',
+    'distributed>=2023.5',
+    'bottleneck>=1.3',
+    'h5netcdf>=1.1',
+    'h5py>=3.9',
+    'netcdf4>=1.6.4',
+    'matplotlib>=3.7',
+    'astropy>=5.2',
     'astroquery>=0.4.6',
-    'tqdm>=4.65.0',
+    'tqdm>=4.66',
     'cython>=3.0.0',
     'pyshtools>=4.10'
 ]
