+++ examples/Chambers2013/README.txt
@@ -15,9 +15,10 @@ Date             : December 6, 2022
 
 Included in the Chambers2013 example directory are the following files:
 
-	- README.txt      : This file
-	- init_cond.py    : A Python Script that generates a set of initial conditions.
-	- scattermovie.py : A Python Script that processes data.nc and generates Chambers2013-aescatter.mp4 or Chambers2013-aiscatter.mp4
+	- README.txt        : This file
+	- init_cond.py      : A Python script that generates a set of initial conditions.
+	- run_simulation.py : A Python script that will run the simulation.
+	- scattermovie.py   : A Python script that processes data.nc and generates Chambers2013-aescatter.mp4 or Chambers2013-aiscatter.mp4
 
 This example is intended to be run with Swiftest SyMBA. For details on how to generate, run, and analyze this example, 
 see the Swiftest User Manual.
