### 2023-10-04 67814bc
Cleaned up the Chambers2013 example and added a new run_simulation script that can generate a short test case.
