+++ examples/Chambers2013/init_cond.py
@@ -1,171 +1,0 @@
-#!/usr/bin/env python3
-
-"""
- Copyright 2023 - David Minton, Carlisle Wishard, Jennifer Pouplin, Jake Elliott, & Dana Singh
- This file is part of Swiftest.
- Swiftest is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License 
- as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
- Swiftest is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty 
- of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- You should have received a copy of the GNU General Public License along with Swiftest. 
- If not, see: https://www.gnu.org/licenses. 
-"""
-
-"""
-Generates a set of Swiftest input files from initial conditions with the SyMBA integrator. All simulation 
-outputs are stored in the /simdata subdirectory.
-
-Input
-------
-None.
-
-Output
-------
-inital_conditions.png : A .png file depicting the simulation initial configuration.
-init_cond.nc          : A NetCDF file containing the initial conditions for the simulation.
-param.in              : An ASCII file containing the parameters for the simulation.
-""" 
-
-import swiftest
-import numpy as np
-from numpy.random import default_rng
-import matplotlib.pyplot as plt
-
-# Initialize simulation object
-sim = swiftest.Simulation(compute_conservation_values=True,  rotation=True, init_cond_format="EL",collision_model="fraggle",encounter_save="none")
-
-# Add bodies described in Chambers (2013) Sec. 2.1, with the uniform spatial distribution and two bodies sizes (big and small)
-Nb = 14
-Ns = 140
-Mb = 2.8e-7 * 14 / Nb
-Ms = 2.8e-8 * 140 / Ns
-dens = 3000.0 * sim.KG2MU / sim.M2DU**3
-
-mtiny = 1e-2 * Ms
-minimum_fragment_mass = 1e-5 * Ms
-nfrag_reduction = 30.0
-rng = default_rng(seed=3031179)
-
-runname = "Chambers (2013)"
-def a_profile(n_bodies):
-    """
-    Generates the profile described in Sec. 2.1 of Chambers:  
-    
-    *In all cases, the surface density R = 8 g/cm2 at 1 AU, varying as a**(-3/2), where a is the orbital semi-major axis. 
-    The region with a < 0.7 AU deviates from this law, declining linearly with decreasing distance until R = 0 at 0.3 AU. 
-    The outer edge of the disk is 2 AU in all cases.
-    """
-    def sample(r_inner, r_break, r_outer, slope1, slope2):
-       """ 
-       Define the functions to pull random semi-major axes from a distribution using a rejection sampling method
-       This defines a 2-slope model with a break at r_break
-       Based on (https://stackoverflow.com/questions/66874819/random-numbers-with-user-defined-continuous-probability-distribution)
-       """
-       while True:
-          x = rng.uniform(r_inner, r_outer, size=1)
-          
[…]
