### 2023-10-04 ba9f088
Decrased the number of fragments made in the example movie script to make it run faster, and rearranged script to make the variable that controls fragment number more obvious.
