### 2023-10-05 #5409 @warrickball
I'll open a PR to tweak a few bits of typography but I want to flag two things that need attending.

First, there should be some text explaining Swiftest's relationship to other N-body integrators. Do any exist? If they do, how is Swiftest complementary? I.e, why would one use Swiftest instead of another package? I don't know how REBOUND or [Sverre Aarseth's codes] compare, for example. It's not my field, so there might be others (and perhaps neither of these is worth mentioning).

Second, it's not clear what the dashed line in Fig. 1 means. The text says it's
> an ideal 1:1 speedup relative to Swiftest in serial
but none of them go through the top right corners (cores=speedup=24) of the plots. Perhaps the speedup should be expected to be some other ratio?
