### 2023-10-05 #5409 @profminton
I'll address the comparison with other n-body codes.

You aren't the first person to point out the confusion with that plot, but I forgot to go and update it to make it more clean. I'll try to get those taken care of in the next day or two.
