### 2023-10-07 #5409 @profminton
I've updated the paper to plot the speedup relative to single processor Swiftest instead of Swift, which makes it more intuitive. I also fleshed out some of the text and discuss more of the details regarding the integrators we have implemented and how they compare to other symplectic integrators.

The version that I've tagged for release is v2023.10.2
