### 2023-10-09 #5409 @warrickball
Thanks @profminton. I've started a PR in which I'll fix some remaining editorial items. Because the paper is part of the release that gets archived, hold off on creating it until we've settled on the paper. I'm looking for a better reference for *Swifter* will also ask around about why some of the citations give initials or first names as well as surnames in the text.

The overall content looks fine, so there should be little to do on your side once I've figured this out.
