### 2023-10-09 #5409 @warrickball
I concede defeat on using [this URL](https://www.boulder.swri.edu/swifter/) as a reference for *Swifter*, since there's no other authoritative citation and some published works refer to *Swifter* with no citation and this URL in a footnote.

For the rest, I opened [this PR](https://github.com/profminton/swiftest/pull/1) to make a few editorial changes. Let me know once that's merged I'll do the final check on DOIs, at which point we can mint a new archive and I'll recommend the paper for acceptance.
