### 2023-10-10 #5409 @warrickball
Great! I discussed this with other editors, and for the moment I think the answer is no unless an editor-in-chief says otherwise. The code hasn't changed, so we don't mind if the paper in the archive differs very slightly (only on editorial matters, too) from the published version.

I'll run some last checks right now but if the DOIs resolve correctly we're basically done.
