### 2023-10-10 #5409 @warrickball
@profminton, the last thing is to amend the Zenodo archive metadata so that the author list and license (I presume you want GPLv3) match the JOSS paper. You should be able to do this without creating a new archive.
