### 2023-10-10 #5409 @profminton
@warrickball I've attempted to edit the metadata, but I'm a bit confused as to actually go about doing it. Under the Metadata tab it has this text:
> Zenodo automatically extracts metadata about your repository from GitHub APIs. For example, the authors are determined from the repository's contributor statistics. The automatic extraction is solely a best guess. Add a .zenodo.json file the root of your repository to explicit define the metadata. The format of file is the same as for our REST API (use e.g. below JSON to get started).

I added a `.zenodo.json` file to the repository, but I'm not sure how to make Zenodo see it unless I make a new release. Please advise.
