### 2023-10-10 #5409 @warrickball
I just logged in to Zenodo using the "Log in with GitHub" option and navigated to an archive I created. I see an orange "Edit" button at the top right and about one-third the way down the page is the list of authors, where I can add, remove or amend authors. A bit further down are the licence details. After amending the data, the "Publish" button is greyed out because you need to "Save" first. Pressing the "Publish" button then raises a warning like:

> Once the record is published you will no longer be able to change the files in this upload.
This is because a Digital Object Identifier (DOI) will be registered immediately after publishing. You will still be able to update the record's metadata later.

which is safe to ignore because we aren't changing any files. (I think this warning appears every time you update the metadata but you can do so as many times as you like.)
