### 2023-10-10 #5409 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
