### 2023-10-10 #5409 @dfm
Many thanks to @hannorein, @mikegrudic, and @aatrani for reviewing and to @warrickball for editing! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you!!

@carlislewishard, @profminton — Your paper is now accepted and published in JOSS! :zap::rocket::boom:
